Makes html_viewer only swap to remote when necessary
Previously on navigation the code preemptively swapped from local to
remote. This is wrong and problematic if the frame was going to be
local after the navigate.

The fix is to change the frame server code to consult the content
handler id when doing a navigation. If the id is the same as the
existing id, then the same renderer is going to be used. If the same
renderer is used OnConnect() is supplied a different argument to indicate the
new connection corresponds to an existing frame. Additionally a
ViewTreeClient is not obtained for the frame (this is because the
html_viewer already has a View for the frame).

BUG=479172,490221,523408
TEST=covered by tests
R=ben@chromium.org

Review URL: https://codereview.chromium.org/1327673004

Cr-Commit-Position: refs/heads/master@{#346794}