Mac Overlays: Allow sandwich strategy to emit multiple overlays
Cut the region of the overlay candidate into rects and emit a quad and overlay
for each rect. This helps with situations where there are logos or controls in the
corners of a video (which then cause a huge top overlay).

BUG=522627
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1315053004

Cr-Commit-Position: refs/heads/master@{#346816}