CREDENTIAL: Update PasswordCredential's constructor.
Rather than a list of parameters, PasswordCredential is defined[1] to
take a PasswordCredentialData object. We defined that object long ago
(in [2]), but apparently never actually used it. This patch fixes that
oversight.

As a drive-by, this patch also stops exposing the PasswordCredential
and FederatedCredential to Workers, as per spec.

[1]: https://w3c.github.io/webappsec/specs/credentialmanagement/#interfaces-credential-types-passwordcredential
[2]: https://codereview.chromium.org/1162883002

BUG=494880

Review URL: https://codereview.chromium.org/1314673007

git-svn-id: svn://svn.chromium.org/blink/trunk@201723 bbb929c8-8fbe-4397-9dbb-9b2b20218538