Fixes possible race in new connection setup
The frame tree uses change_ids to track the tree status. When a
connection is made to a frame it is entirely possible for the new
connection to have a change_id > change_id seen by tree. To deal with
this HTMLFrameTreeManager gets an observer. DocumentResourceWaiter, as
necessary, adds itself as an observer and waits for the change ids to
match. Once they match, loading continues.

BUG=479172,490221
TEST=none
R=fsamuel@chromium.org

Review URL: https://codereview.chromium.org/1325903003

Cr-Commit-Position: refs/heads/master@{#347186}