webapps: Add a registry to keep track of web apps with data
Web apps which are storing data using SharedPreferences will need to be cleaned
up at some point. To do this, we need a way of knowing the web apps which have
stored data.

The way used in this CL is preferrable to iterating over the Files as less IO
happens during cleanup plus we don't need to have ugly hardcoded paths. Hardcoding
would be necessary as Android does not have a way of finding all the
SharedPreferences an application has created.

BUG=508627

Review URL: https://codereview.chromium.org/1312403004

Cr-Commit-Position: refs/heads/master@{#347189}