Don't transfer viewport to an invalid/dirty <use> instance
If an update is pending for a <use> shadow tree, the width/height
needn't be updated, since the update of the shadow tree will do it.
Make sure to clear the reference to the instance root when
invalidating the shadow tree.
Also split clearResourceReferences into a few different parts.

BUG=516051

Review URL: https://codereview.chromium.org/1323443003

git-svn-id: svn://svn.chromium.org/blink/trunk@201738 bbb929c8-8fbe-4397-9dbb-9b2b20218538