Fix flakiness in SitePerProcessBrowserTest.ProxyCreationSkipsSubtree.
One of the frame navigations in this test is checking the frame tree
after waiting on a TestFrameNavigationObserver and assuming that the
frame's child frames don't exist at that point, but that may not
be the case.  This CL changes that navigation to a page without
subframes, which don't seem to be needed/used in the test.
https://crbug.com/514420#c5 has more details.

BUG=514420

Review URL: https://codereview.chromium.org/1322333002

Cr-Commit-Position: refs/heads/master@{#347253}