Fixing binary size increase caused by tracing in blink GC
This CL is the result of cleanly cherry-picking the following CLs:
1. Fix race at GCInfo::index (www.crrev.com/1312963002).
2. Warn about extractTypeNameFromFunctionName usage
     (www.crrev.com/1307743008).
3. Fix binary size increase because of class name loading
     (www.crrev.com/1317563005).
4. Fix GC_PROFILING builds following r201671
     (www.crrev.com/1308053007).

The static strings causes binary size increase in the release branch
(crbug.com/527054). This CL cherry-picks the changes made in the
ToT for fixing this regression, and the minimal depending changes,
required for a clean cherry-pick.

BUG=527054
R=haraken@chromium.org, primiano@chromium.org, sigbjornf@opera.com

Review URL: https://codereview.chromium.org/1307233008 .

git-svn-id: svn://svn.chromium.org/blink/branches/chromium/2490@201793 bbb929c8-8fbe-4397-9dbb-9b2b20218538