Verify that cert status is valid before forgetting user exceptions
Adds a check in SSLPolicy when user exceptions for bad certificates are
forgotten to make sure that the certificate information is valid. Simply
adds a check that the cert_id is not 0. Also adds unit tests to validate
this chain of events.

In SSLPolicy, if a user has made an exception for an invalid certificate
for a host by clicking on an interstitial, that exception is forgotten
if a good certificate is ever seen for the host. Unfortunately, when we
checked if the certificate has an error or not, we forgot to check *if
the certificate info is valid*, and thus we were getting certificate
information that that was not valid, and by default lists no certificate
errors. Thus, Chrome thought it found a certificate without an error,
when what it really found was invalid certificate information.

BUG=516808
TBR=creis@chromium.org,estark@chromium.org

Review URL: https://codereview.chromium.org/1311453007

Cr-Commit-Position: refs/heads/master@{#346849}
(cherry picked from commit f4684d1bb87fb11d93888922b1f61e52f0d61042)

Review URL: https://codereview.chromium.org/1324503003 .

Cr-Commit-Position: refs/branch-heads/2490@{#198}
Cr-Branched-From: 7790a3535f2a81a03685eca31a32cf69ae0c114f-refs/heads/master@{#344925}