Convert DeviceManagerDelegate to PermissionProvider mojo interface.
This changes the device::usb::DeviceManager service so that instead of
expecting a C++ implementation of the DeviceManagerDelegate interface it
depends on the client connecting to the devices app to provide an
implementation of the PermissionProvider mojo service interface. This
will allow the permissions checker to be injected by the FrameMojoShell
when a renderer requests a connection to the devices app.

BUG=492204

Review URL: https://codereview.chromium.org/1316203006

Cr-Commit-Position: refs/heads/master@{#347998}