Throw errors for nested targets in GN.
People sometimes nest targets or configs, usually with the assumption that
this limits the visibility of a config to within a target. But this nesting
provides no visibility restrictions over declaring it outside of a block.

For clarity, force certain types of blocks to be non-nested.

Adds missing ordering documentation to the ldflags variable.

Review URL: https://codereview.chromium.org/1314773005

Cr-Commit-Position: refs/heads/master@{#348186}