Don't Allow Service Workers for Unloaded Extensions
Extensions are interesting, in that they can be unloaded and reloaded,
and their state should be the same after this (especially since it can
happen on every browser start/stop).  Any service workers that an
extension has should not be running during the time that the extension
is not loaded, but on reload, these workers should be put in place again,
and should maintain their cache.

For this, check if the extension is loaded in AllowServiceWorkers(), and
stop any running service workers when the extension is unloaded.

BUG=519326
TBR=kinuko@chromium.org (small changes to misc content/ files)

Review URL: https://codereview.chromium.org/1313213002

Cr-Commit-Position: refs/heads/master@{#348208}