Clear webapp storage when site data is cleared
From a privacy perspective, all webapp data should be cleared
when site data is cleared by the user. Add the functionality
to do this.

Note: this CL does not delete the files themselves but no
identifying information about either the webapp or the user
is left in the SharedPreferences so there should be no problem.
This is because of a platform limitation which does not allow
SharedPreference consumers to delete the SharedPreference
file.

BUG=508627

Review URL: https://codereview.chromium.org/1329083002

Cr-Commit-Position: refs/heads/master@{#348370}