Revert special implementations of layerForScrollbarContainer.
Rebasing the clip layer to be on the visual viewport causes the
scrollbars to disappear in popups, for reasons I don't fully understand
yet.  For now, this part of the change can be rolled back, as it only
has an effect with overlay scrollbars, and therefore has little visible
impact given that I already gave up on merging in Android scrollbars.
This is something we can try to reland on the next attempt to merge
Android scrollbars.

This leaves intact the behavior change that nonoverlay scrollbars stick to
the screen on pinch zoom.  It means the scroll offset they report is the layout
viewport's (which was already the case for Blink-painted ones, which don't
use the clip layer for this calculation).

BUG=528154,523056

Review URL: https://codereview.chromium.org/1331363002

git-svn-id: svn://svn.chromium.org/blink/trunk@202165 bbb929c8-8fbe-4397-9dbb-9b2b20218538