[CSS Grid Layout] Content sized tracks must clear the override height
Content sized tracks depend on grid item's height to determine their
height. If the item was stretched and needs a new layout, we must clear
the override height to properly.

However, we can't just clear the override height because we need to know
it was set, hence we would need grid item's intrinsic height instead of
its logicalHeight in case it does not need such new layout.

BUG=530476

Review URL: https://codereview.chromium.org/1335213004

git-svn-id: svn://svn.chromium.org/blink/trunk@202196 bbb929c8-8fbe-4397-9dbb-9b2b20218538