roll libyuv from r1481 to r1483
A capturer bug causing incorrect size field was fixed, so the
ValidateJpeg can now depend on, and verify, the size field matches the
buffer.  Scanning the last 1024 bytes of the buffer is both faster,
and safer sanity check, that the buffer contains a valid jpeg image.

There is also a change to YUV to RGB conversion for arm 32 bit, which
should have no functional or performance change.

Contains the following revisions
https://chromium.googlesource.com/libyuv/libyuv/+log/602fdbef9..fcacbfb27f4

BUG=chromium:497297,libyuv:478

Review URL: https://codereview.chromium.org/1336403002

Cr-Commit-Position: refs/heads/master@{#348973}