Fix local to absolute coordinate mapping in ContainerNode::boundingBox.
In the case of an empty anchor tag, we find an adjacent LayoutText to base the
coordinates on.  If the container of that LayoutText had an overflow clip, we
failed to subtract its scroll offset when converting from local to absolute.

To fix this, we now call localToAbsolute directly on the LayoutText, instead of
starting from its container.

BUG=132290

Review URL: https://codereview.chromium.org/1346043002

git-svn-id: svn://svn.chromium.org/blink/trunk@202359 bbb929c8-8fbe-4397-9dbb-9b2b20218538