Use "fast seek" and usetoc flags for FFmpeg mp3 decoding.
Only the usetoc flag is critical to fixing the issue. Here's the story:

1. We did not previously set this flag, but a bug in ffmpeg caused it
to default to "on" anyway. Prior to commit c3a7366, the code for
parsing the flag was:

  int fill_index i= mp3->usetoc && duration > 0;

The default value for the usetoc flag was -1, so this evaluated to true
for any non-zero duration file and enabled the TOC behavior.

2. We *do* want the usetoc behavior, since this gives us fast accurate
seeking whenever the TOC is present.

3. Several commits came in with the latest ffmpeg roll that overloaded
and modified default behavior for the usetoc flag. When the dust
settled the flag defaulted to the new value of 2, which invokes the
very slow (but accurate) sequential-read-seek. This is whats happening
in the bug repro.

Key ffmpeg commits: (see mp3_read_header)
c3a7366 - default usetoc = 0 (still does fast seek for CBR)
58fade2 - default usetoc = 2 (slow seek for CBR and VBR)
748d481 - default usetoc = (s->flags & AVFMT_FLAG_FAST_SEEK) ? 0 : 2;

4. This change fixes the bug by explicitly setting usetoc = 1. This
restores the previous Chrome behavior. Both CBR/VBR use the TOC when
available. When no TOC exists CBR will use the fast seek method and
VBR will fallback to slow seeking.

5. This change also sets AVFMT_FLAG_FAST_SEEK. This really has no
effect at this time, but the next FFmpeg roll will include commit
c43bd08, which picks up this flag to enable fast (but less accurate)
seeking for VBR whenever the TOC is not found. That commit also
includes yet another tweak to defaulting which will make explicit
setting of usetoc=1 redundant when AVFMT_FLAG_FAST_SEEK is also set.

BUG=530043
TEST=manual verification of CBR and VBR files, including bug repro.

Review URL: https://codereview.chromium.org/1352573003

Cr-Commit-Position: refs/heads/master@{#349318}