Close popups on tap outside
Right now tapping outside a PagePopup will not close the popup on Linux.

The popup semetimes closes on Windows but it is because the tap removes focus from the element.

This CL cancels the PagePopup on GestureTap events.
Code in WebViewImpl::handleGestureEvent will be used on Windows, and WebPagePopupImpl::handleGestureEvent is used on Linux.

BUG=510950

Review URL: https://codereview.chromium.org/1346113002

git-svn-id: svn://svn.chromium.org/blink/trunk@202429 bbb929c8-8fbe-4397-9dbb-9b2b20218538