[Eraser] First pass at removing the notification center panel.
This just rips out the center and settings component, and does not
refactor any of the delegate/interface classes to match the new reality
(that is to be done later).

Also TODO: removing the entry from the OSX main menu (didn't want
to do xib stuff here), so for now the entry is just disabled.

R=peter,rsesek,oshima,bauerb
TBR=atwilson

BUG=530376

Review URL: https://codereview.chromium.org/1334363002

Cr-Commit-Position: refs/heads/master@{#349425}