Switch from using codec ids to codec name hashes.
Uses 32-bits of the SHA-1 hash of the codec name as the sparse
histogram identifier.  Adds a unittest for producing this enum
mapping so it can be updated easily.

FFmpeg is always changing codec ids, so using the raw AVCodecID
is proving to be untenable.

BUG=none
TEST=playback shows the right hashes.

Review URL: https://codereview.chromium.org/1348563002

Cr-Commit-Position: refs/heads/master@{#349472}