webview: Move calls to update the root scroll delegate to LTHI.
This moves the calls to update the delegate out to the LayerTreeHostImpl
instead of having it on LayerTreeImpl. This has a few benefits.

1) Things that are only for the active tree don't make the most sense
on LayerTreeImpl in general, since it's an abstraction around things
that occur on both the pending and active tree.
2) This avoids the needs to set the root scroll delegate to null and
back during activation and pinch updates. Instead of updating the root
delegate reactionarily when things change, we just call it at the end
of steps that might change the root scroll delegate's data (and early
out in BrowserViewRenderer if they are unchanged instead).
3) Avoids plumbing around concepts of the root scroll delegate through
LayerImpl and LayerTreeImpl.

R=boliu
BUG=531746
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1345353002

Cr-Commit-Position: refs/heads/master@{#349520}