Fix 32-bit icon images so that they render properly.
The 32-bit images' AND masks had transparent pixels for ares of <50%
alpha transparency. This was causing Windows to display black spots and
exhibit very bad scaling quality on these images (in particular, on the
Windows 10 taskbar).

I have regenerated the AND masks such that any pixel with >0% alpha
transparency is opaque in the mask. The icons no longer exhibit black
spots, and look good when scaled.

BUG=526622

Review URL: https://codereview.chromium.org/1355523002

Cr-Commit-Position: refs/heads/master@{#349567}