Improve back button behavior.
When Chrome loads a URL from another app that resolves to an intent that
ends up in another application, get rid of Chrome from within the back
stack before the tab is closed. This allows the user to go back to the
original application that launched it (instead of getting stuck on a
different Chrome tab, potentially with a back stack).

BUG=487938

Review URL: https://codereview.chromium.org/1346283004

Cr-Commit-Position: refs/heads/master@{#349777}