IndexedDB: Free transaction memory as write back is constructed
When committing an Indexed DB transaction to the leveldb backing
store, the in-memory transaction map is iterated to populate a leveldb
WriteBatch. The in-memory transaction is then flushed.

To avoid a big memory spike during this process (and possible OOM),
discard individual records from the transaction during the iteration.

BUG=527995
R=cmumford@chromium.org

Review URL: https://codereview.chromium.org/1358893002

Cr-Commit-Position: refs/heads/master@{#350041}