Revert of Reland debugging for ipc related crash. (patchset #1 id:1 of https://codereview.chromium.org/1354063002/ )
Reason for revert:
Supposedly, this CL only changes the signature of the crash, not the presence of the crash. There is some (but not totally clear) evidence that this CL is causing a crash itself.

https://code.google.com/p/chromium/issues/detail?id=527588#c32

Original issue's description:
> Reland debugging for ipc related crash.
>
> This CL is almost identical to https://codereview.chromium.org/1350823002/. The
> only difference is that the part of the CL that causes the crash has been
> omitted.
>
> BUG=527588
>
> Committed: https://crrev.com/a7502c6dff276bce3c76d680c5d613cb6ece19be
> Cr-Commit-Position: refs/heads/master@{#349828}

TBR=avi@chromium.org,tsepez@chromium.org
NOPRESUBMIT=true
NOTREECHECKS=true
NOTRY=true
BUG=527588

Review URL: https://codereview.chromium.org/1354863004

Cr-Commit-Position: refs/heads/master@{#350062}