MediaCodecPlayer implementation - stage 6 (respect audio permission)
We ask MediaPlayerManager for permission every time we restart
after the pause or resume playback while performing the seek.
The manager's method RequestPlay() is called on the UI
thread, therefore request is done asynchronously: we wait
on kStateWaitingForPermission until the result comes on the
Media thread.

If permission is granted we continue with prefetch, otherwise the
player goes to paused state.

BUG=407577

Review URL: https://codereview.chromium.org/1354553002

Cr-Commit-Position: refs/heads/master@{#350083}