MacViews: Add a case for Mac in TextfieldTest.KeysWithModifiersTest
The test ensures that characters are not inserted into textfields when a
system modifier is held down.

On Mac, "Command" is a modifier, but "Alt" isn't (it inserts "alternate"
characters like the German "Eszett" for Alt+s). So the test just needs
to check against a different string on Mac.

The test was added in r347514 and currently fails with

../../ui/views/controls/textfield/textfield_unittest.cc:742: Failure
Value of: textfield_->text()
  Actual: TEx134
Expected: ASCIIToUTF16("TEx234")
Which is: TEx234

BUG=471488

Review URL: https://codereview.chromium.org/1343273002

Cr-Commit-Position: refs/heads/master@{#350107}