Consider Layout tree cycles when resolving the <pattern> content element
When SVGPatternElement determines the attributes of a pattern, it will
not factor in any cycles brought upon it by the actual content elements.
The cycle solver however does this analysis when it is run on a
resource.
To bridge this DOM vs. Layout tree "gap", make sure to inspect the chain
of linked resources when determining the content element to use when
generating the pattern data.

BUG=104448

Review URL: https://codereview.chromium.org/1356053002

git-svn-id: svn://svn.chromium.org/blink/trunk@202631 bbb929c8-8fbe-4397-9dbb-9b2b20218538