DeprecatedPaintLayerStackingNode should walk DPL
https://codereview.chromium.org/1199413006 changed the way we
collect DeprecatedPaintLayers for the purpose of painting.
The change made the implicit assumptions that stacking contexts
would have DeprecatedPaintLayer, else there is no way we could
correctly paint. It turns out that this is not an assumption
that holds in our code. The reason is that some LayoutObject
just can't have DeprecatedPaintLayer (any objects that are not
LayoutBoxModelObject) or they don't want one (LayoutTableCol),
even if they should have one.

The old code worked because it didn't discriminate these cases
and was a lot more robust to the widespread abuse in the code.
Thus the logical solution is to revert the change.

BUG=527927

Review URL: https://codereview.chromium.org/1346103003

git-svn-id: svn://svn.chromium.org/blink/trunk@202653 bbb929c8-8fbe-4397-9dbb-9b2b20218538