Fix WebUSB notification on Chrome OS with no windows open.
This patch fixes WebUSB notifications so that Chrome does not crash when
opening the landing page URL on Chrome OS when there are no windows
open. This is done with a ScopedTabbedBrowserDisplayer which will create
a new browser window if one is not visible. The notification strings are
also updated to match the most recent mocks.

BUG=492204

Review URL: https://codereview.chromium.org/1360873003

Cr-Commit-Position: refs/heads/master@{#350360}