Fix a use of the non-thread-safe mojo::Callback.
mojo::Callback is not thread-safe. It must be created, copied, and
destroyed on the same thread. It is not safe to post a task to another
task runner which has a mojo::Callback as a bound parameter. Because of
this, a level of indirection is necessary to ensure a mojo::Callback is
always used on its original thread.

For BackgroundSyncClientImpl, a map is used to save the callback to the
Sync method and associate that callback to an id. That id is then safe
to be passed around between threads.

BUG=532058
TESTED=Locally augmented mojo::internal::SharedData with a ThreadChecker to
       ensure mojo::Callbacks are only referenced on a single thread.

Review URL: https://codereview.chromium.org/1363533002

Cr-Commit-Position: refs/heads/master@{#350400}