[Android, Presentation API, Cast] Proper use of MediaRouteChooserDialog events, reentrance of SessionWrapper.close()
onRouteSelected is not called when the same route is selected and that causes
onDismiss() to report the dialog being closed and the connection is terminated by MediaRouter.
Now onCancel() is used to distinguish between the user not selecting anything from
selecting a route (including the current one) in onDismiss().

BUG=524131

Review URL: https://codereview.chromium.org/1349163007

Cr-Commit-Position: refs/heads/master@{#350544}