Pass correct page URL for context menus
A regression erroneously caused a blank page URL to be passed in the
context menu creation parameters from the renderer process to the
browser, if the menu was created in a subframe.

This corrects the problem, and allows context menus to properly account
for having remote main frames in site per process mode.

BUG=534209
R=dcheng@chromium.org

Review URL: https://codereview.chromium.org/1359693003

Cr-Commit-Position: refs/heads/master@{#350595}