SyncableService::DeactivateDataType isn't needed.
SyncableService::DeactivateDataType isn't needed and should be
removed for 3 reasons:
1) It is called only from Stop() method of
old implementations of DataTypeController:
- FrontendDataTypeController
- NonFrontendDataTypeController
UI/NonUI DataType controllers don't call this method from
Stop() and work fine without it.
2) Datatype deactivation already done from
DataTypeManagerImpl::OnSingleDataTypeWillStop() which is called
just before calling the DTC::Stop() method.
3) As a result of ongoing refactoring of USS code there will
be two separate deactivation methods - one for directory
datatypes, another for USS datatypes. It would be difficult for
SyncableService::DeactivateDataType to figure out which of
the two specific deactivation codepaths to take.

Note: SyncDataTypeManagerImplTest tests still pass with
this change. There is a test that verifies deactivation of
Bookmarks data type which is based on FrontendDataTypeController.
A few tests that bypass DataTypeManager had to be tweaked
to not expect SyncableService::DeactivateDataType() to be
called when DTC::Stop() is called.

BUG=515962

Review URL: https://codereview.chromium.org/1364043002

Cr-Commit-Position: refs/heads/master@{#350634}