Remove WebContentsObserver::DidCommitNavigation
This CL removes the WebContentsObserver::DidCommitNavigation. On navigation
commit, the WebContentsObserver::DidFinishNavigation will be called, making
clear that the NavigationHandle tracks a navigation from start to finish. Note
that if the navigation committed, the document load will still be ongoing after
DidFinishNavigation is called.

BUG=504347

Review URL: https://codereview.chromium.org/1350673003

Cr-Commit-Position: refs/heads/master@{#350637}