Fixes crash in SRTFetcher when trying to check if metrics reporting is enabled.
BUG=529283

Note: The function that I called will only return true if both metrics and crash are enabled for the user. I considered adding a new function ChromeMetricsServiceAccessor::IsMetricsReportingEnabled(), but since this code seems to be subject to refactoring, I'm not sure it's a good idea to do that now. However, if the coverage of both populations (UMA enabled vs. UMA and Crash enable) is too different, it may be worth doing that.

Review URL: https://codereview.chromium.org/1356423002

Cr-Commit-Position: refs/heads/master@{#350649}