Support DomKey on Android devices.
Generate the DomKey from the actual key pressed and the unicode
character presented. Build a keycode table to return the appropriate
non-printable character if necessary.

Note that backspace, tab, escape and delete are magically mapped
provided they actually generate the printable character since DomKey
maps the printable characters to those.

Enter which generates the '\n' character on Android needs to be avoided
to convert that to the DomKey because ENTER is mapped as the '\r'
sequence.

Ctrl-<key> is not handled properly yet; see bug 527111 for more details

BUG=227231

Review URL: https://codereview.chromium.org/1356833003

Cr-Commit-Position: refs/heads/master@{#350698}