[Android] Don't depend on android support v13 library in Crosswalk Runtime Library
The commit "[BackgroundSync] Reland of Launch Chrome on Android for One-Shot" in
https://codereview.chromium.org/1140813006/#ps340001 is
On Android devices BackgroundSync needs to launch the browser the next time the device
goes online if a one-shot sync is registered.
Because olny BackgroundSyncLauncherService.java is using the android library support.
so skip it when building the component of content_java. It will be compiled
in app-side if customer want to try this feature.

BUG=XWALK-5092