[Ozone-DRM] Queue buffers onto the surface
Queue all related buffers onto the surface and only send them to the
window once we're ready to commit them.

This allows us to decouple the ScreenManager from GbmPixmap, which will
be useful once the DRM thread is in place since GbmPixmap and
GbmSurfaceFactory are a objects owned on the GPU main thread while
ScreenManager will live on the DRM thread.

BUG=536113

Review URL: https://codereview.chromium.org/1366273002

Cr-Commit-Position: refs/heads/master@{#350887}