Ensure that the global mouse position is updated to the unlocked global position prior to restoring the cursor in UnlockMouse
Currently when the UnlockMouse operation occurs in RWHVA, we move the cursor to its original position (Position at the
time of LockMouse). This triggers a synthesized mouse move in Aura, which shows up in RWHVA. The movement delta
is calculated as a difference of the current mouse event's screen position and the global mouse position which is then
sent to the renderer. However this shows up as a big delta because the global mouse position contains the last mouse move
position when the RWHVA was in a locked mouse state.

Fix is to update the global mouse position to the unlocked global position as we want status quo to be restored.

BUG=162110
TBR=jam

TEST=Manual. Please navigate to https://scheib.github.io/HTMLMisc/PointerLockLog.html. Click on Request pointer lock and allow the page to lock the mouse. Exit the mouse lock via the ESC key. There should not be a large movementX value printed on the page.

Review URL: https://codereview.chromium.org/1369663002

Cr-Commit-Position: refs/heads/master@{#350892}