Fix WebSpeech API for OOPIF based <webview>.
In OOPIF based <webview>, BrowserPluginGuest isn't used much and the
plan is to remove its usage in --site-per-process. However, the
Web Speech API relies on BrowserPluginGuest to return its embedder
WebContents. I've made it so that we ask for the WebContentsImpl to
return its embedder or outer WebContents instead of going through
BrowserPluginGuest.

TBR=jam@chromium.org for chromium.fyi.json
BUG=533647
Test=Try WebSpeech API from a Chrome App <webview>, with --site-per-process.
This should work and not crash the browser/ process.

Review URL: https://codereview.chromium.org/1351643003

Cr-Commit-Position: refs/heads/master@{#351018}