Fix incorrect window show state after session restore.
If a user quits Chrome with two windows, both docked, on restore Chrome
will notice that neither window is fully visible and select the first
window in its restoration list to make visible. However, the profile
also has a notion of an active window, which session restore auto-
matically makes fully visible. The result is that depending on the order
in which the windows were placed in the dock, both windows can end up
restored fully visible instead of only one of them (which ideally should
be the last active window).

This change causes session restore to restore the last active window to
fully visible if all windows are hidden and the last active window
appears in the window restoration list.

BUG=536201

Review URL: https://codereview.chromium.org/1366313003

Cr-Commit-Position: refs/heads/master@{#351164}