Add "Cast..." to overflow and contextual menus.
This changes adds a new entry point for the Media Router dialog by adding a "Cast..." menu item into the overflow/wrench menu as well as page/audio/video contextual menus.

This new menu item is currently only surfaced if Media Router is enabled (--enable-media-router) and if the user is not in Incognito mode. Once Media Router is re-enabled for Incognito, that check should be removed.

BUG=534841

Review URL: https://codereview.chromium.org/1360323002

Cr-Commit-Position: refs/heads/master@{#351184}