Reland of Add a NavigationThrottle to the public content/ interface
This CL relands patch https://codereview.chromium.org/1269813002/ that was
reverted because it broke some extension tests on ChromeOS. Original commit
message:

This CL adds a NavigationThrottle class to the public content/ interface. A
NavigationThrottle is used to control the flow of navigations. It lives
entirely on the UI thread. Eventually, all components that want the
functionality of a ResourceThrottle for navigations should transition to a
NavigationThrottle: the new architecture for navigations (browser-side
navigation) will not support ResourceThrottles for main resources load. See
https://docs.google.com/document/d/1ICLLQoC9EsZ-bWH4ZKRhPCIoZKn6pOj02SlGl6SKH6Y/edit?pli=1#heading=h.fmxjmgvbgg7x
for the design doc.

This Cl also transition the InterceptNavigationResourceThrottle to the
NavigationThrottle model.

BUG=504347
TBR=mnaganov@chromium.org, sky@chromium.org, asargent@chromium.org, creis@chromium.org, nasko@chromium.org

Review URL: https://codereview.chromium.org/1363483007

Cr-Commit-Position: refs/heads/master@{#351284}