Do not perform a non-incremental compile when previous compile failed.
This reverts a not-fully-thought-out change to md5_check.py from
http://crrev.com/d8cef607861

Compiles fail for legitimate reasons, and we shouldn't delete md5
information in this case.

The concern behind that logic in the first case was that buggy --incremental
logic might be the reason for compile to fail. So, this commit also adds
a new environment variable to force all targets to be rebuilt:

FORCE_REBUILD=1

BUG=533442

Review URL: https://codereview.chromium.org/1374263002

Cr-Commit-Position: refs/heads/master@{#351345}