Implement the IsLightDump api in chrome side.
This implements new SkTraceMemoryDump interface changed in
www.crrev.com/1310123007.

The dump manager in base decides to request a heavy or light dump
depending on the user's configurations. The light dumps are more
frequent than heavy ones, but has onl the totals from the dump
providers. This is to enure that the trace is not huge as well as the
dumps are frequent enough in the traces. This change passes the dump
args that is provided on OnMemoryDump call to the SkTraceMemoryDump
object, which implements RequestDetails method.
The discardable details are enabled always for consistency across
dumps.

It also changes the argument in constructor to string from char* since
it is not cler if the constructor stores it or copies it.

TBR=reveman@chromium.org
BUG=499731
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1324453008

Cr-Commit-Position: refs/heads/master@{#351541}