Exposing a new easyUnlock.setupConnectionGetDeviceAddress JS API.
This CL exposes a new JS API to find the bluetooth address for the
remote device corresponding to a given connection id.

This is necesssary as the bluetooth address of a device changes after
it's paired (but it's updated in the underlying
BluetoothLowEnergyConnection object).

BUG=528277

Review URL: https://codereview.chromium.org/1370133005

Cr-Commit-Position: refs/heads/master@{#351549}