[Sync] Fix signin issues for Android tests.
This change introduces the SigninTestUtil, which handles common tasks
for configing the test signin environment and adding test accounts.

Something with the AccountTrackerService seems to have broken most of
the Android sync tests, at least on M. The seedSystemAccounts() call has
a callback that would crash the test if ChromeSigninController had an
account name set when it happened. SigninTestUtil properly ensures that
the AccountTrackerService is updated at the same time a new account is
added and signed in.

Additionally, launchActivityWithIntent() seems to never return on M so
this change uses a different way of starting the PassphraseActivity.

BUG=517590

Review URL: https://codereview.chromium.org/1365253002

Cr-Commit-Position: refs/heads/master@{#351635}