Make sure official GN builds on Linux have debug info.
On all platforms, we want the official builds to have
debug info so that we can get sensible stack traces. On non-Linux platforms, this was already the case, since the default is to contain symbols, but on Linux, we had a special case where non-debug builds did not have symbols.

This patches fixes that so that only non-debug, non-official Linux bugs do not have symbols.

R=brettw@chromium.org
BUG=530733

Review URL: https://codereview.chromium.org/1373323002

Cr-Commit-Position: refs/heads/master@{#351686}