Only (and always) put URLs in NSLinkAttributeName.
When the NSLinkAttributeName attribute is used, AppKit automatically turns it into a draggable link. If anything but a URL is used, that link then becomes a bomb to explode drag targets. This change ensures that only URLs are put in there, and that a URL is always put in there.

BUG=528228
TEST=existing tests

Review URL: https://codereview.chromium.org/1364113003

Cr-Commit-Position: refs/heads/master@{#351805}