Renaming CRD host OS details key to be constant with usage.
I added new keys to the xmpp packet heartbeat heartbeat.os-name which
at the time was suppose to replace heartbeat.log.os-name but it is not
constant with naming for the protos and variables and is easily confused
with heartbeat.log.os-name. Renaming the key to heartbeat.host-os-name for
constancy.

R=sergeyu@chromium.org
BUG=534902

Review URL: https://codereview.chromium.org/1375223004

Cr-Commit-Position: refs/heads/master@{#351835}