Back up and restore RuntimeEnabledFeatures during testing.
For use in preventing REF mutations made as part of a LayoutTest from leaking across tests. More background:

While many REFs are only effective when set at startup, it should be feasible to build a feature behind a REF and flip it via LayoutTest JS. In the interest of self-protection, we reset REF state at the end of each test to whatever it was at the beginning.

The REF::Backup class nomenclature is modelled loosely after InternalSettings::Backup as seen in:

https://chromium.googlesource.com/chromium/src/+/66515dbbfcd5701ebbc7c49e1ac8c046ffc117c9

BUG=534429

Review URL: https://codereview.chromium.org/1377683002

Cr-Commit-Position: refs/heads/master@{#351874}