Make naming of task related classes and variables consistent.
Before, both action and task terms are used for tasks. This CL renames all
of action occurences to task, with few exceptions where it wouldn't make
sense.

The reason for that is upcoming ActionsModel class which uses a term action
internally in many places, and it would cause confusion.

TEST=Tested manually + all current tests should pass.
BUG=None

Review URL: https://codereview.chromium.org/1372293004

Cr-Commit-Position: refs/heads/master@{#351956}