[DocumentThreadableLoader] Don't touch member variables after clearResource()
Cancelling resource loading may result in synchronous invocation
of "load" event. E.g. handlePreflightFailure() calls clearResource()
to release the RawResource, but it leads to invocation of body's
onload handler, and in the onload handler we can call
XMLHttpRequest.open() and it clears the DocumentThreadableLoader
instance away. And then, we're back to handlePreflightFailure()
(clear() function) and see that the member variables are already
destructed.

This CL reorders the member clean up code and clearResource() call
so that we don't touch the members after clearResource().

BUG=534992
R=hiroshige,mkwst

Review URL: https://codereview.chromium.org/1375263002

Cr-Commit-Position: refs/heads/master@{#351995}