Mask the AccountManager{Future,Callback} with a simple Callback.
This abstracts the implementation of SystemAccountManagerDelegate so
that it is easier to replace it with other implementations.

This will facilitate child account detection in Android M.

This cl also deprecates the existing synchronous getAccounts method
and adds an asynchronous method instead though this is presently
only used in ChildAccountService.

Depends on:
http://crrev.com/1368663002
http://crrev.com/1369503002

BUG=534377,517697

Review URL: https://codereview.chromium.org/1353393002

Cr-Commit-Position: refs/heads/master@{#352016}