Fix build conditions for minidump_stackwalk et al. in GN
Some of the breakpad tools are not meant to be built as iOS (or android)
targets, and do not work on Windows; however, the conditions for
them were incorrectly specified in the build file. This patches
fixes that to make it clear that they should only be built with the host
toolchain.

R=brettw@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1378463005

Cr-Commit-Position: refs/heads/master@{#352075}