TextAutosizer should handle cross-process navigation and RemoteFrame.
In https://codereview.chromium.org/1374783005/ the fix tested for whether
the frame has a FrameView or not. The problem turns out to be a bit
different - the check for RemoteFrame and the mainFrame do not point
to the same object, which leaves the code still open to problems.

This CL is ensuring that both the check for RemoteFrame and the usage of
the main LocalFrame point to the same object.

BUG=357747

Review URL: https://codereview.chromium.org/1375613004

Cr-Commit-Position: refs/heads/master@{#352249}