WebMediaPlayerAndroid: create all texture bits if needed on a MediaPlayerPlay signal.
When the MediaPlayerPlay signal is sent, the WebMediaPlayer might no
longer have a valid stream texture proxy and peer for it. We need to
make sure it is correctly set, the same way would if the user presses
the play button.

A real example where this fails is if a MediaSession resumes a video
that was being played fullscreen and left the fullscreen state. It will
resume in a non fullscreen state but the WebMediaPlayerAndroid instance
will not be reset correctly.

BUG=535137

Review URL: https://codereview.chromium.org/1384273002

Cr-Commit-Position: refs/heads/master@{#352380}