gpu: Remove Skia Runner
skia_runner was intended to allow skia to test the
command buffer on their servers, but we have since created
command_buffer_gles2 library that they can use as a backend.

This code is just wasting compile time, possibly
maintenance, etc.

:'(

(it's in history if we ever need it again)

Review URL: https://codereview.chromium.org/1381913003

Cr-Commit-Position: refs/heads/master@{#352454}