Reconstruct customscrollbars if owningLayoutObject changes.
CustomScrollbars owner changes when the body's LayoutObject is attached
which occurs during style recalc (before layout). However in the case
of overflow:scroll, the scrollbars are created when we call
styleDidChange() on the LayoutView, which happens even sooner than
that. So during updateAfterLayout,in needsScrollbarReconstruction if
owningLayoutObject changes for customscrollbar reconstruct the
scrollbar.

BUG=522389

TESTED=scrollbars/custom-scrollbar-inactive-only-on-windowinactive-selector-expected.html
now passes with root layer scrolling

Review URL: https://codereview.chromium.org/1307173007

Cr-Commit-Position: refs/heads/master@{#352644}