Fix SVGUseElement "instance tree loading" status check
SVGUseElement::instanceTreeIsLoading was not checking the loading status
of the immediate <use> targer (the |targetInstance| argument to the
method.) It also wasn't checking the correct element for loading status,
since correspondingUseElement() will return the containing (innermost)
shadow host - which is not the <use> that will carry the relevant
loading status.
Make sure to check the immediate target, and check the loading status of
any <use> elements encountered (instead of attempting to look at the
"corresponding <use>").
Also rewrite the checker method to use Traversal<...>::next instead of
being recursive, and simplify resourceIsStillLoading() some.

BUG=232901

Review URL: https://codereview.chromium.org/1388093002

Cr-Commit-Position: refs/heads/master@{#352645}