Add a size estimation mechanism to StoragePartitionHttpCacheDataRemover.
StoragePartitionHttpCacheDataRemover, the class that enables cache clearing based on time constraints [begin, end), is now able to estimate the size in bytes of entries that would be deleted, without actually deleting them.

The actual computation was so far implemented only for the Blockfile cache backend. Simple and Memory backends to be done in follow-up CLs (but the infrastructure is in place, since disk_cache::Backend is purely virtual).

Update: I missed Backend V3 and testing subclasses. They were added in Patch 3.

Added one browsertest for this functionality.

BUG=510028

Review URL: https://codereview.chromium.org/1304363013

Cr-Commit-Position: refs/heads/master@{#352830}