[css-grid] Implement grid gutters
Authors can now specify the gutters between grid lines, i.e., the space
between two consecutive grid lines. This can be done using the new
'grid-column-gap 'and 'grid-row-gap' properties (or the 'grid-gap'
shorthand).

From the track sizing algorithm POV, gutters are treated as fixed size
columns. The primary consequence is that grids are enlarged (depending on
the number of tracks). Gutters also affect the sizing of content-sized
tracks and fr tracks as long as the grid have spanning items. Those tracks
will become smaller as gutters will consume part of the item's size, so the
tracks won't need to grow as much as they used to.

Added several new test cases to verify that gutters are properly considered
when sizing and also to check that they do not modify the current behavior.
As many existing tests were reused I took the chance to refactor some
testing code related to alignment so that it could be reused by many
different tests.

Review URL: https://codereview.chromium.org/1309513008

Cr-Commit-Position: refs/heads/master@{#352839}