Presentation API: Presentation interface shouldn't be an EventTarget.
The spec changed (moved events to PresentationRequest) but we forgot to
remove the EventTarget inheritance. This CL is a simple one that removes
the web visible EventTarget inheritance so we can cherry-pick to M47.
Another CL will do the cleanup in the C++ code.

BUG=539883

Review URL: https://codereview.chromium.org/1388103002

Cr-Commit-Position: refs/heads/master@{#352871}