Gives 1px preference to docking on the right side of screen
I wanted a centered window that has kShowStateKey property set to
ui::SHOW_STATE_DOCKED to get docked consistently on the right edge of a
screen. This was not the case when <screen width> + <window width> was
an odd number which could result in tests flakiness. This change corrects
the calculation such that a window with an odd width centered in an
even-wide screen (which is technically nearer to the left edge) would
still get docked on the right when getting state change to
ui::SHOW_STATE_DOCKED.

BUG=539417

Review URL: https://codereview.chromium.org/1397513002

Cr-Commit-Position: refs/heads/master@{#352884}