[Media Router] Show existing issue when dialog is opened.
Currently, any issue that exists before the Media Router dialog is opened will not be sent to the dialog as part of the initial state. This change registers the IssuesObserver for the UI when the dialog is opened to check for any existing issues. It is unregistered when the dialog is closed and UI destroyed.

This change pulls the (un)registration of the IssuesObserver into separate functions so it's not explicitly tied to constrution/destruction of the observer.

BUG=538239

Review URL: https://codereview.chromium.org/1371423003

Cr-Commit-Position: refs/heads/master@{#352897}