[Android] Fix a timing issue on startup that could leave the UI inoperable.
If you start up Chrome on Android, immediately hit the home button,
wait long enough for the native libraries to fully load, and then return
to Chrome, it will sometimes leave the UI unfunctional with no tab
content shown.  You'll be able to focus the omnibox, but nothing beyond
that works.

This was caused by a slightly different timing of the predraw listener
used to signal first draw and the activities logic of determining
whether the first draw occured.  This removes the activities handling
and makes the predraw listener the only signal.

BUG=538341

Review URL: https://codereview.chromium.org/1384373002

Cr-Commit-Position: refs/heads/master@{#352934}