[css-grid] Include freeSpace in GridSizingData struct
During the layout process we keep the free space for rows and columns in two
variables that are passed to a few methods along with the GridSizingData
struct.  Clearly those two variables should be part of GridSizingData as
they're temporary values used just for the sake of the layout.

Review URL: https://codereview.chromium.org/1387063003

Cr-Commit-Position: refs/heads/master@{#353035}