Makes TabStrip give extra space to first tabs
If there is extra space when sizing tabs (this happens if
available_width / num_tabs * num_tabs != available_width) we now give
the extra space to the first tabs.

Prior to this change it was a bit random as to which tabs would get
the extra space.

BUG=533540
TEST=see bug
R=pkasting@chromium.org

Review URL: https://codereview.chromium.org/1372373003

Cr-Commit-Position: refs/heads/master@{#353058}