Fix more GN check issues in chrome/common:constants
Moves widevine_cdm_constants out of chrome/common:constants and into chrome/common. This has a dependency on widevine. It was included in common:constants for the path. But this path was used only once globally in chrome's path_service. It's reasonable in this case to just inline it (the whole point of the path service is to centralize the path constants).

chrome/installer/util and chrome/common include headers from each other. This adds an allow_circular_include annotation to tell GN about this unfortunate situation.

Moves the ui/base paths one to be an Android-only dependency as described in the BUILD.gn file.

Review URL: https://codereview.chromium.org/1397923002

Cr-Commit-Position: refs/heads/master@{#353177}