Refactor the API for setting dynamic resource load priorities
Currently, only images can change priority after blink sends
the resource load out to chromium. After each layout, we
iterate through a registry of LayoutObjects with associated
images and update those images' priority based on whether
the LayoutObject is in the viewport and how much of the
viewport it fills.

Change this API to be more general, so that it could more
easily support other resource types and other factors for
updating priority.

Move the current logic to ResourceFetcher, which has a list
of all in-progress resource loads. For each, ask the
associated ResourceClients for updated priority information.
This should not entail any behavior change at this time.

BUG=536147

Review URL: https://codereview.chromium.org/1378743002

Cr-Commit-Position: refs/heads/master@{#353188}