[android-tools] add memtrack_helper for GL/EGL memory profiling
This CL introduces a memtrack_helper daemon. It is NOT intended
to be shipped in production. It is supposed to be used as a
helper process when running telemetry + memory-infra or for
local profiling workflows.
The helper is supposed to be start via adb shell and makes
mtrack graphics information accessible via a UNIX socket.
See crrev.com/1383733002 for the client counterpart in Chrome.

More details in the design-doc: https://goo.gl/4Y30p9

BUG=529451

Review URL: https://codereview.chromium.org/1383583002

Cr-Commit-Position: refs/heads/master@{#353501}