Merge the simple support string experiment to the counters experiment.
The simple support string experiment was detached from the browsing data counters experiment in case it needed to be launched sooner. This is no longer the case, so we're merging it back.

Furthermore, this CL changes how the Help Center article link is shown. Instead of linkifying the "will not be cleared" phrase at the end of the support string, we instead append a "Learn more" link. Since we now have a "Learn more" link in the footer, we also need to hide the one that previously existed in the main area of the dialog. The new "Learn more" link will now show an article explaining both deleted and non-deleted datatypes, instead of the original plan that the "Learn more" link shows what is deleted and "will not be cleared" link what is not.

BUG=535890

Review URL: https://codereview.chromium.org/1389303003

Cr-Commit-Position: refs/heads/master@{#353515}