Add an OrientationIterator class for vertical text segmentation
This is an essential component for later landing a RunSegmenter class
that helps pre-segmenting the text in preparation for text shaping. The
OrientationIterator returns the next longest consecutive run of text
that does or does not need upright rendering in vertical layout, see the
gtest file for details.

BUG=542181
R=eae,behdad,kojii

Review URL: https://codereview.chromium.org/1395333003

Cr-Commit-Position: refs/heads/master@{#353523}