[Extensions Toolbar] Stop using visibility pref for visibility in the toolbar
We were using the BrowserActionVisibilityPref as an indication of whether or not
to show the icon in the toolbar, vs in the overflow menu. Unfortunately, this is
very confusing for switching between having the redesign switch on and off, as
actions which were overflowed then need to be re-shown from the
chrome:extensions page.

BUG=538786

Review URL: https://codereview.chromium.org/1390003003

Cr-Commit-Position: refs/heads/master@{#353579}