Add a FontFallbackIterator to help with shaper driven segmentation
When changing HarfBuzzShaper to let the shaper try shaping as much as
possible with the initial font, then fall back through the list of web
fonts, selected by unicode-range, and so on, we need an interface for
the shaper to access the next available font. Let's add
FontFallbackIterator for that.

BUG=542666
R=eae,behdad

Review URL: https://codereview.chromium.org/1399413002

Cr-Commit-Position: refs/heads/master@{#353754}