mus: Fix nested iframe compositing
A View's Surface can reference Views that are multiple
levels below it in the view tree if there is a chain of
local frames within the page before hitting an out of
process iframe.

Every frame gets a View, but only out-of-process iframes
get surfaces as well.

This CL fixes the multi-level iframe nesting bug my making
sure that each surface is only referenced once. Each
ServerView keeps track of referenced views in its
compositorframe, and that list of views referenced is grown
in DrawViewTree to ensure that we don't create
SurfaceDrawQuads for views that have already been
referenced by descendants.

This CL also ensures that each CompositorFrame can only
reference views within its subtree, fixing a TODO.

Finally, this CL also allocates a surface ID for a view as
soon as an ancestor wishes to display it on screen. That
way the child view can submit CompositorFrames to that
surface once it has content to submit without forcing the
embedder to submit a new compositorframe first.

BUG=537458, 533029

Review URL: https://codereview.chromium.org/1381843004

Cr-Commit-Position: refs/heads/master@{#354091}