Get searching forward/backwards working.
This adds next/prev buttons to the find bar, and lets the user navigate
with find. We properly wrap in the one frame case, and in the majority
of multiframe cases.

There are minor edge cases still untouched. We also don't properly
handle wrapping around the document when the last frame in the tree
doesn't match.

BUG=496978

Review URL: https://codereview.chromium.org/1396183003

Cr-Commit-Position: refs/heads/master@{#354100}