cc: Rename PositionImage to DrawImage and remove rect.
This patch removes the rect from position image since it might be
misleading once we scale the matrix to account for contents scale.

It also removes position image to draw image to better reflect that this
struct stores an images with a matrix that is going to be used to draw
it.

R=danakj
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1403643004

Cr-Commit-Position: refs/heads/master@{#354156}