Improve error code for service worker bailing due to no update found
ServiceWorkerWriteToCacheJob is the URLRequestJob responsible for
fetching and writing the updated script. It fails with network
error when it wants to abort the update because the new script is
the same as the old one.

Currently that results in ERR_FAILED errors appearing in the
DevTools console and netlog, which is confusing and hard
to debug because that error also occurs for actual network errors.
This patch changes the error to FILE_EXISTS, so it's
more clear why the job "failed".

BUG=541797

Review URL: https://codereview.chromium.org/1404753002

Cr-Commit-Position: refs/heads/master@{#354218}