Cleanly integrate //components/metrics/profiler on iOS
This CL cleanly integrates //components/metrics/profiler on iOS. The main work
is decoupling TrackingSynchronizer from content::ProfilerController. To do so,
this CL introduces TrackingSynchronizerDelegate, an interface that abstracts
TrackingSynchronizer's interactions with the platform on which it's operating.
It also introduces //content and iOS implementations of
TrackingSynchronizerDelegate to cleanly support the functionality that was
previously present in TrackingSynchronizer controlled by #idfef's.

The remaining work in this CL is to update buildfiles now that
//components/metrics/profiler is free of //content dependencies and can be
cleanly shared on iOS.

BUG=529195, 472210
TBR=jochen

Review URL: https://codereview.chromium.org/1401043003

Cr-Commit-Position: refs/heads/master@{#354249}