Restart the browsing data counters when the CBD dialog is opened.
This is a followup to https://codereview.chromium.org/1342093003/. In that CL, we restarted the dialog on the OptionsPageUIHandler::InitializePage event, which means that the counters were restarted even if the dialog was reloaded (which doesn't trigger OptionsPageUIHandler::InitializeHandler()).

It turns out that this is not sufficient. The correct event to listen to is the Page.didShowPage in JavaScript. This is called whenever the dialog is opened, regardless of whether it happened by clicking a button, navigating directly to the page, or reloading the page.

BUG=510028

Review URL: https://codereview.chromium.org/1404793007

Cr-Commit-Position: refs/heads/master@{#354266}