Plumb NotifierId to status icon balloons
When constructing Notifications, a NotifierId must be passed indicating which component of Chrome is responsible for the notification.

Previously, as of https://codereview.chromium.org/1395093002, a status icon DesktopNotificationBalloon would use a fixed string for the NotifierId.

But DesktopNotificationBalloons is just a middle-man, that could actually be used by various components within Chrome. Thus, this patch adds a NotifierId parameter to StatusIcon::DisplayBalloon, which is provided by the calling component.

BUG=542232

Review URL: https://codereview.chromium.org/1387383004

Cr-Commit-Position: refs/heads/master@{#354286}