Introduce IOSChromeMetricsServiceClient
This CL introduces a MetricsServiceClient implementation intended for the
//ios/chrome embedder. This implementation is based on
ChromeMetricsServiceClient (which the iOS port previously used a forked version
of), but is adapted for the needs of iOS.

Notable changes from ChromeMetricsServiceClient:
- All notification listening is eliminated. The three notifications that are in
fact fired on iOS are TAB_PARENTED, LOAD_STARTED, and LOAD_STOPPED. Listening
for these is replaced by usage of TabParentingGlobalObserver and
GlobalWebStateObserver respectively, whose callbacks fire at identical times to
the notifications.

- The asynchronous flow for initializing system profile metrics is greatly
streamlined, as the iOS port doesn't use most phases of the flow.

- ChromeMetricsServiceClient::RecordCommandLineMetrics() is not ported as it (a)
relies on //chrome and (b) does not currently gather meaningful metrics. See
crbug.com/375794 for details.

NOTE: In the codereview page for this CL, PS1 has the forked downstream copy of
chrome_metrics_service_client.* that the iOS port was using at the time of
creating this CL in order to ease the task of doing before-and-after
comparisons.

BUG=512421

Review URL: https://codereview.chromium.org/1398443009

Cr-Commit-Position: refs/heads/master@{#354324}