Use new earcons in ChromeVox Next.
Add a new Earcon class that generates sounds using
the new cvox2 EarconEngine. Swap it in dynamically
when we're in a ChromeVox Next mode. Add support for
testing earcons to MockFeedback and write an initial
test that all of the correct earcons are outputted
for controls.

BUG=491243

Review URL: https://codereview.chromium.org/1319093003

Cr-Commit-Position: refs/heads/master@{#354366}