Add extensions code to use common updater in components/update_client/
This is another piece of work towards implementing differential
extensions update and share autoupdate code with the chrome components
system. It adds code to the extensions/browser directory that makes use
of the common autoupdate code via the update_client::UpdateClient class
to send update checks to the server, download a full .crx or
differential update, and hand back an unpacked directory with the
contents of the new extension version.

This CL does not yet add any code in chrome/ that actually uses these
new capabilities; that will be in a forthcoming CL.

BUG=490418

Review URL: https://codereview.chromium.org/1362043005

Cr-Commit-Position: refs/heads/master@{#354370}