Remove switch statements on ContentSettingsType in website_settings{_ui}.cc
This removes some switch statments on ContentSettingsType as we plan to make
ContentSettingsType not an enum. In website_settings_ui.cc it moves the
constants into a static array which is searched instead. This gives a bit of a
code cleanup. In website_settings.cc it converts them to if/else statements. In
that case, in the long run, both if/else statements will be removed.

BUG=526932

Review URL: https://codereview.chromium.org/1393763007

Cr-Commit-Position: refs/heads/master@{#354724}