Add a cache counter to the clear browsing data dialog.
Currently, cache counting is only implemented for the infinite time interval (i.e. the entire cache).

The solution is that in the UI, we will always show the same value (size of the entire cache), but mark it as either exact value or the upper bound depending on the time interval selected. Therefore, this class currently ignores the time period supplied by the UI, and always requests the calculation of size of the entire cache.

BUG=510028

Review URL: https://codereview.chromium.org/1402983002

Cr-Commit-Position: refs/heads/master@{#354834}