Adds logic to close always on top dialogs before/after running tests
If a test (or child process) triggers an assert the runtime shows a
dialog. The system dialog is always on top and not tied to the length
of the program. Because interactive ui tests generate mouse events
this always on top system dialog may interfere with future tests and
cause all sorts of flake.

This patch runs code before/after tests. If it finds an always on top
dialog it closes it (CloseWindow()). If a find any other unknown
always on top windows the class name is logged.

BUG=544037, 523255
TEST=test only change
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/1412443005

Cr-Commit-Position: refs/heads/master@{#354858}