media: DecoderSelector supports generic decoders that can handle encrypted streams.
Here's the new behavior when the stream is encrypted:

if (not_on_android && decryptor_supports_decrypt_and_decode) {
  ChooseDecryptingVideoDecoder();
} else if (decryptor_supports_decrypt_only) {
  ChooseTheFirstDecoderThatSupportsTheDecryptedStream();
} else {
  DCHECK(no_decryptor);
  ChooseTheFirstDecoderThatSupportsTheEncryptedStream();
}

With this CL, we can play streams encrypted with Clear Key key system using
WebMediaPlayerImpl and GpuVideoDecoder on Chrome for Android.

This CL also updates GpuVideoDecoder so that it supports encrypted stream on
Android. This is needed to support other key systems where we need to use
MediaDrm in the GPU process.

BUG=542431
TEST=Updated/Added unittests.

Review URL: https://codereview.chromium.org/1403953003

Cr-Commit-Position: refs/heads/master@{#354886}