Reland "[Android] Call Tab.hide() when the activity is stopped"
There's some residual bookkeeping that relies on the Tab.hide() signal
being called when Chrome is backgrounded. Restore that call when
the activity is stopped.

This change originally landed in r354532, but was reverted due to
breakage of TabTest#testTabRestoredIfKilledWhileActivityStopped. This
was due to some incorrect assumptions in the test validations, which
have been resolved.

BUG=434401

Review URL: https://codereview.chromium.org/1387433003

Cr-Commit-Position: refs/heads/master@{#354888}