Oilpan: Fix crashing fullscreen-plugins-dont-reload.html caused by r353782
The crash is happening because the Widget is disposed when the m_widget->m_persistedPluginWidget transfer happens. When the updateWidgetTimer is fired after the reattach happens, WebPluginContainerImpl::show is called on an already disposed object which causes the crash.

The adds releasePersistedPluginWidget and releaseWidget which can be used to transfer the Widgets without calling dispose() on them.

BUG=None

Review URL: https://codereview.chromium.org/1407193003

Cr-Commit-Position: refs/heads/master@{#355010}