HTMLOptionElement::text should not refer to |label| content attribute.
https://html.spec.whatwg.org/multipage/forms.html#dom-option-text
> The text IDL attribute, on getting, must return the result of stripping and
> collapsing whitespace from the concatenation of data of all the Text node
> descendants of the option element, in tree order, excluding any that are
> descendants of descendants of the option element that are themselves script
> elements in the HTML namespace or script elements in the SVG namespace.

New behavior matches to the standard, IE11, and Firefox.
Also, we confirmed IE11 didn't refer to |label| attribute in the standard mode
and the quirk mode.

Implementation:
There are several internal usages of text(), and changing the text() behavior has
risk of regressions.  So, this CL introduces textForBinding() for |text| IDL
attribute with the new behavior, and keeps the |text()| with the old behavior.

BUG=429293

Review URL: https://codereview.chromium.org/1416833002

Cr-Commit-Position: refs/heads/master@{#355044}