Make "Courier New" first choice in font-face-unicode-range.html
"Courier" is an AAT font on Mac, leading to problems with
unicode-range: subsetting on Mac, when using the new shaping
code. These issues are captured in bug 545555. However, in order to
avoid disabling this test-case, "Courier New" is made the primary font
choice, leading to an OpenType font being picked on Mac.

BUG=545557
R=eae,behdad

Review URL: https://codereview.chromium.org/1420583002

Cr-Commit-Position: refs/heads/master@{#355241}