Turn on VideoRendererAlgorithm for WebMediaPlayerMS by default.
The algorithm is intended to increase the rendering smoothness for WebRTC. Previously, it is behind flag --enable-rtc-smoothness-algorithm. Now it is on by default, and you can disable it by using flag --disable-rtc-smoothness-algorithm.

The targeted workflow has been tested for a couple of months, and it went well. Now we've decided to bring it on to increase the test population, and let the code path to go over more comprehensive test on multiple platforms and for other workflows to make sure nothing is broken.

BUG=514873

TEST=Keep an eye on Cast, as it uses WebMediaPlayerMS, but uses a different code path than webrtc to get frames.

Review URL: https://codereview.chromium.org/1413163003

Cr-Commit-Position: refs/heads/master@{#355303}