Estimation of loading task cost to use 98th percentile.
The cost of loading tasks is strongly bimodal.  There are typically
hundereds (if not thousands) of very cheap loading tasks and a handful
of super expensive ones.  By increacing the sample count to 1000 and
the estimation percentile to 98th we get a more realistic worst case
estimate.

The effect of this change will be to block loading tasks much more
frequently if we are expecting a main thread touch start. This should
reduce first_gesture_scroll_update_latency on the
smoothness.scrolling_tough_ad_cases benchmark.

There may also be minor improvements to other smoothness benchmarks.

BUG=497761, 544185

Review URL: https://codereview.chromium.org/1422703002

Cr-Commit-Position: refs/heads/master@{#355573}