AltGr keys were not working.
Introduced by https://codereview.chromium.org/1409823002
the masking of control characters caused a regression on
Windows. Windows maps AltGr to Ctrl-Alt; so natively
the Ctrl-Alt are both set causing the character to be masked
out when a AltGr key was entered.

Ensure that AltGr is not down.

BUG=529883,546464

Review URL: https://codereview.chromium.org/1414853004

Cr-Commit-Position: refs/heads/master@{#355794}