Fix incorrect window size on session restore.
Under the right conditions, a window that is restored during a browser
restart and never moved during that session will have an empty bounds
in session restore during the next browser restart. As a result that
window will be assigned the default browser window bounds, which could
be much different (in the steps in the referenced bug the restored
window jumps from its original size to the full size of the screen).

Normally whenever a window's size is changed, even during window
restoration, its size is sent to session restore, but at start-up the
window may not be fully initialized with the session service and so
session restore ignores the update. The solution is to ensure that
the window always updates its size with session restore when it
becomes main (it was almost always doing that previously; the case
where it wasn't is what led to the bug).

BUG=536280

Review URL: https://codereview.chromium.org/1370753003

Cr-Commit-Position: refs/heads/master@{#355803}