mac: In static library builds, link against a static libc++.a
To achieve this, just add a -Lthird_party/libc++-static flag to the link line,
and add a postbuild that checks that Chromium Framework depends on neither
libstdc++.dylib nor libc++.dylib Use the existing verify_order postbuild
for this, and let it not run in component builds (since what it checks for
isn't interesting in that config, and we do depend on system libc++ in
component builds).

This change is small but subtle, see thread
"[chromium-dev] Intent to implement: Statically linking libc++ for Chrome/Mac"
and the document linked from comment 14 on the bug for details.

Ideally, this has no observable behavior change. If it looks like this
breaks tests somewhere, especially on 10.6, please revert. (The bots
like it, and the things I tried on 10.6 worked too, though.)

BUG=400091
R=mark@chromium.org

Review URL: https://codereview.chromium.org/1413863003 .

Cr-Commit-Position: refs/heads/master@{#355826}