ui: Add RGBX_8888 buffer format.
This adds a new RGBX buffer format which is the RGBA equivalent of
BGRX. This is useful to have when importing buffers and it makes
sense to have this when we have BGRX.

This also makes our shared memory implementation of GpuMemoryBuffers
and GLImages support these formats without an alpha channel. This is
useful for testing purposes and not having to special case BGRX makes
for cleaner code.

The GLImage unit test have been updated to test all 32bpp buffer
formats.

BUG=538325
TEST=gl_unittest --gtest_filter=GLImage*/GLImageCopyTest/*.CopyTexImage

Review URL: https://codereview.chromium.org/1415723006

Cr-Commit-Position: refs/heads/master@{#355839}