[sync] Set up notification-free infra for triggering local sync refresh
Sync code that is due to be componentized and shared with iOS both sends and
listens for the SYNC_REFRESH_LOCAL notification. This is problematic, as iOS
cannot use the //content-based notification system.

This CL puts the pieces in place for getting rid of this notification: It adds
an interface to SyncService that a client can call to trigger refresh. Future
work can change all senders of the notification to instead call into the
SyncService and then eliminate the notification altogether.

BUG=512779,512059

Review URL: https://codereview.chromium.org/1418513005

Cr-Commit-Position: refs/heads/master@{#356021}