Assume CheckMediaAccessPermission() checks are for Flash when from HTTP origins
PermissionBubbleMediaAccessHandler::CheckMediaAccessPermission has no knowledge
of whether or not the request is from Flash - the information isn't plumbed
through at all. This means that when it does its permission check, it can get
the wrong value (this bug has existed for years). We used to assume the request
was coming from Flash, but that changed in
https://codereview.chromium.org/1300603002 to assuming it was coming
from WebRTC. That broke device enumeration in Flash on http because Flash is
more lenient with allowing the mic/cam permission to be persisted in http sites.

Fortunately, post-M47 WebRTC requests can't be made from http origins at all so
we can assume that all http requests are for Flash to work around this. This
special case should be removed after HTTP access for Flash has been deprecated
(crbug.com/526324).

BUG=542414

Review URL: https://codereview.chromium.org/1411993005

Cr-Commit-Position: refs/heads/master@{#356203}