ui: Add support for creating GLImage instances from shared memory pools.
This adds an offset field to the GpuMemoryBufferHandle struct that
allows shared memory backed GLImages to be created from a pool of
shared memory. This is useful when importing buffers as it allows
the clients to manage the memory used for buffers.

BUG=538325
TEST=gl_unittests --gtest_filter=GLImageSharedMemoryPool/GLImageCopyTest/0.CopyTexImage
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1417363006

Cr-Commit-Position: refs/heads/master@{#356212}