Update version parameter format in SpdyAltSvcWireFormat.
Update version parameter of alternative services to a quotation mark enclosed,
comma separated list of positive integers (from a single naked integer).  Also
place it at the end of the list both as a member and when serialized.  This
makes sense because it is an experimental parameter (as opposed to, for example,
max-age, which is serialized first).  It was following port previously for more
efficient packing, but now that it's a vector, that became moot.

This CL lands server change 106303839 by bnc.

BUG=392575

Review URL: https://codereview.chromium.org/1410343005

Cr-Commit-Position: refs/heads/master@{#356300}