Mac: Always use surfaceless mode
The goal of this change is to unify all image transport surfaces on Mac.

Prior to this, if we do not support remote CoreAnimation APIs, then we
will create an ImageTransportSurfaceFBO, which will send the IOSurface
that it creates in the GPU process to the browser process.

Change this to use the ImageTranportSurfaceOverlayMac. Using this
surface will trigger the use of
GpuSurfacelessBrowserCompositorOutputSurface instead of
GpuBrowserCompositorOutputSurface. This output surface allocates an
IOSurface-backed GpuMemoryBuffer in BufferQueue, which is rendered to.

When frames are swapped in ImageTranportSurfaceOverlayMac, and remote
CoreAnimation is not supported, send the backbuffer's IOSurface handle
to the browser in GpuHostMsg_AcceleratedSurfaceBuffersSwapped.

There is one snag here, where we are unable to open the IOSurface by
its handle if the GpuMemoryBuffer that was created by BufferQueue
has been destroyed. To fix this, make BufferQueue hold on to the
GpuMemoryBuffer while it may be in use.

Also, in GpuTransportFactory's CreateOverlayCandidateValidator, only
create a validator if the remote CoreAnimation API is supported, because
the non-remote-CoreAnimation path only supports drawing a single layer.

BUG=546795

Review URL: https://codereview.chromium.org/1416363002

Cr-Commit-Position: refs/heads/master@{#356301}