Remove Lanczos2 filters
This filter code is only used in tests and benchmarks. The code
savings is rather small, but it's a bit odd to have a unit test
comparing Lanczos2 vs Lanczos3 when you're using the exact same
implementation for both of them, just varying the parameter for
filter width.

R=fmalita@chromium.org

Review URL: https://codereview.chromium.org/1428613002

Cr-Commit-Position: refs/heads/master@{#356382}