Set up a FeatureList instance for unit tests.
This way, code that uses the FeatureList API that runs
as part of unit tests won't hit a DCHECK that a FeatureList
doesn't exist.

In a follow-up CL, I plan to expand this a bit to add support
for tests overriding default feature state. But would like to land
this initial version first, as it unblock a couple of early adopters
of this API.

BUG=548289

Review URL: https://codereview.chromium.org/1416503004

Cr-Commit-Position: refs/heads/master@{#356477}