Fix AppViewTest.KillGuestCommunicatingWithWrongAppView's Flakiness
Fixed a test which was flaky on Linux (and multiple other platforms/builds). The listed bugs and crbug.com/519039 are somehow associated with this test. The ones listed below are fixed but crbug.com/519039 is not tested yet.

The issue with the bug was the nested message loops associated with the test listeners and the chrome.test.sendMessage(*, true) API (send message with reply). The new test uses content::ExecuteScript to deliver messages.

BUG=533318,507940

Review URL: https://codereview.chromium.org/1403373004

Cr-Commit-Position: refs/heads/master@{#356859}