Optimize plane usage when primary is obscured by Overlay
Currently, when playing Video in fullscreen and using Overlays, we
update both Primary and Overlay planes. This means hardware needs to
scan out primary and Overlay plane. This is redundant in case buffer
bound to Overlay is not transparent and covers the complete region of
primary. In such cases, if primary can support the given format, we
can bound the buffer to primary itself and avoid updating the overlay
plane. This avoids the need to scan an extra plane with size equal
to the resolution.

BUG=370522

Review URL: https://codereview.chromium.org/1423793003

Cr-Commit-Position: refs/heads/master@{#356938}