Carry over REFERENCE_TIME for the frames using GPU Memory Buffer
For the frames using GPU memory buffer, the REFERENCE_TIME is lost, and
then WebMediaPlayerMS will not be able to apply rendering algorithm to
enhance the rendering smoothness.

In this CL, we make those frames to carry over the REFERENCE_TIME from
original frame.

BUG=545165

Review URL: https://codereview.chromium.org/1418923006

Cr-Commit-Position: refs/heads/master@{#357156}