Clean up the various ways to cancel parsing in a Document
Before this change, there are 2 ways we cancel parsing, one of which is only used in FrameLoader and appears very hacky (most notably, because it never actually touches the DocumentParser). Moreover, 2 of the 4 callsites for the current definition of Document::cancelParsing() are unnecessary, and a 3rd just needs to destroy the old parser so a new one can be created.

Remove/change those calls and change cancelParsing() to something a bit more obvious and spec compliant.

BUG=

Review URL: https://codereview.chromium.org/1415053003

Cr-Commit-Position: refs/heads/master@{#357213}