Make RenderFrameImpl::GetRenderWidget() return local root RenderWidget
In the case of out-of-process iframes, RFI::GetRenderWidget() returns
the RenderView, but the RenderWidget part of that is not used for
anything. Since the RenderView/RenderWidget in the case of a remote
main frame doesn't receive a Show message, GetRenderWidget()->is_hidden()
always returns false. This prevents media players from starting.

This patch finds the local root, and returns the RFI's |render_widget_|
if it has one (otherwise, it is the main frame and RenderView is the correct
thing to return.

BUG=546255

Review URL: https://codereview.chromium.org/1406403006

Cr-Commit-Position: refs/heads/master@{#357446}