Clamp clip layer to outer viewport size for scrollbars.
When there is a nonoverlay scrollbar, the inner viewport is larger than
the outer by the size of the scrollbar.  The outer viewport can still be
free to scroll within the content area, so we should take the minimum of
both sizes.

BUG=547668
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1414573007

Cr-Commit-Position: refs/heads/master@{#357613}