Make decoder source manager a JPEGImageReader member
A decoder source manager is required for the lifetime of the
JPEGImageReader class, and it makes little sense to allocate
(free) the source manager in the class c-tor (d-tor).

Make the decoder source manager a class member, so the class
can be created without error (removes JPEG_ERROR state, also
the allocate/free steps, of course).

BUG=439655

Review URL: https://codereview.chromium.org/1429743003

Cr-Commit-Position: refs/heads/master@{#357631}