Remove obsolete "temporary seek" track buffer logic.
Originally added nearly 3 years ago in
http://crrev.com/11691006

Updates to SourceBufferStream have made this block obsolete. If
there are GOPs between the end of the track buffer and the start
of the new buffers, current logic will leave those GOPs untouched,
so temporarily seeking to those GOPS is not needed.

The tests that were added to exercise this block still exist today
and pass irrespective of whether the block is present.

TEST=all media unittests.

Review URL: https://codereview.chromium.org/1410353010

Cr-Commit-Position: refs/heads/master@{#357719}