Temporary RELEASE_ASSERT to find frame loads during DocumentLoader detachment.
This CL mostly relands r357479 but without the check that ancestor documents
haven't started unload as that turned out to fire very frequently.

I've also moved the loadEventProgress member of Document to be one of the first
members so that it's likely to be included in a minidump memory region.

BUG=519752

Review URL: https://codereview.chromium.org/1414693010

Cr-Commit-Position: refs/heads/master@{#357751}