Devtools Animations: Update and fix flaky iframe animation test
This test was flaky because the normalised monotonic time values invariant
was broken. This was broken because each document had its own clock which
ticked independently to other documents on the same page, maintaining the
invariant acros a document, but not between pages. This is fixed in
https://codereview.chromium.org/1410313004/

BUG=479181

Review URL: https://codereview.chromium.org/1427433005

Cr-Commit-Position: refs/heads/master@{#357854}