Switch SSLClientSocketOpenSSL to the new BoringSSL renego API.
The current double-negative API was confusing. Fortunately, we needed to make
it an enum in BoringSSL for some other consumer, so that was an excuse to fix
it and deprecate the old one. Also we default to rejecting renego now, so
there's no need to explicitly configure it that way.

BUG=none

Review URL: https://codereview.chromium.org/1410163004

Cr-Commit-Position: refs/heads/master@{#357869}