Collapsing borders disallow padding on table
Our code had a bug where we would wrongly remove the padding from
the available space for table-row-group and table-row when
collapsing borders. CSS is clear that paddings are not allowed on
table with collapsing borders.

On good side of this change is that the DevTools now report the right
size for table-row-group and table-row.

BUG=529763

Review URL: https://codereview.chromium.org/1422943003

Cr-Commit-Position: refs/heads/master@{#357987}