Tweaking COMPOSITOR_GESTURE to improve touchstart latency
If there's no touch handler, we don't need to prioritize compositing tasks
(in fact it's safe to deprioritize them).

If there is a touch handler, we initially deprioritize compositor
tasks when a gesture starts.  This is because it's a
good time for the main thread to run any expensive tasks that may have
piled up. After 300ms we expect the gesture to end, with another subsequent
gesture being likely.  To make sure the subsequent touchstart is
processed quickly, we start prioritizing compositor tasks again and
block loading and timer tasks if they seem expensive and there is
a touch handler on the page.

BUG=

Review URL: https://codereview.chromium.org/1420323004

Cr-Commit-Position: refs/heads/master@{#358079}