SkDevice shouldn't be accessed outside of Skia; it will soon be deprecated. This refactoring adds a function to Chromium's skia namespace that wraps the code required to get write access to the pixel data backing a *raster* SkCanvas.
Instead of returning a SkBitmap with complicated lifetime
semantics, we use the simpler SkPixmap where possible.

The previous version of ImageDataPlatformBackend::Map() forcibly
changed the bitmap descriptor from opaque to premultiplied; this
isn't guaranteed to actually work. Instead we change the
initialization of the bitmap so that it's premultiplied instead
of opaque.

R=reed@chromium.org,bbudge@chromium.org,piman@chromium.org,msw@chromium.org
BUG=543755
TEST=gfx_unittests,ppapi_unittests

Review URL: https://codereview.chromium.org/1420283003

Cr-Commit-Position: refs/heads/master@{#358166}