ExclusiveAccessBubbleViews: Use BoxLayout instead of GridLayout.
BoxLayout is more appropriate, and does not suffer the 'invisible
children still take up space' issue of GridLayout (crbug.com/548551).
(There is a work-around for this issue with Label, but not with other
view types which we'll need in the future.)

Removed the code for Label and ButtonView to make their GetPreferredSize
return 0 when hidden (no longer necessary).

This CL should not have any observable effect.

BUG=352425

Review URL: https://codereview.chromium.org/1410413007

Cr-Commit-Position: refs/heads/master@{#358238}