Let RenderFrameImpl set navigationStart based on CommonNavigationParams
This change originally was reverted here:
https://codereview.chromium.org/1423453007/
Due to a flawed test. The test in question tries to set Blink's navigationStart timestamp based on the browsers navigation_start, but blink essentially ignores that for it's reporting to consumers, only using the TimeTicks value passed in for relative measures (i.e. all the other metrics).

Original description:
This entails setting the timestamp before we set it in blink, so DocumentLoadTiming had to be modified to allow for that. We set the value in didCreateDataSource.

didCreateDataSource was also refactored slightly so that same-page navigations don't need to call it anymore.

This is 2/3 of the effort to add a navigation_start timestamp to DidStartProvisional load. The full CL we're breaking up is here:
https://codereview.chromium.org/1425823002/

BUG=548335

Review URL: https://codereview.chromium.org/1409883008

Cr-Commit-Position: refs/heads/master@{#358532}