Add detection for repeated audio in capturing.
This CL is to add a light-weight detector for repeated audio that has been observed in some recent recording. The issue might have been resolved in a recent fix, see
https://chromium.googlesource.com/chromium/src/+/8d9071da52c70d300bfc0cdc0448c564b39764f4

It is still good to add a detector and UMA stats to verify the fix and fire alarms for occurrences of audio repetition due to other potential causes.

The repetition detector was planned to be placed in WebRTC Audio Processing Module, but per discussion, it is better to be placed in content renderer. The algorithm has been reviewed, see
https://codereview.webrtc.org/1287663002/

BUG=520425
TEST=build with custom Chromium that produces repeated audio.

Review URL: https://codereview.chromium.org/1357013006

Cr-Commit-Position: refs/heads/master@{#358555}