Change terms URL to chrome://credits.
The existing terms seen at chrome://terms have a link to
http://code.google.com/chromium/terms.html and various translations
of it. This link target no longer exists as it was deleted due to
never being updated. As a result, replace it with a link to
chrome://credits, which is at least up to date and exists.

BUG=chromium:91111

Review URL: https://codereview.chromium.org/1418453013

Cr-Commit-Position: refs/heads/master@{#359034}