Allow higher unicode characters in XMB files.
The XMB tool has a regex of invalid XML characters, which erroneously
contained all unicode characters in the supplementary planes (U+10000 to
U+10FFFF). The tool would silently replace these characters with spaces
when generating XMB files, which caused problems recently when an emoji
character was added to a grd file. The translation console supports
these characters, so GRIT should too.

The XMB tool now supports these characters, and now raises an exception
if an invalid character is used (instead of silently replacing it with a
space).

BUG=498288

Review URL: https://codereview.chromium.org/1440313003

Cr-Commit-Position: refs/heads/master@{#359621}