[GN][Chromecast] Replace //chromecast:config with feature-specific configs.
This config target is not good style - it has remained in the build since the very first attempt to port chromecast to GN. Remove it now, and replace it with feature-specific configs. In this case, playready_config has been added to all relevant targets.

BUG=516899

Review URL: https://codereview.chromium.org/1435413002

Cr-Commit-Position: refs/heads/master@{#359678}