media::VpxVideoDecoder cleanup
Couple of deep-cut cleanups:

- Inner class MemoryPool is ref-counted thread safe when it
doesn't need to, removed it, using WeakPtr for the only Bind
use case, which anyway already had a BindToCurrentLoop().
- moved MemoryPool's registration & dereg with
MemoryDumpManager to its ctor/dtor.

- reshuffled a few method to follow declaration order (sorry
that it makes review harder!).
- removed VpxVideoDecoder's |task_runner_|, which is used
mostly for thread checking (with a note in Reset()).
- merged Decode() and DecodeBuffer() following a TODO
suggestion. This makes |decode_cb_| superfluous, removed.
- added vpx error stringification (useful!).
- clarified in DCHECK()s the supported pixel formats vs codec
and other assumptions that were not so evident to me.
- using libyuv::I420 copy ISO Copy{Y,U,V}Plane(). Also,
MakeOpaqueAPlane() > libyuv::SetPlane() and
CopyAPlane() -> libyuv::CopyPlane().
- changed some if-if to early returns.

Note that VP9 could have supported Alpha channel if
looking at the code in ConfigureDecoder(), but not if
we looked at CopyVpxImageTo(). Made this evident.

BUG=554196, 554737
TEST=Tried bags of tests including this
functionality, in particular
./out/Debug/media_unittests --gtest_filter="PipelineIntegrationTest.*_VP*"
and anything relating to playing a WebM file
from Chrome (big buck bunny).

TBR=bbudge@chromium.org for the tiny change in
pepper shim, since is just removing the task runner
which is anyway used without having to cache it.

Review URL: https://codereview.chromium.org/1438233003

Cr-Commit-Position: refs/heads/master@{#359880}