Fix video frame overlay codepath for Chromecast
We stayed on m43 until recently and had no tests to verify that our
video frames were correctly generating overlay quads in the compositor.
This is a collection of fixes required to get them working again.

Some explanation of specific points:

1.  Pass-through GLImage case.
We use a native pixmap to plumb the overlay frames through to the
compositor, but handle actual overlay rendering externally (via our
hw-accelerated media playback path).  Previously, we relied on a
GLImage subclass specifically for this case, which got lost in other
refactoring: see https://codereview.chromium.org/1258713002

The new approach here is to:
* Support the pass-through case directly in GLImageEGL case, i.e. allow
  |egl_image_| == EGL_NO_IMAGE_KHR.
* GLImageEGL::BindTexImage must return false for this case, and so
  ScheduleOverlayCHROMIUM should now treat this as a no-op instead of
  a GL error.

2.  XRGB video frames
We have always used this type for our dummy video frame, because the
compositor is picky about which formats can be allowed for overlays.
Support for it was removed here:
https://codereview.chromium.org/1150863008/
and we simply restore it here.

3. Adding ClientPixmap/ClientPixmapFactory for Cast ozone
This CL: https://codereview.chromium.org/1128113011 means we need
these implementing in Cast ozone platform.

Some additional tests will be coming soon (they're dependent on other
CLs so excluded here) to hopefully reduce such breakage in future :)

BUG=internal b/25230251

Review URL: https://codereview.chromium.org/1414113010

Cr-Commit-Position: refs/heads/master@{#359989}