[css-grid] Fix alignment with grid gutters and negative free space
This is a regression caused by r358816. The problem was that the free space
was not properly computed when the grid container was narrower than the
track sizes + gutters size for a given axis. The culprit was a clamp to
0 that we were doing just to avoid confusing negative free spaces with the
special meaning of freeSpace==-1 (which meant indefinite freeSpace).

In order to fix this bug and avoid future potential issues with magic
numbers, we're removing the ==-1 check to know whether the freeSpace is
indefinite or not. Instead we pass that information as an extra argument to
the track sizing algorithm.

BUG=556155

Review URL: https://codereview.chromium.org/1446913003

Cr-Commit-Position: refs/heads/master@{#360044}