Run Oilpan GC in blink_platform_unittests.
This fixes the issue that caused this revert:
  https://codereview.chromium.org/1450953002/

Essentially, this becomes the first platform unit test to have a heap-allocated
mock, which we have to clean up after. This is similar to the logic in
RunAllTests.cpp for webkit_unit_tests and heap_unittests.

OriginAccessEntryTest is adjusted to restore the old platform so that it doesn't
break subsequent tests (or in this case, the heap's attempt to clean itself up).
Currently it nulls out the platform in many of its tests, and this causes code
in the heap that attempts to record a histogram to crash. I've copied how
PlatformTestingSupport does platform init/restore.

Review URL: https://codereview.chromium.org/1448833003

Cr-Commit-Position: refs/heads/master@{#360126}