Schedule retry for user-permitted uploads.
Previously uploads that are user-permitted but not network-permitted
get ignored when it comes to retry scheduling. This fix makes them
eligible for retry.

* Use isUploadUserPermitted just to check user preference, so that
uploads that only failed this check may be retried in the future.

* Delete crash dump if user does not permit uploads. This makes sure
that uploads returning UPLOAD_DISABLED do not get retried, ever.

Also deleted obsolete code and fixed a TAG prefix.

BUG=553488
R=acleung@chromium.org

Review URL: https://codereview.chromium.org/1451153002

Cr-Commit-Position: refs/heads/master@{#360147}