cc: Fix overlay damage tracking bug
Track the region that is covered by overlays and ensure that that is
included in the damage rect of future frames. This fixes bugs where
something that was previously an overlay becomes a non-overlay due to
reasons that didn't involve doing damage to the frame in that quad's
region (e.g, creating a slightly overlapping overlay quad).

Move the removal of overlays from the damage rect to being done in the
OverlayProcessor instead of being done in each individual strategy.

BUG=551861
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1452943002

Cr-Commit-Position: refs/heads/master@{#360513}