Use unique URLs rather than qualifiers to uniquify RenderProcesses.
This is so code in the browser has something tangible to call ConnectToApplication() on. When a RenderProcessHost creates a new renderer, it creates a URL of the style exe:chrome_renderer#, where # is the RenderProcess ID. It stores this string on the RenderProcessHost's SupportsUserData. Call content/browser/mojo/mojo_shell_client_host.h's GetMojoApplicationInstanceURL() with a particular RPH to get the URL to connect to it.

R=jam@chromium.org
http://crbug.com/558009

Review URL: https://codereview.chromium.org/1455323002

Cr-Commit-Position: refs/heads/master@{#360604}