Fix use-after-free in HistoryCounter and CacheCounter.
HistoryCounter (the synced history part) and CacheCounter use base::Callback for their callbacks. If the WebUI closed and the counters destructed during counting (a task on IO thread on for cache, a web request for history), the callback would refer nonexistent counter instance.

Use weak pointers in the callbacks.

HistoryCounter (local history part) and PasswordsCounter (since it is a PasswordStoreConsumer) rely on CancelableTaskTracker::TryCancelAll to cancel the tasks after destruction. AutofillCounter relies on AutofillWebDataService::CancelRequest. Weak pointers are thus not necessary in those cases.

BUG=558481

Review URL: https://codereview.chromium.org/1464713003

Cr-Commit-Position: refs/heads/master@{#360825}