Fix for a long standing bug in Chrome windows with the auto hide taskbar.
This bug reproduces at times with multiple monitors and occurs when our attempt to
retrieve the taskbar edge information via the ABM_GETAUTOHIDEBAR appbar message fails.
msdn recommends using the ABM_GETAUTOHIDEBAREX call with the monitor rectangle for the multiple
monitor case. However that does not work.

The following fallback does work though.
1. Query for the auto hide status of the taskbar.
2. Retrieve its position and check if the edge information returned in this call matches the
   edge we are looking for. This along with the monitor check is sufficient.

BUG=472139
TEST=Set taskbar to autohide. Connect multiple monitors to the machine and start chrome on the
primary monitor and attempt to reproduce this case. If it does not repro then move chrome to
the other monitor and attempt the test case. If that fails, disconnect the secondary monitor
exit chrome and try. One of these cases should reproduce the problem.

Review URL: https://codereview.chromium.org/1462373003

Cr-Commit-Position: refs/heads/master@{#361188}