Remove ManagePasswordsUIController::NavigateToSmartLockPage
Replace it with the equivalent ManagePasswordsUIController::NavigateToExternalPasswordManager.

Note: This CL does not remove IDS_PASSWORD_MANAGER_SMART_LOCK_PAGE yet, because it is also used in other pieces of the password manager UI. We need to double-check if it is OK to merge that with chrome::kPasswordManagerAccountDashboardURL, or whether it just should be made a separate constant. Using localisation to get the hl= URL parameter is very likely an overkill, see #0 in http://crbug.com/548259.

R=vasilii@chromium.org
BUG=548259

Review URL: https://codereview.chromium.org/1467993002

Cr-Commit-Position: refs/heads/master@{#361304}