[Sync] Adding unit tests for about sync when sync is disabled.
This change involved refactoring SyncInternalsMessageHandler to allow it to be testable, particularly in regards to how it created the base::Value object tree for the about sync page. The unit tests for SyncInternalsMessageHandler shouldn't depend on/understand any of the base::Value generation logic, so they needed a way to change how it was done. This change shouldn't change the behavior of SyncInternalsMessageHandler at all.

BUG=549048

Review URL: https://codereview.chromium.org/1442583005

Cr-Commit-Position: refs/heads/master@{#361424}