Transform context menu coordinates to root frame's coordinate space
Currently when a user causes a context menu to be created by an out-of-
process iframe, it appears in the wrong location because the coordinates
of the child frame do not match the coordinates of the overall viewport.

This patch adds a mechanism to RenderWidgetHostView to transform a point
to the coordinate space of an ancestor RenderWidgetHostView, and causes
RenderFrameHostImpl to apply the transform to the target point of a
context menu request received from the renderer.

BUG=546634

Review URL: https://codereview.chromium.org/1417353007

Cr-Commit-Position: refs/heads/master@{#361441}