jni: Pass method parameters as JavaParamRef in media/midi.
Pass all object parameters to JNI methods in JavaParamRef<> wrappers.
This matches previous changes made to do this for JNI non-method
functions.

Remove the "ObjectRef" typedef as it's not correct for all the things
using it to be the same type - some should be JavaRef and some should be
ScopedJavaLocalRef.

BUG=519562
R=yhirano@chromium.org

Review URL: https://codereview.chromium.org/1475913003

Cr-Commit-Position: refs/heads/master@{#361646}