New-age, all-the-rage infobar stacking animations.
Now, only a single infobar is shown at a time. If multiple infobars are
triggered, only the top edges of the additional infobars appear, stacked
behind the front infobar.

This is a minimal working implementation. Follow-up CLs will address the
various TODOs added in this CL and clean-up and update tests.

Summary of changes:
 - Added InfoBarContainerLayout to handle all the animations. It deals
   entirely with Views and knows nothing about the rest of the infobar
   code. It has a simple interface that allows views to be added,
   removed, or swapped.
 - Deleted the old infobar animation system, i.e. ContentWrapperView,
   AnimationHelper, and parts of InfoBarContainer.

BUG=396223

Review URL: https://codereview.chromium.org/1456283004

Cr-Commit-Position: refs/heads/master@{#361662}