Page popups should respect zoom factor of the anchor element.
If the element is in non-100% zoom level, popups associated to the element
should be in the same zoom level.

This CL affects the following popups:
 - Calendar popups for input[type=date], input[type=month], input[type=
   datetime-local], and input[type=week]
 - Suggestion popups for them and input[type=time].
 - Color suggestion popup

Note that the behavior of select popups on Windows, Linux, and CrOs should not
be changed by this CL because it already respects the zoom level.

Note: Updates of existing tests:

* popup-menu-appearance-styled.html
  Because we pass a font-size in float, not int.

* popup-menu-appearance-zoom.html
  The popup border is correctly zoomed.
  The scrollbar is unexpected and I couldn't find the reason why
  the scrollbar was shown.  However, chrome and content_shell
  doesn't show such scrollbar.  So, this is not a production
  issue.

* popup-menu-load-webfont-after-open.html
  Because we pass a font-size in float, not int.
  The popup position in -expected.png is an error of my local
  machine, I think.

BUG=555326

Review URL: https://codereview.chromium.org/1454613002

Cr-Commit-Position: refs/heads/master@{#361756}