PNaCl: Enable the Subzero translator for x86-32 without a flag.
Tweaks the logic from https://codereview.chromium.org/1005173006/ so that the Subzero translator is always used for O0 translation on x86-32, regardless of the --enable-pnacl-subzero flag.

The flag is retained so that it can be used in the future to phase in Subzero for other targets, such as x86-64, arm32, and mips32, in the same way as was done for x86-32.

BUG=532718

Review URL: https://codereview.chromium.org/1457183002

Cr-Commit-Position: refs/heads/master@{#362047}