Revert "Treat '...' as a non-parent path component outside of Windows."
This reverts commit f61e0a571a76b94eb5ce059ea25babff6a2bb3a3
(https://codereview.chromium.org/1469153006). It broke some Blink tests as
FilePath::ReferencesParent is also used in things relating to the HTML
FileSystem API.

Ideally, that API should be immune to bizarre Windows-isms and use a different
notion of FilePath from the platform FilePath (which is sensitive to bizarre
Windows-ism and others). But that'll require some work to disentangle. In the
meantime, revert the problematic change so the tests pass.

BUG=464760,563588
TBR=mark@chromium.org,jschuh@chromium.org

Review URL: https://codereview.chromium.org/1488873002

Cr-Commit-Position: refs/heads/master@{#362362}