[css-grid] Fix percentage heights for grid items in quirks mode
The patch adds a condition in
LayoutBox::skipContainingBlockForPercentHeightCalculation().
In order that we skip auto-height containing blocks for grid items.

Remove the fixed size for the grid container in
fast/css-grid-layout/percent-resolution-grid-item.html
as it was not needed.

Created a copy of that test to check quirks mode behavior.

BUG=562458
TEST=fast/css-grid-layout/quirks-mode-percent-resolution-grid-item.html

Review URL: https://codereview.chromium.org/1483613003

Cr-Commit-Position: refs/heads/master@{#362401}