OOPIF: History navigations for new child frames.
When a subframe is created during a history navigation, we now make an
OpenURL call to the browser process that looks up which
FrameNavigationEntry matches the new frame (based on Blink's unique name
approach).  If no matching entry is found, we will fall back to the
original URL for the frame.

BUG=502317
TEST=See bug for repro steps.

Review URL: https://codereview.chromium.org/1422333009

Cr-Commit-Position: refs/heads/master@{#362579}