Add bounded information to renderer crash stats.
Similar to http://crrev.com/1473753002, this CL extends the native
side UMA recording to also record renderers that crash while not
having a strong binding on android.

This will allow us to confirm our suspicion that the vast majority,
over 99%, of unbound renderer crashes result in no crash report. If
that is indeed the case then we can remove this stat in the future,
but this confirmation is necessary to understand crash sources
holistically. This information is also not present in crash reports,
which means that adding a stat like this one is the only way to add
visibility to whether a renderer crashed while bounded or not.

BUG=510327

Review URL: https://codereview.chromium.org/1469023003

Cr-Commit-Position: refs/heads/master@{#362713}