mac: Sign app_mode_loader [badly]
Re-land of 2720613d5ea1, which was reverted in 51c465c92024.

We can't sign the app_mode_loader bundle because it's altered at runtime
as new app mode loader stubs are created. We can sign just the
executable. The resulting executable won't validate in its bundle
normally, and spctl won't like it, but it can be verified with
--ignore-resources, or in isolation from its bundle.

BUG=550972,565859

Review URL: https://codereview.chromium.org/1486863003

Cr-Commit-Position: refs/heads/master@{#363208}