Measure time since last use when Chrome is launched.
This will be correlated it with MobileStartup.MainIntentAction histogram.

While correlation is not directly possible in the dashboard itself,
because both histograms are recorded per session and the logs are split
by session, it can be done through dremel/lingo.

As noted in https://codereview.chromium.org/1376383002
there are some faux measurements in some cases, but they seem to be filterable
through dremel as well and I could not find a way to do this in code.

BUG=521030

Review URL: https://codereview.chromium.org/1482943003

Cr-Commit-Position: refs/heads/master@{#363262}