arc: Add ArcServiceManager to own all future ARC services
Currently the ArcBridgeService is the only ARC related service, but we have
more services coming. To initialize those services along with the
ArcBridgeService, this CL adds an ArcServiceManager which owns, initializes
and destroys them. This way we will not have to litter
ChromeBrowserMainPartsChromeos with multiple ARC services.

BUG=chromium:558499
TEST=This is a refactoring, check that the ArcBridgeService is still operating
     as expected.

Review URL: https://codereview.chromium.org/1496193002

Cr-Commit-Position: refs/heads/master@{#363522}