Fix for beta compilation error.
On https://codereview.chromium.org/1503113002 a fix for change password forms was merged to M-48. And also it was improved a test, but in this test method GetPendingCredentials was used, that was implemented only in M-49. In this CL the test was returned to the previous state.
TBR=vasilii@chromium.org

BUG=567499

Review URL: https://codereview.chromium.org/1505203002 .

Cr-Commit-Position: refs/branch-heads/2564@{#267}
Cr-Branched-From: 1283eca15bd9f772387f75241576cde7bdec7f54-refs/heads/master@{#359700}