Reader manager can only clear it's own reference
Since the ReaderModeManager has a one-to-one relationship with
ChromeActivity, it is possible for tabs in document mode to clear
the static manager reference in DomDistillerUIUtils that it doesn't
own. This change makes sure that the manager cleaning up UI utils is
the same one that set it.

BUG=562360

Review URL: https://codereview.chromium.org/1512653002

Cr-Commit-Position: refs/heads/master@{#363827}