Prep CL so that WebRTC's RTCCertificate::Expires CL can land.
The call to expires_timestamp_ns() made it so that this CL had to revert: https://codereview.webrtc.org/1494103003/. expires_timestamp_ns() was a dummy implementation anyway just returning 0, and when the CL that was reverted can re-land it is renamed to Expires() which does return something useful. A follow-up CL will start using it as soon as we roll.

BUG=544894

Review URL: https://codereview.chromium.org/1508933005

Cr-Commit-Position: refs/heads/master@{#364027}