Prevent desktop from being panned when user is interacting with System UI.
This change uses the 'ScaledEdgeSlop' value to create a small region
inside the screen edge in which pan gestures will be ignored.  This
allows the user to interact with System UI on any edge of the screen w/o
affecting the desktop or cursor.

BUG=566174

Review URL: https://codereview.chromium.org/1512123005

Cr-Commit-Position: refs/heads/master@{#364605}