mus: Move focus/activation when the focused/active window becomes invalid.
If the focused/active window is destroyed, hidden, or reparented, then make sure
focus/activation moves to the next eligible window. To do this, update the
ServerWindowDrawnStateTracker to notify its observer *before* the change
actually takes place. This allows the FocusController to have a complete view of
the tree before the change happens, and so it can shift focus/activation to the
next window.

BUG=548422

Review URL: https://codereview.chromium.org/1516863005

Cr-Commit-Position: refs/heads/master@{#364763}