Adding initial touch mode.
This change adds the first (of two) touch input modes which is Implemented
in the SimulatedTouchInputStrategy class.  This class takes the local
touch events and sends the appropriate mouse events to the remote machine.

Enabling mode switching and the second touch input mode will be
submitted in separate CLs after this has landed.

BUG=454549

Review URL: https://codereview.chromium.org/1506313004

Cr-Commit-Position: refs/heads/master@{#364862}