Route touch-events for WebViewGuest directly to guest renderer.
This CL is the beginning of a refactor to route input events directly
to a WebViewGuest's renderer instead of taking a round-trip through
BrowserPlugin first.

Touch events were chosen as the first event type to re-route, since
guests with touch-handlers installed currently bypass our attempts to
force pinch-zoom to be handled in the embedder. Also, they require a
second gesture recognizer (owned in RenderWidgetHostViewGuest) to
create gestures based on the acked touch events.

This CL does away with the extra GestureRecognizer, and routes the
touch event acks back to the main GestureRecognizer instead, thus
simplifying the event handling logic, and providing more consistent
behaviour when touch handlers are present.

BUG=445319

Review URL: https://codereview.chromium.org/1412923009

Cr-Commit-Position: refs/heads/master@{#365034}