Serialize font shorthand as empty string if any longhand is not specified (if the font shorthand is set all the longhands will be set).
Currently we only require font-size and font-family to be set for the font shorthand to be serialized as non-empty (font-size, font-family, and all the remaining longhands that are not 'normal').

Firefox serializes font as empty string unless all font longhands are set in the stylesheet (including non standard font longhands), but this behavior is inconsistent.

BUG=398239

Review URL: https://codereview.chromium.org/1519643002

Cr-Commit-Position: refs/heads/master@{#365166}