Fix bug deleting mapping from id to accessibility objects.
On Windows we maintain a map from unique IDs to accessibility
objects, and this map wasn't being cleared properly.
This bug existed previously but was harmless until
crrev.com/1497863003 made it store pointers in the map, then
it led to a crash. This patch fixes the underlying issue so that
ids get erased from the map when a node is deleted.

BUG=568737

Review URL: https://codereview.chromium.org/1529933002

Cr-Commit-Position: refs/heads/master@{#365330}