Mac: Do not leave IOSurface ports open longer than necessary
The IOSurface IsInUse/IncrementUseCount/DecrementUseCount API can be
used to determine when it is safe for pool IOSurfaces to be reused.

When an IOSurface's is set as the content of a CALayer, its use count
is incremented. Its use count is decremented only when the IOSurface is
no longer being used by WindowServer.

Also, when IOSurfaceCreateMachPort is called, the IOSurface's use count
is incremented, an is only decremented when the port is closed by both
the providing process and by any receiving processes. This is the
mechanism by which the above CALayer use count is implemented (the
WindowServer process calls IOSurfaceIncrementUseCount before it closes
the port from which it received the IOSurface).

This is the mechanism that we should be using to track when it is safe
to recycle GpuMemoryBuffer resources in cc::. At present, we just keep
resources around for ``hopefully enough frames'' in GLRenderer.

We are currently prevented from using this mechanism because
BrowserIOSurfaceManager holds on to a Mach port for IOSurfaces to track
them. Change BrowserIOSurfaceManager to hold the actual IOSurfaceRef,
and only keep the Mach ports around until they are stuffed into a
mach_msg call.

With this in place, we can now use the IOSurfaceIsInUse signal to
determine if it safe for a GpuMemoryBuffer-backed resource to be
recycled.

While we're in the neighborhood, change the RegisterIOSurface function's
behavior to fail if the IOSurface's port is not successfully opened by
the browser process, and update and add tests accordingly.

BUG=558701

Review URL: https://codereview.chromium.org/1529683002

Cr-Commit-Position: refs/heads/master@{#365416}