Fix flaky logo spin animation restart logic
This change updates the click handler for the Chrome logo in the help
page to force a style recalc between removing the "spin" class and
adding it back again. This ensures that any running animations are
cancelled and a new animation starts.

Previously the code was relying on setTimeout() to wait for a style
recalc which was never guaranteed thus the animation would not always
restart.

BUG=568485

Committed: https://crrev.com/b1bad612c1806c51985dd1b5220b053882ab21f0
Cr-Commit-Position: refs/heads/master@{#365242}

Review URL: https://codereview.chromium.org/1525953002

Cr-Commit-Position: refs/heads/master@{#365440}