Mac: Fix extension dropdown accessory view in file dialogs.
Currently, for save as file dialogs, the chosen extension is not
appended to the saved file. Also, for open file dialogs, file choices
are not correctly restricted based on the extension group chosen in
the file dialog. This CL fixes these issues, by binding a
target(ExtensionDropdownHandler) to the extension dropdown control.
Also, a new "All files" option is added to the extension dropdown to
account for the case when file_types->include_all_files is true.

BUG=439857, 334920

Review URL: https://codereview.chromium.org/1487283002

Cr-Commit-Position: refs/heads/master@{#365457}