Implement a site engagement backend for app banner triggering.
This CL adds a flag and field trial group enabling app banners to use
the site engagement service for triggering. This allows banners to be
triggered based on broader criteria than simply navigations.

An additional UMA histogram is also added to compare the time from when
an origin was first recorded as visited by the app banner system until
the time when a banner should be shown.

This CL revealed a bug in the testing for the site engagement service,
where the service would be NULL under some testing contexts. This
issue is fixed, and a new app banner settings helper test for site
engagement is added. Existing tests are refactored to properly reset
the app banner trigger system to its default values, ensuring that tests
can be run in different orders without failures.

BUG=553858

Review URL: https://codereview.chromium.org/1413343008

Cr-Commit-Position: refs/heads/master@{#365689}