Support multiple RenderViewHosts per WebContents in Android WebView.
This CL makes sure the resources in the cc::Surfaces are returned to the
correct synchronous compositor.
In order to achieve this, each ChildFrame (a wrapper around
CompositorFrame) will be tagged with the BrowserViewRenderer-generated
ID of the frame-producing compositor. A separate cc::SurfaceFactory will
be created for a new compositor.  The SurfaceFactory that corresponds to
the old compositor will be destroyed when a new RVH becomes current and
sends a CompositorFrame tagged with the new compositor's ID.

When a resource is no longer needed, it will be inserted into a map
keyed by the ID of the compositor.

Other WebView APIs like onDraw, SetViewVisibility, etc. will be applied
to the compositor owned by the current RWHVA.

BUG=533516
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1456763002

Cr-Commit-Position: refs/heads/master@{#365726}