Adding TouchInputStrategy.
This change introduces the new TouchInputStrategy class.  This class
receives local touch input events and transforms them into remote touch
events which are transmitted to the host for injection.

This class is the counterpart to the SimulatedTouchInputStrategy, the
difference being that this class is used when the remote host supports
touch injection (vs. the Simulated...Strategy which is used for mouse
only hosts).

In order to inject remote touch events, this class keeps a queue of all
received MotionEvents which it then injects if the appropriate gesture
is detected.  Once a gesture has started, then events are injected as
they are received.

BUG=454549

Review URL: https://codereview.chromium.org/1525203002

Cr-Commit-Position: refs/heads/master@{#366001}