Show favicons for not-yet-restored webpages
Currently, we are not showing favicons for the tabs that are not yet
restored. So if you close Chrome (by swiping out from Android Overview),
and come back and open tab switcher, you'll notice that most of the
favicons are missing.

This is not intended and the reason is that favicon is fetched from
webcontents, but webcontents become accessible only after the tab is
restored.

The proposed solution is to leave the current logic as is, but to fetch
and update favicons in the background.

Also, we are updating TitleCache in multiple places, so I have refactored
bitmap factories into title cache in order to remove duplicate code.

BUG=516418

Review URL: https://codereview.chromium.org/1507223002

Cr-Commit-Position: refs/heads/master@{#366041}