Allow ServiceWorker events to have custom durations
BackgroundSync events need to be able to run for a different amount of
time than other events, such as push or fetch. This CL allows each event
request to set its own expiration time and stores them in a priority
queue based on increasing expiration time. It adds some tests to the
ServiceWorkerVersion changes and lets BackgroundSync set its own
custom time (by default 3 minutes but field trails will
be able to change it in an upcoming CL).

This CL also adds the ability for selected types of requests to not stop
the service worker if they time out. Right now this only pertains to the
sync event.

BUG=569001

Review URL: https://codereview.chromium.org/1525743002

Cr-Commit-Position: refs/heads/master@{#366105}