[Cronet] Delete CronetUploadDataStreamAdapter synchronously
Destroying CronetUploadDataStreamAdapter was posted
asynchronously to executor's thread when the request is
done. This will cause a leak if the executor fails to
execute the task. It's legal to delete the adapter on
any thread as long as it is being protected by a lock.
This CL deletes the adapter object synchronously.

BUG=564946

Review URL: https://codereview.chromium.org/1496063002

Cr-Commit-Position: refs/heads/master@{#366448}