[Download] Treat RESUMING_INTERNAL as IN_PROGRESS in DownloadItem.
The RESUMING_INTERNAL state is used to indicate that the DownloadItem
has commenced a resumption attempt. The DownloadItem will remain in this
state until the network request receives a response.

Formerly, we treated this internal state as INTERRUPTED for all entities
that are observing the download item, including the UI. However, in
retrospect, this doesn't make sense and prevents the DownloadItem from
cleanly handling cases where it needs to transition back to an
interrupted state following a resumption attempt. In the latter case,
the download would otherwise try to transition from INTERRUPTED ->
INTERRUPTED which doesn't make a whole lot of sense to an outside
observer.

In addition, keeping the download in the INTERRUPTED state prevents an
outside observer from being able to cleanly tell when a resumption
attempt has begun. Furthermore, the INTERRUPTED state exhibited while an
automatic resumption is being attempted is not actionable to all outside
observers.

To rectify these, this change makes the RESUMING_INTERNAL state be
visible to external observers as IN_PROGRESS.

BUG=7648
R=svaldez@chromium.org

Review URL: https://codereview.chromium.org/1544653002

Cr-Commit-Position: refs/heads/master@{#366542}