[Media Router UI] Fix route-details/sink-list view flashing.
If the dialog is initially in route-details view, and the route is
closed before the 3s startup timer fires, then the dialog will flash
between the route-details and sink-list view before settling on
sink-list. See bug for details.

The fix is to only allow maybeShowRouteDetailsOnOpen to be invoked
once, on startup. It is removed from |rebuildRouteMaps_| and is now
called from setInitialData. Also removed |localRouteCount_| field from
media-router-container since it's no longer needed.

BUG=571463

Review URL: https://codereview.chromium.org/1538363003

Cr-Commit-Position: refs/heads/master@{#367031}