Partial revert of 'Refactor CompositorWorkerManager'
This is a partial revert of:
https://codereview.chromium.org/1449953002

Most of that CL is inactive code, but it did change the way the
compositor thread was created. Instead of a base::Thread, I'd used a
WebThreadImplForWorkerScheduler and this seems to have had some
performance impact. While I try to figure out why, this CL should get
us back to our prior performance.

BUG=565789

Review URL: https://codereview.chromium.org/1550093002

Cr-Commit-Position: refs/heads/master@{#367051}