Add get_concurrent_links.py as the implicit dependency for all targets.
It's essentially a part of the toolchain, as it controls the link
concurrency. The file is being changed quite infrequently, but
when it does there's no currently a way to try the change before it's
submitted. The consequences of a mistake are pretty severe, like multiple
bots timing out or the build being killed due to OOM.

BUG=574119

Review URL: https://codereview.chromium.org/1559853003

Cr-Commit-Position: refs/heads/master@{#367415}