Do not prompt and/or create copies when Android logins are filled into websites.
Presently, when logins from Android apps are filled into the corresponding websites and are used for a successful login, the user is prompted to save a new credential, and ultimately a copy of the credential, keyed by the web site, gets saved to the password store.

This is a very confusing UX, and also results in duplicate logins being created.  Later, these can lead to an out-of-date password stored for the Android app if the Web password is ever updated in the future.

Instead, in this scenario, do not prompt at all, do not save any new credentials, but simply update the usage count of the existing Android credential in-place while leaving the other meta-attributes untouched.

BUG=540609

Review URL: https://codereview.chromium.org/1571813002

Cr-Commit-Position: refs/heads/master@{#368421}