Reverts Switches way to set IO buffer size on Mac to HAL based AudioObjectPropertyData APIs
This CL reverts https://codereview.chromium.org/1529773006/ and two additional CLs that are correlated (and had to be done first). These are:

https://codereview.chromium.org/1551613002/ and
https://codereview.chromium.org/1551783002/

The reason for reverting is that we currently break Web Speech on Mac since it tries to use a larger input buffer size
than what is possible to set using the HAL based APIs.

TBR=tommi
BUG=504363

Review URL: https://codereview.chromium.org/1575893002

Cr-Commit-Position: refs/heads/master@{#368575}