cc: Allow returning overlay resources upon swap completion
On Mac, the only way to reliably know that overlay resources are no
longer in use is by the function IOSurfaceIsInUse (which is checked by
the functions ResourceProvider::InUseByConsumer and
ResourceProvider::CanLockForWrite).

IOSurfaceIsInUse will start returning true as soon as a CALayer has its
contents set to an IOSurface, which happens during swap in the GPU
process (in particular, when a Mach port for the IOSurface is created
for sending to the WindowServer process), and so the earliest time that
resources may be returned to the renderer process is swap completion.

Repurpose a now-extinct flag to indicate that overlay resources are to
be returned immediately upon swap acknowledgement, and add tests to
ensure this behavior.

BUG=558701
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1523063005

Cr-Commit-Position: refs/heads/master@{#368751}