Pull the Activity context from WindowAndroid if possible
WebView uses ContentViewCore's context for displaying dialogs and
wraps it at construction time. This change makes sure that if there
is an activity that can be reached through WindowAndroid, we use
that as the context inside ColorChooser. Also make sure if there is no
activity, we
dont attempt to create a color chooser and try to show it.
Update ActivityWindowAndroid to take a Context type so
that in Android WebView, the activity context can remain wrapped.

This change is a combination of
https://codereview.chromium.org/1419843002 and
https://codereview.chromium.org/1572913003 and in general changes
WindowAndroid behavior so that a Context reference is kept rather than
an Activity reference and makes it possible to use a ContextWrapper.

BUG=550410, 570429
R=boliu@chromium.org

Review URL: https://codereview.chromium.org/1578353002 .

Cr-Commit-Position: refs/branch-heads/2564@{#545}
Cr-Branched-From: 1283eca15bd9f772387f75241576cde7bdec7f54-refs/heads/master@{#359700}