Only generate bookmark app icons when there is none available to resize.
Previously icons were generated when there was no same size or larger
icon to resize. That is, smaller icons would not be scaled up.

With this change icons should now be scaled up if available, and only
generated when there are no icons at all from the page or manifest.

TEST=Creating bookmark apps (Wrench / More tools / Add to XXX...)
should only generate an icon with a letter if there is not favicon or
manifest icon.
BUG=575045

Review URL: https://codereview.chromium.org/1577823004

Cr-Commit-Position: refs/heads/master@{#369108}