Don't dismiss picture buffers when resetting the avda codec.
Rather than dismissing the picture buffers when AVDA::ResetCodecState
is called, just tell the strategy that the output buffers are no
longer valid.  Otherwise, the strategy believes that the picture
buffer will never be used again.  This isn't true; the picture buffer
will be returned via Reuse.

The net result of all this is that the AVDA will request picture
buffers only when it encounters a format changed message, even
if it is reset.

BUG=576483

Review URL: https://codereview.chromium.org/1581803005

Cr-Commit-Position: refs/heads/master@{#369301}