Cleanup dual_mode handling in ShellIntegrationWin after Metro removal.
Follow-up to https://codereview.chromium.org/1581473002

Per MSDN documentation it appears that we have no choice but to set the
dual_mode property to false when it is already set: "Removing a property
values from a property store is not supported and could lead to unexpected
results." (though this comes right after the documentation saying removal is
one of the three use cases of IPropertyStore::SetValue()..!)
https://msdn.microsoft.com/en-us/library/windows/desktop/bb761475(v=vs.85).aspx

Can be cleaned up further as a follow-up by getting rid of the dual_mode
field on ShortcutProperties altogether but that requires more thought.

Also, adding TODO for self to remove shortcut migration from the startup
path now that we have a mean to do this in setup for per-user shortcuts
via Active Setup.

BUG=558054, 577697
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/1586143002

Cr-Commit-Position: refs/heads/master@{#369527}