compositor worker: Use a WebThread for the compositor thread.
The first attempt (crrev.com/362924) had to be reverted (crrev.com/367051)
because it caused perf regressions (crbug.com/565789). It turns out, the
regression was caused because of using the default task-runner from the
scheduler, instead of directly using the thread's task runner. To fix this,
create a slightly specialized version of WebThreadImplForWorkerScheduler
that uses the thread's task runner as the default ask runner.

BUG=565789, 430155

Review URL: https://codereview.chromium.org/1589463002

Cr-Commit-Position: refs/heads/master@{#369532}