Fix the keyboard shortcuts page in the help app not showing.
Clicking the learn more link in the keyboard overlay used to show a blank
page. The article with id 188743 has been changed to 1047364 on gKMS, and a
CL to fix that has landed here:
https://chrome-internal-review.googlesource.com/#/c/244298/
The article ID has to change on the UI side too.

BUG=556439
TEST=manually, clicking the learn more link shows the proper page of the help app.

Review URL: https://codereview.chromium.org/1585843005

Cr-Commit-Position: refs/heads/master@{#369550}