Exposing OpenH264 build targets regardless of use_openh264 flag value.
Without the flag, the OpenH264 targets are not built as part of "All". We'll still want to only build OpenH264 if a flag is turned on, but this makes it possible for other repos (i.e. WebRTC) to depend on the target based on its own flags/conditions.

For example, if WebRTC has its own build flag, its default value can be changed within the same repo, making it possible to run all trybots on a CL without having to land a default value change in Chromium first (which we can't do because the default must remain to not build it!).

This CL also adds -Wno-header-hygiene to GYP gcc/clang flags. This is not necessary for the chromium bots, but I think it's necessary for some webrtc standalone bots.

BUG=468365

Review URL: https://codereview.chromium.org/1584963004

Cr-Commit-Position: refs/heads/master@{#369711}