tools/battor_agent: Loosen the BattOrAgent construction requirements
Constructing a BattOrAgent doesn't perform any IO, so it's not critical
that it be done on the IO thread. Loosening this restriction allows
us to create the BattOrAgent without having to switch threads when
we're inside the Chrome tracing agent and on the UI thread.

TBR=nednguyen@google.com
BUG=542837

Review URL: https://codereview.chromium.org/1595563002

Cr-Commit-Position: refs/heads/master@{#369768}