Allow a Snackbar to force its display when it will be deleted due to timeout
If a snackbar has the forceDisplay property it will be shown when
SnackbarManager#dismissAllSnackbars() is called from a timeout.
If SnackbarManager#showSnackbar() is called while a forceDisplay
snackbar is showing, the new snackbar will be added to the stack
and the forceDisplay snackbar will not be overwritten.

BUG=547254

Review URL: https://codereview.chromium.org/1575603003

Cr-Commit-Position: refs/heads/master@{#369821}