Preparing web ui data source unit test for async
This CL adds callbacks for the tests to wait for
async behavior.  The current code seems ok expecting
the resource fetch to happen immediately, but I
have a future CL that will add some i18n filtering
on resource load that will happen in a
SequencedWorkerPool.  That motivated this change to
wait for the resource load.

BUG=None

Review URL: https://codereview.chromium.org/1583343003

Cr-Commit-Position: refs/heads/master@{#369858}