Deflake Android fullscreen tests.
The flaky tests all dealt with manually hiding the top controls
based on user gesture.  The flakiness comes from the fact that
determining whether the top controls can be scrolled is controlled
in the renderer compositor and even if we were to monitor Tab.java's
latest value that it has sent, it does not take affect until some
time later.

As a result, this adds a new function that attempts to scroll the
top controls in a loop until it takes affect and then proceeds.

BUG=488393,339668,512299

Review URL: https://codereview.chromium.org/1598433004

Cr-Commit-Position: refs/heads/master@{#369880}