O(1) descendant slots by keeping track of insertion/removal of slots in a shadow tree.
In slots assigning, we have to traverse a shadow tree to collect all slot elements in the tree.
This may take O(N), where N is the number of a node in the tree.

To make it O(1) amortized, this CL keeps track of insertion or removal of slot elements
and re-use the previous result if nothing has changed.

According to the micro benchmark, v1-distribution.html, which is included in this CL,
the distribution calculation became 20% faster, in case of N=100.

With this CL:    avg 6680.41 runs/s
Without this CL: avg 5486.71 runs/s

BUG=531990

Review URL: https://codereview.chromium.org/1589903002

Cr-Commit-Position: refs/heads/master@{#369992}