[Docs] Move heap profiler docs into repository
This moves the page that was at [1] to in-tree Markdown documentation
and updates the information on there too.

This is the second part of an incremental effort to move all MemoryInfra
documentation (and eventually all tracing documentation) off of Google
Sites and into the repository.

[1]: https://sites.google.com/a/chromium.org/dev/developers/how-tos/trace-event-profiling-tool/memory/heap-profiling-with-memory-infra

TBR=oysteine@chromium.org

Review URL: https://codereview.chromium.org/1598393004

Cr-Commit-Position: refs/heads/master@{#370105}