Change default command buffer result for BrowserGpuChannelHostFactory.
When a rapid start/stop of the gpu process occurs for Chromecast,
the GpuChannelHost in BrowserGpuChannelHostFactory doesn't get flagged
as lost even after the gpu process and its host are dead.

The default command buffer result is updated to
CREATE_COMMAND_BUFFER_FAILED_AND_CHANNEL_LOST so that when creation
fails due to a null GpuProcessHost, a leftover gpu channel that was
connected to the previous gpu process is flagged as lost.

BUG=internal b/26542692
TEST=Rapid sequence of start-ups/tear-downs for gpu process.

Review URL: https://codereview.chromium.org/1607333002

Cr-Commit-Position: refs/heads/master@{#370463}