Cut down on ancestry announcements
Currently, when crossing out of a container (or anything with a leave output rule), we might get into a situation with the following output:
exited article, ..., entered article, ...

In Classic, we explicitly checked for this case. We need to do the same in next.

TEST=go to a site with a bunch of landmarks. Observe we don't get these announcements.

Review URL: https://codereview.chromium.org/1588033003

Cr-Commit-Position: refs/heads/master@{#370594}