Experimental CompressibleString UMA
This CL replaces String/AtomicString objects for JavaScript source
strings with CompressibleString. CompressibleString introduced by
this CL is a mock and just records UMA. CompressibleString has
compress/decompress methods but does not actual
compression/decompression. We're trying to experiment with this CL
to judge if we can introduce actual compression algorithm.

The new UMA Memory.CompressibleStringCount records the frequency
with which JavaScirpt source strings are compressed and decompressed
in foreground and background tabs. If decompressing in background
doesn't occur often compared to compressing in background, we would
be able to say decompressing would be effective to reduce memory.

BUG=574317

Review URL: https://codereview.chromium.org/1583263002

Cr-Commit-Position: refs/heads/master@{#370693}