Replace TestingProfile::GetCookieMonster with GetCookieStore.
I'm trying to remove CookieStore::GetCookieMonster, since it's
weird and ugly to have most consumers of a class have hard
dependencies on it being a particular subclass of that class.

Also make a test using the method no longer hold onto references
to the CookieStore, in preparation for de-reference-counting the
class.

BUG=579653

Review URL: https://codereview.chromium.org/1613473002

Cr-Commit-Position: refs/heads/master@{#370779}