[Android WebView] Implement initial settings and callback support for Service Workers.
Service Workers are not tied to AwContents as regular webviews are and
therefore do not invoke callbacks provided in WebViewClient and
WebChromeClient.

This patch provides initial implementation of AwServiceWorkerClient,
AwServiceWorkerSettings and related infrastructure. In particular it
implements the shouldInterceptRequest callback mechanism for service
worker related requests. The callback is invoked when a service worker
script is registered and also for network fetches from within the
service worker script.

This patch also contains a fix for TestWebServer to make it work with
service workers.

BUG=566027

Review URL: https://codereview.chromium.org/1544863002

Cr-Commit-Position: refs/heads/master@{#370998}