Service Worker: Change the criteria for bumping the last update check time
As per the latest spec change, when an update successfully loaded a script
resource from the network (having bypassed the HTTP cache), the registration's
last update check time should be updated regardless of whether the installation
of the fetched resource succeeds or fails.

While the existing code updates the time stamp only when the subsequent
installation of the newly received script is successfully completed (or the
loaded script is byte-for-byte match to the incumbent script resource), this CL
makes it bump the time stamp as soon as the update succeeded.

This CL added a condition for checking whether the script resource was served
from the network:
if (net_request_->response_info().network_accessed &&
    !(net_request_->response_info().was_cached)) // This line is added.

Spec discussion: https://github.com/slightlyoff/ServiceWorker/issues/514#issuecomment-135641718

BUG=539709

Review URL: https://codereview.chromium.org/1381153004

Cr-Commit-Position: refs/heads/master@{#371485}