Remove the code that prevents task blocking if not expecting frames
In abstract preventing task blockimg if not expecting frames sounds like
a good idea.  In practice it caused a large perf regression since we
probably should be blocking tasks that take longer than 50ms even if we're
not expecting frames because touch starts can and will get delayed.

Removing this condition is safe because:
a. We are now using task blocking in more limited cases
b. For the usual compositor scrolling case we increased the blocking
threshold to 50ms

BUG=574343

Review URL: https://codereview.chromium.org/1636513003

Cr-Commit-Position: refs/heads/master@{#371521}