Avoid dereferencing find() without checking for == end()
If (due to an unexpected failure) one of the find() calls in
audio_video_metadata_extractor_unittest.cc fails (returns end()) then
dereferencing it triggers undefined behavior. This makes the results of
the test undefined. In most cases the test failure will still be caught
but there are no guarantees.

This change adds a helper function to simplify using find() in a way
that is safe even in the face of unexpected failures.

This was discovered while investigating crbug.com/581768.

R=dalecurtis@chromium.org

Review URL: https://codereview.chromium.org/1643633005

Cr-Commit-Position: refs/heads/master@{#372219}