blink_gc_plugin: Make RecordInfo::Get{Fields,Bases} return deterministic order
Tests are failing on the Windows bots because diagnostics get emitted in
non-deterministic order. This is a follow-up to
https://codereview.chromium.org/1645613003/ which fixed one instance, but
there were several more.

This should fix it for real by storing bases in a std::vector and using a
SourceLocation-based comparator for the fields map.

BUG=581782
R=sigbjornf@opera.com, thakis@chromium.org

Review URL: https://codereview.chromium.org/1645763004 .

Cr-Commit-Position: refs/heads/master@{#372387}