Merging to M48
Ensure that the code in the DXVA DX11 decoder which can potentially access the DX11 device context on the decoder thread is locked with the ID3D10Multithread interface.

The ProcessInput call on the color converter does access the DX11 device context on the decoder thread to check states.
This currently executes outside the lock. Moving it inside the lock. With this change the color converter ProcessInput
and ProcessOutput calls are in the lock. Added a class AutoDX11DeviceLock to make the locking more seamless.

The other change is to get rid of the begin streaming calls for the converter and the Flush calls on the context.
These are not necessary now with the single device

If this does not fix the Win8 bot flakiness the last step to try would be to move the copy texture call to the main thread.
Failing that we need to go back to the two device model. All symptoms here point to an NVIDIA driver quirk.

BUG=548383
TBR=dalecurtis

Review URL: https://codereview.chromium.org/1645333003 .

Cr-Commit-Position: refs/branch-heads/2564@{#641}
Cr-Branched-From: 1283eca15bd9f772387f75241576cde7bdec7f54-refs/heads/master@{#359700}