Update chrome for new prefs location.
This is a search-and-replace update for includes with "base/prefs" ->
"components/prefs" and the headers re-sorted in the chrome directory.

DEPS files were updated to allow the components in question to depend on
components/prefs. chrome/DEPS was updated to whitelist all components rather
than listing all the ones it uses out individually. Checkdeps is a tool to
enforce layering, and Chrome is allowed to depend on any component, so listing
each one it uses doesn't help anything.

This should be a no-op from a build perspective.

BUG=583034
TBR=jam@chromium.org

Review URL: https://codereview.chromium.org/1658793002

Cr-Commit-Position: refs/heads/master@{#372827}