net: add BindAndListen() method to UnixDomainServerSocket
This method was designed to support the Unix Domain Socket case, where the
address + port pair does not make sense, as it uses files for sockets
instead. This makes the callsites cleaner.

With this addition there is no need to override ListenWithAddressAndPort
from ServerSocket.

BUG=None
TEST=net_unittests
R=mmenke@chromium.org,mkwst@chromium.org
TBR=gunsch@chromium.org,boliu@chromium.org,sergeyu@chromium.org,newt@chromium.org

Review URL: https://codereview.chromium.org/1644003002

Cr-Commit-Position: refs/heads/master@{#372859}