cc: Add support for multiple observers to BeginFrameSource.
Multiple observers can be added to BeginFrameSource. The
SetNeedsBeginFrames method was removed because it doesn't make sense
with multiple observers. BeginFrameSources are expected to produce
BeginFrames as long as there are registered observers. There are helper
methods in BeginFrameSourceBase to make this easy.

Changes had to be made to DelayBasedTimeSource as its MISSED BeginFrame
logic was moved to SyntheticBFS for dealing with multiple observers.

BUG=580352
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1611633004

Cr-Commit-Position: refs/heads/master@{#373031}