[Sync] Request dependency services in sync client with implicit access type.
Currently all services that sync client depends on are requested with
ServiceAccessType::EXPLICIT_ACCESS. Explicit access should be used for
operations that are caused by user input.

The typical behavior of service factories is when service is requested
with implicit access in incognito mode service factory returns nullptr.

Sync should not be syncing in incognito mode so this behavior should be
acceptable.

R=skym@chromium.org
BUG=558320

Review URL: https://codereview.chromium.org/1647833006

Cr-Commit-Position: refs/heads/master@{#373041}