Implement in-proc VerifySyncTokens and use for android video
In Android WebView, video still uses InProcessCommandBuffer.
So the video sync tokens cannot be verified by the compositor
using CommandBufferProxyImpl. This breaks verification of
all sync tokens, and causes corruption when video is present.

Fix by implementing verifying the token immediately with
the InProcessCommandBuffer. And this should be ok as the
parent compositor also uses InProcessCommandBuffer on the
same thread as video InProcessCommandBuffer.

Review URL: https://codereview.chromium.org/1663503002

Cr-Commit-Position: refs/heads/master@{#373122}