[css-grid] Tracks with flex factor sum below 1 allow stretch alignment
We had an assertion in the content alignment logic for stretching tracks
to ensure there are not flex tracks when there is available space. This
assertion was set under the assumption that any flexible track would
exhaust the available space, required to apply the content alignment
logic.

However, if the sum of all the flex factors is below 1, the flexible
traks will be sized so there is still available space for applying the
content alignment css rules. This assertion was valid until r201486.

BUG=583344

Review URL: https://codereview.chromium.org/1667673002

Cr-Commit-Position: refs/heads/master@{#373225}