Ignore in-page navigations and pre-rendered pages for site engagement.
In-page navigations should not trigger or reset engagement detection.
Pre-rendered pages currently trigger navigation engagement, even if they
are ultimated discarded. It is nontrivial to observe whether
a pre-rendered page is ultimately swapped-in (and should receive
navigation engagement), as the transitions assigned to these pages can
be misleading.

This CL disables engagement triggers in DidNavigateMainFrame for in-page
and pre-rendered navigations. Input detection will be activated in
WasShown(), which is called when pre-renders are swapped in.

BUG=565087

Review URL: https://codereview.chromium.org/1660133003

Cr-Commit-Position: refs/heads/master@{#373434}