Re-land: Minor fixes for WidgetRemovalsObserver::OnWillRemoveView
Previously it was called whenever a view was removed, even if it
was just moved within the same widget.

Now it's only called if it's actually removed entirely, or moved
to a different widget.

This patch fixes a memory leak in the previous changelist:
http://crrev.com/1624343003

BUG=581127

Review URL: https://codereview.chromium.org/1662083002

Cr-Commit-Position: refs/heads/master@{#373578}