[sql] Remove misleading AutoRecoverTable() parameter.
|extend_columns| was intended to be used in the case where the target
schema differed slightly from the schema of the table being recovered.
Unfortunately, it actually implemented part of the solution for when the
target has fewer columns and described it as the solution for when the
target has more columns.  Remove the unnecessary code and parameter.

[In general, table schema only add new columns, removing is more
infrequent.]

BUG=none

Review URL: https://codereview.chromium.org/1666473003

Cr-Commit-Position: refs/heads/master@{#373586}