Attempt to fix a crash in the H.264 decoder on Windows in the GPU process.
This crash occurs in the DXVAVideoDecodeAccelerator::Decode functions in a number of places.
1. While allocating an IMFSample interface for the packet coming in.
2. While copying the encoded data into the the sample.
3. While posting a task to the decoder thread.

It is unclear as to why these crashes occur. We have not been able to reproduce these in house.

The callstack in the dump for 1 above, indicates that we are trying to allocate a sample for the default
stream size of 0x1000, while the incoming packet is much lesser than that. I don't think that this crash
is due to OOM. However it seems better to allocate what is needed.

The other change is to set the current length on the media buffer before unlocking it as per msdn guidelines.

BUG=455390

Review URL: https://codereview.chromium.org/1666043002

Cr-Commit-Position: refs/heads/master@{#373684}