Fix handling of 1.0 threshold.
Prior to this patch, a threshold of zero was handled correctly, i.e., it meant "any non-zero number of pixels", but a threshold of 1.0 was not handled correctly: "threshold < ratio" would always be false for threshold==1.0, so the threshold would never be crossed.

This patch switches to a "<=" comparator, which more accurately implements the method as named (firstThresholdGreaterThan), and fixes the ratio=1.0 case.  Making that change necessitates adding some special handling for the zero ratio case, since "ratio <= threshold" will always be true, even for a zero threshold.  The zero threshold will be crossed when the ratio transitions from zero to any non-zero value.

BUG=583456
R=ojan@chromium.org,esprehn@chromium.org

Review URL: https://codereview.chromium.org/1663583004

Cr-Commit-Position: refs/heads/master@{#373792}