Require OS X 10.7 for Chromium's clang binaries.
This updates clang from r259395 to r259396. The latter
revision is an lldb change which we don't bundle, so this
is a no-op on Windows and Linux.

On OS X, remove the hoops we used to jump through to be able
to run clang on OS X 10.6.  As far as I can tell, this
isn't needed anywhere anymore (I checked Chromium, NaCl,
v8, and pdfium bots).

If this does break your Mac builder bot, please let me
know!

BUG=none

Review URL: https://codereview.chromium.org/1663003004

Cr-Commit-Position: refs/heads/master@{#373900}