Don't focus plugin when created in a sub-frame.
At present, when a plugin is created we give its element focus. This can
cause problems when the plugin is created in a subframe, as it
may unexpectedly take focus from the other frame(s). This CL restricts
by only giving focus to the plugin when it is loaded in the main frame.

BUG=574648

Review URL: https://codereview.chromium.org/1673333002

Cr-Commit-Position: refs/heads/master@{#374183}