Redefined the bit WebTouchEvent.causesScrollingIfUncanceled.
The "causesScrollingIfUncanceled" name is confusing because this bit
essentially means that a touchmove has exceeded the touch slop region.
A touchmove with this bit set doesn't necessarily cause scrolling,
even when uncanceled, e.g. when the touch sequence started on a node
with "touch-action:none".

This CL redefines the bit: this bit remains unset as long as the
touch-point is within the slop region, and is set after the first
touchmove beyond the slop region.

The bit is set by gesture-provider, and currently used in two places:
for touchmove slop suppression and for suppressing PointerEvent firing
for events that cause panning/zooming. This new definition was chosen
to be useful for both the current slop suppression and to enable us to
detect tap-like touchend events from blink (eg. for crbug.com/582140).
We are fixing the PointerEvent use case through crbug.com/567740.

BUG=582140,567740

Review URL: https://codereview.chromium.org/1645613007

Cr-Commit-Position: refs/heads/master@{#374204}