[Android] Use fallback icon if there is no large favicon on the history page
This CL makes the history page use a rounded rectangle with the first letter of
the URL for history entries for which there is no large enough favicon.
Previously, the history page did a really bad resize for entries for which there
is no large enough favicon.

This CL:
- Introduces an Android version of FallbackIconService:
FallbackIconServiceImplAndroid which uses RoundedIconGenerator
- Encodes the SkBitmap produced by FallbackIconService to PNG on a worker
  thread. Encoding to PNG takes on average 5ms on a Nexus 5.

BUG=566012

Review URL: https://codereview.chromium.org/1596273003

Cr-Commit-Position: refs/heads/master@{#374247}