[mojo-edk] Reland: Simplify multiprocess pipe bootstrap
This is a reland of
https://crrev.com/b3ea203171e07f5c7e476e94d210ec4ad53ce5b0

which was reverted for breaking browser tests in debug builds.
The problem was essentially that faster init -> faster shutdown
-> exposed bad racy behavior in ChannelInit.

This reland rmeoves the bad racy behavior from ChannelInit.

This introduces a new MergePort message at the Ports layer
for joining two independent port cycles which each have
an unused (i.e. unwritten, unread, unsent) receiving port.

MergePort allows us to create a MessagePipeDispatcher which
is immediately usable but which will eventually be linked to
a MessagePipeDispatcher on another port cycle, potentially in
another process.

The basic idea is to create a fully functional port pair but
only bind one port to an MPD. Do this on each end and
merge the dangling ports asynchronously.

The simplification here allows a lot of code to be deleted
from NodeController, some of which is deleted in this CL.

Future work will convert existing bootstrap sites back to
using synchronous bootstrap, including the token-based APIs.

BUG=584764
TBR=amistry@chromium.org
TBR=ben@chromium.org

Review URL: https://codereview.chromium.org/1682043002

Cr-Commit-Position: refs/heads/master@{#374441}