Fix pointer arithmetic bug in SdkMediaCodecBridge::CopyFromOutputBuffer()
Currently, SdkMediaCodecBridge::CopyFromOutputBuffer() adds an offset twice to a base pointer, leading to an incorrect result. This bug was introduced in https://crrev.com/1472943002 and this CL fixes that problem.

Also, the interface of MediaCodecBridge::CopyFromOutputBuffer() is simplified - instead of returning a bool to signal if the source capacity is big enough, it now simply CHECKs this condition. No one is checking the return value of CopyFromOutputBuffer() anyway.

BUG=445320

Review URL: https://codereview.chromium.org/1686043002

Cr-Commit-Position: refs/heads/master@{#374874}