Add knowledge about blocking and passive event listeners.
Originally we though we could just get away with assuming blocking
represented blocking and passive. But it is the case when there
are touch event listeners that aren't hit we must fire touch events
to the main thread which aren't really necessarily.

Change this to be an enumeration of kBlockingAndPassive.

BUG=489802
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1686833002

Cr-Commit-Position: refs/heads/master@{#375054}