Extracts InterfaceRegistry from ConnectionImpl.
For most existing Mojo uses outside of Chrome, we have been relying on Connection::AddService<T>() as the means of exposing interfaces to other applications, because primarily we have been doing this via the InterfaceProviders exchanged @ Shell::ConnectToApplication.

However there have been times in the past and are likely to be again in the future where we will want to pass an InterfaceProvider to another method, so having a canonical implementation of an InterfaceRegistry seems useful.

I pulled all the relevant bits out of ConnectionImpl into this utility and then made ConnectionImpl wrap it. I suspect in time we could replace some of the custom InterfaceProvider impls currently in Content with this.

R=rockot@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1681933005

Cr-Commit-Position: refs/heads/master@{#375146}