Fix windows 10 check in VideoCaptureDeviceTest.
The test VideoCaptureDeviceTest.CaptureMjpeg is disabled on Windows 10 (http://crbug.com/570604). But the detection of Windows 10 is currently broken, most likely due to the introduction of VERSION_WIN10_TH2 as a distinct version. This cl updates the version check, to see if that solves the problem.

BUG=586556

Review URL: https://codereview.chromium.org/1694743002

Cr-Commit-Position: refs/heads/master@{#375422}