Track number of instances with a corrupt allocator.
The existing indication of the number of allocations that
failed due to a corrupt allocator doesn't provide much of
use because it's impossible to tell across how many actual
instances that is spread.  This new code will be set only
once per instance with corruption providing an important
data point.

BUG=546019

Review URL: https://codereview.chromium.org/1693463003

Cr-Commit-Position: refs/heads/master@{#375625}