Fix the crash of bookmarks deletion in the CBD dialog on Android.
This crash was introduced in CL https://codereview.chromium.org/1530123002/. When the user selects bookmarks for deletion in the Android CBD dialog, they are deleted Java-side. However, bookmarks are still passed in the list of datatypes to pref_service_bridge. During the construction of BrowsingDataRemover::RemoveDataMask, this hits NOTREACHED.

This should not happen - either the Java caller should remove bookmarks from the list of data types, or the C++ side should no-op when the deletion of bookmarks is requested, as there is no mask for them. This CL implements the latter solution.

We also add a test checking that this time none of the checkboxes in the dialog causes a crash.

BUG=569870

Review URL: https://codereview.chromium.org/1697123002

Cr-Commit-Position: refs/heads/master@{#375850}