Change the logic to find the best matching Service Worker to debug
The current logic uses the length of the service worker script URL's directory to find the best matching service worker to debug.
But the directory may not be same as the service worker's scope and this logic may choose wrong service workers.

This CL change this logic to use the actual scope of the service worker, and check whether the page URL matches to the scope.

BUG=584296

Review URL: https://codereview.chromium.org/1713933002

Cr-Commit-Position: refs/heads/master@{#376958}