Dump DisplayListRasterSource memory
This change implements MemoryDumpProvider on DisplayListRasterSource.
This dumps the estimated memory usage of the display list based on
the existing GetPictureMemoryUsage call.

Note that for thread safety, MemoryDumpProviders must be registered/
unregistered on the same thread they dump memory on. This is currently
done on the compositor main thread. This CL adds a ThreadChecker to
ensure this.

BUG=567465
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1725553002

Cr-Commit-Position: refs/heads/master@{#377109}