Use the MonitorFromPoint API for the cursor position to cover cases where the MonitorFromWindow API fails.
In some cases the MonitorFromWindow API for the taskbar fails to return the correct monitor. Examples include the autohide
taskbar on the left of the secondary monitor or the right of the primary monitor. This causes maximized Chrome windows to
draw over these taskbars.

Fix is to use the MonitorFromPoint API for the cursor position which seems to work for these cases.

BUG=472139

Review URL: https://codereview.chromium.org/1735743002

Cr-Commit-Position: refs/heads/master@{#377661}