Update test configs for chromium.webkit bots.
It looks like a fair amount of bitrot had set in and the entries in
this file were not up to date. The recipe will not error out if
entries are missing, and so we were getting the default sets of tests.
This CL brings things up to date.

It looks like swarming may not currently work on Mac10.10 or 10.11,
so I'm leaving it off on these bots for the moment; I will investigate
that separately. This bot also updates manage.py to blacklist (ignore)
the Mac bots.

TBR=thakis@chromium.org,maruel@chromium.org,phajdan.jr@chromium.org

Review URL: https://codereview.chromium.org/1739163003 .

Cr-Commit-Position: refs/heads/master@{#377797}