Fix inspector crash on usage of invalid variables
This patch fixes a crash in the inspector which occurs when invalid
variables are set on an element. This is either due to the variable
being used in a cycle or referencing another invalid (possibly unset)
variable without a fallback.

The bug was introduced in patch crrev.com/988acae8. This patch allowed
the HashMap on StyleVariableData to contain entries with null values,
for invalid values, instead of using a CSSVariableData with an empty
token vector. The inspector is the only other piece of code which
directly looks at the map.

BUG=589396

Review URL: https://codereview.chromium.org/1739723003

Cr-Commit-Position: refs/heads/master@{#377816}