Protect the provisional loader from detaching during prepareForCommit
This patch fixes a crash caused by a frame losing both its DocumentLoaders.
When the main document loader is detached, various subresources can hook
into 'abort' and try to stop the provisional load that is about to commit.
This causes both document loaders to be detached, and the frame to be in a
non-detached limbo state.

BUG=394055

Review URL: https://codereview.chromium.org/1714063002

Cr-Commit-Position: refs/heads/master@{#377900}