Introduce LayerListIterator
This iterator (unlike the other iterators which are intended to operate
on the RSLL) will traverse a collection of LayerImpls. It is intended
to be used whenever we currently do a walk over the LayerImpl tree.

It currently operates on a tree of layers, but if we migrate all code
to work in terms of this iterator, when we change the underlying
iterator implementation to work on lists, all code will be updated in
unison.

BUG=557194
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1758823003

Cr-Commit-Position: refs/heads/master@{#378827}