Fix overlay scroll bar color on elements with dark background
When overlay scroll bars are used on elements with CSS `overflow` set to use
scroll bars, and the background of that element is dark, the color of the
scroll bar is now set to the "light" style. This was previously only done for
the document (body), not for elements with a scrollable area. Whenever the
background changes, the scroll bar style will be updated accordingly.

Even after this change, the scroll bar is rendered with the wrong color if the
scrollable element doesn't have a background of its own (but a parent element
uses a dark background). Fixing the issue for scrollable elements with
background set explicitly is an improvement to the current state though.

BUG=588709

R=wangxianzhu@chromium.org

Review URL: https://codereview.chromium.org/1738503003

Cr-Commit-Position: refs/heads/master@{#378843}