[Android] Add SetBackgroundColor to avoid white screen on startup
If set the WindowBackground of Activity to RED, and webpage page's backgroudn to RED,
when application starts or resumes after click back key, it will showup a white screen
shortly. With this SetBackgroundColor, we can change the first screen to a fix color,
which can help to avoid white screen.

BUG=XWALK-4809, XWALK-4995