Suppress lint warning.
This patch is to suppress the lint warnings for missing density folder
and unused resources.
The exactly way is to add the ignore path to specified list, e.g.
  <issue id="UnusedResources">
    <ignore path="xwalk/app/android/app_hello_world/res/values/strings.xml" />
  </issue>
but this way does not work all the time, I have tracked the command line,
python code and config xml, there is no difference.

Ignore all the warnings for missing density folder and unused resources
as a temporary solution.

BUG=XWALK-6276