ARC: Remove error status setting in GetNetworks.
This change removes the setting of the status code in GetNetworks. Its redundant, instead in error scenarios an empty list is given back. This change also updated GetNetworks to take in an enum instead of a boolean flag. The old version is deprecated moving forward but supported for existing older versions.

BUG=590963
BUG=b/26496701

Review URL: https://codereview.chromium.org/1751793002

Cr-Commit-Position: refs/heads/master@{#379117}