docs: fix the numbering in "Converting an existing Mac checkout" section
Due to the code block in the second item, the Markdown engine is not
smart enough to figure out that the next one is related to the other
items and thus should be '3' not '1'.

To fix this, we number the list ourselves rather than letting the Markdown
engine do this work for us.

BUG=None
TEST=load the document in http://dillinger.io/ before this change,
confirm it has the issue, then change it and confirm it has the right
numbering for the list.
R=sdefresne@chromium.org

Review URL: https://codereview.chromium.org/1775473002

Cr-Commit-Position: refs/heads/master@{#379590}