Quit BlinkTestRunner::TestFinished early if main frame is not local.
content::BlinkTestRunner::TestFinished assumes that main frame is local.
This assumption only accidentally holds true today when running layout
tests with --site-per-process flag - it holds true with OOPIFs, because
today OOPIFs do not receive ShellViewMsg_SetTestConfiguration message
which means that BlinkTestRunner::is_main_window_ is set to false in
OOPIFs (even if they actually *are* in the main window).

This assumption is needed for
test_runner::TestRunner::CheckResponseMimeType and for the call to
test_runner::DumpLayout from BlinkTestRunner::CaptureDump (which targets
the main frame only).

This CL makes sure that the assumption will hold true even if an
equivalent of ShellViewMsg_SetTestConfiguration message is sent to
OOPIFs in the future (i.e. after crrev.com/1750063002).  This is done,
by having OOPIFs ask the browser to forward TestFinished to the
BlinkTestRunner where main frame is local (similarily to how this was
already asked from renderers associated with secondary test windows).
Because of this change, ...SecondaryWindow... is replaced with
...SecondaryRenderer... where applicable.

Note - this CL prevents a crash in a future CL
(https://crrev.com/1750063002), by enforcing the assumption described
above (and exiting early if the assumption doesn't hold).  OTOH, this CL
does NOT actually cause ShellViewHostMsg_TestFinishedInSecondaryRenderer
to be delivered to the browser in all the cases - there are extra issues
that prevent that from happening (these issues are addressed in a later
CL proposed at https://crrev.com/1765623003/).  Not delivering the
message is harmless and actually prevents test failures at this
point (when runtime test flags are not yet replicated - until we land
https://crrev.com/1715573002/).

BUG=587175

Review URL: https://codereview.chromium.org/1746393002

Cr-Commit-Position: refs/heads/master@{#379870}