Fix shift warnings from VC++ 2015 Update 2
Update 2 gives warnings when you shift an int and then assign to a
64-bit value. This behavior can lead to unintended truncation if you
shift too far. These warnings existed in VC++ 2013 as well, but VC++
2015 Update 2 issues the warnings more aggressively.

Sample code that is warned about:

  int64_t val = 1 << retry_num;

Sample fix:

  int64_t val = static_cast<int64_t>(1) << retry_num;

or (if retry_num is guaranteed to be small enough):

  int val = 1 << retry_num;

We could suppress the warnings, but I think they are good-ish warnings
so I'd rather not do that.

TBR=jam@chromium.org
BUG=440500

Review URL: https://codereview.chromium.org/1774003002

Cr-Commit-Position: refs/heads/master@{#379984}