One accessibility tree per frame.
Historically WebKit had a single accessibility tree that spanned all
frames. That simplified a lot of things like focus tracking.

The introduction of local and remote frames to support out-of-process
iframes and <webview> have added back a lot of that complexity. Having
accessibility trees span some frames, but not necessarily all of them,
means there were a lot of special cases to test.

To simplify things, this patch makes it so that we now have exactly one
accessibility tree per frame.

On the Blink side that means one AXObjectCache per document, rather
than trying to attach it to the topDocument. The only exception is a
page popup, we'll continue to use the accessibility tree of the
pagePopupOwner since it will always be local.

On the Chromium side we'll have one BrowserAccessibilityManager per
frame. The focused frame is now retrieved directly from the FrameTree,
which solves several tricky issues that were previously race conditions
when moving focus between a frame and a cross-process frame.

Depends on:
https://codereview.chromium.org/1762143002/

BUG=532249
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_site_isolation

Review URL: https://codereview.chromium.org/1761633002

Cr-Commit-Position: refs/heads/master@{#380416}