Move multipart resource handling to core/fetch (1/2)
Currently a multipart/x-mixed-replace response is parsed in the
content layer and dispatched to clients in blink. It is problematic
because the parser calls callbacks in a strange way and special
handling code scatters from core/html to content/child.

This change adds MultipartImageResourceParser in core/fetch. It
is basically copied from multipart_response_delegate.

BUG=570608

Review URL: https://codereview.chromium.org/1693183002

Cr-Commit-Position: refs/heads/master@{#380431}