Use chrome manifest with embedded shell
Moves chrome_manifest.json out of chrome/app/mash
and into content/browser. Changes Shell to use the
resolver when connecting to exe:chrome even when
embedded in the browser.

This brings embedded and external shell behavior
closer together as the manifest will now be used
for the browser app both in and out of mash.

As such, it fixes the fact that the browser cannot
currently connect to exe:chrome_renderer due to
recent changes around capability classes and client
process connections.

BUG=None

Review URL: https://codereview.chromium.org/1787083003

Cr-Commit-Position: refs/heads/master@{#380913}