Remove fixed tests from TestExpectation file
These tests were failing because of a state that was
not reset in the content_shell and it was being carried
over from other tests and resulting these test to fail.
The blocking bug is now fixed and that state is being
reset so these tests are always passing now.

Marking the rest of the tests for another bug as they seem to have another reason for failing.

BUG=591821, 594672

Review URL: https://codereview.chromium.org/1798963003

Cr-Commit-Position: refs/heads/master@{#381211}