Update CertStore in UpdateRequestForTransfer()
Suppose an HTTPS cross-site navigation request is being transferred from
renderer process A to process B. The ResourceLoader::OnResponseStarted()
event stores the certificate for the request in the CertStore,
associating it with process A.

If process A is still alive when the ResponseCompleted() event comes in,
then ResourceLoader::ResponseCompleted() stores the certificate in the
CertStore again, this time associating it with process B and reusing the
same cert id, so everything ends up working okay.

If, however, process A gets cleaned up before ResponseCompleted(), then
ResponseCompleted() will receive a new cert id, and the old one will no
longer work -- but the renderer has already saved the old one, and
that's what gets associated with the NavigationEntry. So we end up with
no certificate for the navigation, which messes up all the security UI
and means that the user can't view certificate details.

This CL updates the cert id that gets sent to the renderer during
UpdateRequestForTransfer(), so that the renderer process B receives the
new cert id rather than the old one which might no longer be valid.

BUG=561754

Review URL: https://codereview.chromium.org/1734723002

Cr-Commit-Position: refs/heads/master@{#381335}