Add 5 experimental accessibility features on Chrome OS.
This just adds the boilerplate for five new accessibility features and hides
them behind a flag. All five have been prototyped and are ready for
implementation.

The first three are related, they just provide visual highlighting to
help low-vision users keep track of things on the screen:
* Caret highlight
* Cursor highlight
* Focus highlight

Hold down a key and click or drag to speak any visible text on screen
* Select to speak

Control the whole computer with a single switch (for motor impaired users)
* Switch access

For now, there's no need to put these in the tray. They can graduate to the
tray when they launch.

BUG=314889,593887,593885

Review URL: https://codereview.chromium.org/1785833002

Cr-Commit-Position: refs/heads/master@{#381353}