Make JumpList NonThreadSafe
JumpList has a WeakPtrFactory and was a RefCountedThreadSafe with
DeleteOnUIThread. On this combination, a WeakPtr keeps valid after all
strong reference of JumpList is gone, until the JumpList instance is
deleted on UI thread. If a strong reference is created from WeakPtr in
this period, JumpList will eventually be deleted twice.

This CL replaces the RefCountedThreadSafe with RefCount, marks JumpList
as NonThreadSafe, and keeps it on UI thread to avoid the issue.

BUG=594270

Review URL: https://codereview.chromium.org/1802253002

Cr-Commit-Position: refs/heads/master@{#381380}