Speculative fix for the NPE in dispatchGetDisplayList
In https://codereview.chromium.org/1638303002, we began potentially
removing the overscroll refresh layout at the end of the animations.

According to this article:
http://stackoverflow.com/questions/15636658/nullpointerexception-when-adding-removing-views-of-a-relativelayout-animations

That could potentially be the source of the crashes we are seeing.

Changing it to post the detach instead of calling it synchronously in
the animation listener.

BUG=590359

Review URL: https://codereview.chromium.org/1800293004

Cr-Commit-Position: refs/heads/master@{#381595}