[Mac] CREDENTIAL: Fix chooser dialog after r380352
https://codereview.chromium.org/1781433002 enabled auto-resize for
constrained web dialogs on Mac. It updated a number of callsites for
`new ConstrainedWindowMac`, but didn't update the callsite in
`password_prompt_view_bridge.mm`.

Tests continued to pass, as the chooser _was_ being popped up, it was
simply constrained to a 0x0 frame. According to
https://codereview.chromium.org/1781433002/#msg15, it's not clear that
we can add a test that would verify that this doesn't regress. That
seems unfortunate. :(

BUG=217034,
R=vabr@chromium.org, apacible@chromium.org, thakis@chromium.org
TBR=vasilii@chromium.org

Review URL: https://codereview.chromium.org/1809043002

Cr-Commit-Position: refs/heads/master@{#381740}