Start using the whitelist icon on the NTP.
When the whitelist is created we save the path to the locally stored
icon. During the suggestions creation save that information and when
trying to get the large icon for the NTP, first check if we have a
locally stored one first.

This does not yet handle the case when whitelist entry point becomes
part of most visited.

For more information on whitelists - go/chrome-whitelists

Screenshot - https://screenshot.googleplex.com/3eTZ3wghRvH.png

BUG=586097

Review URL: https://codereview.chromium.org/1772363002

Cr-Commit-Position: refs/heads/master@{#381756}