Password manager on iOS should understand non-HTML landing pages
Currently, password manager on iOS ignores all pages which are not HTML.
However, if a password form has a, e.g., text/plain landing page, this will
prevent password manager from saving a password on that form. This is
different from the desktop & Android behaviour.

This CL changes password manager on iOS, so that it considers all types of
landing pages, but unless these are HTML, does not try to find password forms
on them.

The test for the changed code is still downstream, and hence not in this CL.

BUG=595717

Review URL: https://codereview.chromium.org/1806333005

Cr-Commit-Position: refs/heads/master@{#381949}