Makes it possible to set a factory that creates MojoShellConnection
And exposes MojoShellConnection::Create() so that the factory can
create the MojoShellConnection in the right way. An alternative is to
declare the closure as one that gets the ShellClientRequest and
is_external. If I did that I wouldn't need to make Create()
public. Let me know which you prefer.

BUG=581733
TEST=test only changes
R=jam@chromium.org

Review URL: https://codereview.chromium.org/1809373003

Cr-Commit-Position: refs/heads/master@{#382024}