Merge copy-pasted code into common base class
The image decoders for PNG, JPEG, and WEBP have the same
base class.

They also have the same copy-pasted block of code to compute
the color space transform from their source color profile to the
output device color profile. Move that to the base class.

Also move the methods to query whether or not a color profile has
been specified. Of note is that in all instances this is called
"hasColorProfile", but actually specifies if there is a transform. If
the source and destination color profiles matched, then there is
indeed a profile, but no transform. Do not change this behavior.

Make use of OwnPtr for QCMS structures.

Delete the helper class OutputDeviceProfile which was
instantiated as a static local, and replace it with a single global
variable with a lock.

Delete some other now-unused data structures and typedefs.

BUG=

Review URL: https://codereview.chromium.org/1816493002

Cr-Commit-Position: refs/heads/master@{#382166}