[MediaRouter, Android] Marking MediaRouterIntegrationTest as non-manual
The MediaRouterIntegrationTests (instrumentation tests) fail on
internal bots. The fake devices do not show up in the dialog.
This CL suppresses the device not found failure, and re-enable
the tests. More logging is added to see what is going on when
the tests fail.

BUG=587870

Review URL: https://codereview.chromium.org/1800653002

Cr-Commit-Position: refs/heads/master@{#382280}