Always apply parent clips to selection rects when painting.
Previously, BoxClipper would try to avoid clipping if it appeared the clip had
no effect, by comparing with contentVisualOverflowRect().
Now that selections paint newline placeholders, the selection rect may
extend beyond the overflow rect. (Selection is not part of overflow, since it is
required for it not to affect layout.)

BUG=594962

Review URL: https://codereview.chromium.org/1817463004

Cr-Commit-Position: refs/heads/master@{#382313}