[Offline pages] Fixing disappearing bolt when switching from incognito
This patch fixes the following issue:
When switching form an incognito tab with offline page opened, to a regular tab with an offline page, the status gets updated to the correct color, but the icon disappears, or actually does not get updated. This happens when tabs are merged with other applications only, which is only affecting phones.

BUG=595649

Review URL: https://codereview.chromium.org/1821003003

Cr-Commit-Position: refs/heads/master@{#382445}