Define a Mojo Notification structure in the Blink API
This will replace the PlatformNotification{Action, Data} structures
that currently live in //content. It's based on a similar CL for Web
Bluetooth: https://codereview.chromium.org/1784613002/

The stack will be switched over in a next CL. Since it's part of a
larger-scale refactoring, I'd like to land this separately to unblock
other people that will depend on this.

BUG=595685

Review URL: https://codereview.chromium.org/1814813002

Cr-Commit-Position: refs/heads/master@{#382564}