Use whitelist large icon for corresponding most visited suggestions.
If we get a most visited suggestion that is also a whitelist entry
point, use the icon from whitelist when creating he suggestions. This
is done in order to have consistently good large icon image on the NTP.
The reasoning for this comes from the fact that a lot of websites do
not have a large icon which would mean that while the user is browsing
and a site transitions from an entry point to a most visited suggestion
we will lose the icon image and will replace it with a color. That is
why we need to do this check and keep the flow consistent.

BUG=586097

Review URL: https://codereview.chromium.org/1787633002

Cr-Commit-Position: refs/heads/master@{#382601}