Make Platform return a WebGraphicsContext3DProvider* always.
Currently we return a WebGraphicsContext3DProvider* for the shared
main thread context (for canvas), but we return a WebGraphicsContext3D*
for webgl.

After this CL, Platform will return a WebGraphicsContext3DProvider*
for webgl also, which can grab the WebGraphicsContext3D* and the
GLES2Interface* off it.

This is largely in preparation for exposing gpu::ContextSupport (or
something wrapping it) on WebGraphicsContext3DProvider so that we can
expose lost context callbacks through there instead of on
WebGraphicsContext3D.

R=kbr@chromium.org, pfeldman@chromium.org, piman@chromium.org
TBR=pfeldman
BUG=584497
CQ_INCLUDE_TRYBOTS=tryserver.chromium.win:win_optional_gpu_tests_rel

Review URL: https://codereview.chromium.org/1822993002

Cr-Commit-Position: refs/heads/master@{#382657}