Add keyboard accessibility to ConstrainedWindowButton
This patch added keyboard accessibility to ConstrainedWindowButton.

ConstrainedWindowButton is used by WebUSB chooser bubble,
Geolocation permission bubble, etc. Previously, when trying to only
interact with these bubbles with the keyboard, there are no focus
styles. User should be able to press tab to navigate to buttons, and
see a visible focus state. It will allow the user to only use keyboard
to control the bubble, which improves the accessibility of the UI
that uses it.

BUG=596671

Review URL: https://codereview.chromium.org/1829393002

Cr-Commit-Position: refs/heads/master@{#383368}