Move make_scoped_ptr into //base/memory/ptr_util.h.
Preparatory CL for renaming scoped_ptr to std::unique_ptr. One of the
changes is replacing all uses of make_scoped_ptr() with WrapUnique(),
which lives in a separate header.

In order to incrementally update files to use the new header, this CL
moves make_scoped_ptr to ptr_util.h but still includes it in
scoped_ptr.h to avoid breaking the build during the transition.

BUG=554298

Review URL: https://codereview.chromium.org/1838323002

Cr-Commit-Position: refs/heads/master@{#383815}