media: AVDA now always notifies GVD of consumed bitstream buffers
In AVDA::QueueInput it was possible that a pending bitstream buffer was
popped from the queue but then a failure caused it to return without
NotifyEndOfBitstreamBuffer. If AVDA was Reset before the error was
propogated to GVD, GVD and AVDA became out of sync.

Now the bitstream buffer notification is done regardless of whether
an error occurred.

BUG=598058

Review URL: https://codereview.chromium.org/1831323002

Cr-Commit-Position: refs/heads/master@{#383836}