Move the run() method in Manager, and make the needs_servers method private.
This is an optional minor refactoring.

Reason: Putting the main public method first after the constructor, rather than in the middle, seems to be a more natural ordering.

The rest of the methods in this class are helper methods not used outside
of this class, and thus should all be private, including needs_servers.

Review URL: https://codereview.chromium.org/1827833005

Cr-Commit-Position: refs/heads/master@{#383859}