Refactor audio and video decoder status into common file.
Completes a long standing refactor that rileya@ initiated in days of yore
before we had DecoderStream. Will make it easier to add a new entry for
PLATFORM_FAILURE in a followup CL.

Should be no functional changes, though it does include a fix to the
pepper video decoder shim which was silenting converting the int type
of VideoDecoder::Status to a int32_t which was expected to be PP_xxx
type value. This change sends PP_OK (== 0) where previously it was
sending VideoDecoder::kAborted (== 1).

BUG=595076
TEST=none

Review URL: https://codereview.chromium.org/1834303005

Cr-Commit-Position: refs/heads/master@{#384020}