Reject promise with InvalidStateError instead of null in createImageBitmap(Blob)
Right now, the createImageBitmap(Blob) reject promise with null when
the input image cannot be decoded, for example, the input image is a SVG
image. A recently change is made to the specs here:
https://github.com/whatwg/html/pull/884
which states that we should reject the promise with InvalidStateError
insetad of null.

This CL changes it such that it will reject the promise with a
InvalidStateError.

BUG=596251

Review URL: https://codereview.chromium.org/1841313003

Cr-Commit-Position: refs/heads/master@{#384045}