Deinline body of ImageSkiaStorage.
I'm going to be adding another function to this class, and it's a bit hard to
read since the class is rather long.

This also does some minor comment fixups and switches from the Matcher functor
to a lambda.

Should be no functional change.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/1841343002

Cr-Commit-Position: refs/heads/master@{#384057}