Remove premultipliedAlpha from WebGraphicsContext3D::Attributes.
This field is read from WebGLContextAttributes, and used in
DrawingBuffer, but is not part of creating a GL context. So just
pass it to DrawingBuffer and store it as a bool on that class.

R=chrishtr@chromium.org, kbr@chromium.org
BUG=584497
CQ_INCLUDE_TRYBOTS=tryserver.chromium.win:win_optional_gpu_tests_rel;tryserver.chromium.mac:mac_optional_gpu_tests_rel

Review URL: https://codereview.chromium.org/1846713004

Cr-Commit-Position: refs/heads/master@{#384369}