Updating Fake IPC Client class to override the base IPC client class.
This CL updates the Fake IPC client class so that it can be used with the
Message Handler class for testing.  Th Fake IPC Client class has two purposes,
one is as a stub which calls callbacks for testing (for the handler class) and
the other actually uses IPC for the tests (for testing the IPC Server class).

BUG=584463

Review URL: https://codereview.chromium.org/1844143003

Cr-Commit-Position: refs/heads/master@{#384434}