Update the chromium.win, tryserver.chromium win entries in MB.
This should provide entries for all of the bots on chromium.win
and tryserver.chromium.win (apart from the _variable bots) that
match how things are currently being built via gclient runhooks
and some of the bots on other waterfalls (like chromium.gpu,
chromium.gpu.fyi, and tryserver.chromium.angle) that are affected
in order to keep them in sync).

In other words, when this CL lands and we land a build-side CL
to flip all of the Win bots to MB, this should have no effect
other than starting to use MB to generate the build files.

R=brucedawson@chromium.org, thakis@chromium.org, kbr@chromium.org, aneeshm@chromium.org
BUG=481692

Review URL: https://codereview.chromium.org/1848173002

Cr-Commit-Position: refs/heads/master@{#384723}