Restore handling of new scroll when waiting to cancel current scroll on cc
https://crrev.com/1678713002 caused this regression. This CL restores the
old handling of a new user scroll when in state WaitingToCancelOnCompositor.

The current behavior is pretty broken when we are in WaitingToCancelAnimation
and a new user scroll arrives. willAnimateToOffset updates the target offset and
goes into WaitingToSendToCompositor. Consequently, ::updateCompositorAnimation
doesn't abort the previous animation and starts a new one with the updated
target and the two animations compete for scroll offsets.

With this CL, when we are in WaitingToCancelAnimation and a new user scroll
arrives, we do nothing and let ::updateCompositorAnimation cancel the running
animation. As subsequent user scrolls arrive, we will now just start a new
animation with the updated target (and the bug is that we will ignore the first
user scroll when in WaitingToCancelAnimation).

The remaining bug is less serious because in general this is a special case when
we are cancelling in the middle of a user scroll and then starting another
scroll right away. This happens in cases where a site listens for the scroll and
updates the scroll offset (programmatic scroll cancels user scroll), and the
user continues to scroll. FWIW, this is the current behavior in M49, and no real
issues have been reported around it. crbug.com/599876 tracks the correct
handling for this case.

BUG=598548

Review URL: https://codereview.chromium.org/1852753002

(cherry picked from commit d115573365c24d26fc93fe9b9d0976d93725e91d)

Cr-Original-Commit-Position: refs/heads/master@{#384714}
Cr-Commit-Position: refs/branch-heads/2661@{#482}
Cr-Branched-From: ef6f6ae5e4c96622286b563658d5cd62a6cf1197-refs/heads/master@{#378081}