MediaLog: Replace task posting in AddEvent with lock
Uses a lock to maintain coherency of sequence of AddEvent().  This
change will help upcoming change(s) that need a coherent view of the
most recently added pipeline and medialog error messages, regardless of
which thread calls AddEvent() and which thread queries for the error
message.

BUG=472253

Review URL: https://codereview.chromium.org/1846913004

Cr-Commit-Position: refs/heads/master@{#385024}