Use Blink's computed filterPrimitiveSubregion() as CropRect.
For filters which affect transparent black, we were falling back to Skia's clip rect to know which pixels to fill. This works for SVG, where we set the clip rect to the filter region. However, this is not possible for CSS filters on HTML content, where there may be multiple reference filters with different filter regions, so we don't actually set a clip rect at all.

The easiest fix is to set the crop rect unconditionally for all filters with a non-empty filter primitive subregion (aka all reference filters), which is what we do here. This also gets us closer to being able to removing the individual crop edges from Skia. This has minor pixel diffs in the affected tests.

BUG=599933

Review URL: https://codereview.chromium.org/1840303006

Cr-Commit-Position: refs/heads/master@{#385031}