Slightly simplify scrolling code
This new code is equivalent because adjustedScrollOffset() basically
does scrollPositionDouble() + scrollOffset(). The previous code here
did -offset + adjustedScrollOffset(), which is therefore equivalent
to just scrollPositionDouble().

I am not 100% sure why this code is necessary but I believe it is for
the Scrollbar->offsetDidChange() calls.

R=szager@chromium.org

Review URL: https://codereview.chromium.org/1863483002

Cr-Commit-Position: refs/heads/master@{#385081}