Fixing a Dr. memory failure in the remoting unit tests.
The test I checked in yesterday had a race condition in it which was detected
by the Dr. Memory bot (they run more slowly than the normal bots).  On a
normal bot, my runloop quits when the first (of ~11) message is written but
the other messages are processed and added to the vector by the time it is
checked.  On the Dr. Memory bot, this condition was exposed and the size check
failed.  The fix is to ensure that we are processing each message as it is
written/received and validating the size afterwards.

This is a reland of: https://codereview.chromium.org/1844423002/
Due to changes to the base files, I am sending this out as a new patch instead
of using the old CL.

BUG=599494

Review URL: https://codereview.chromium.org/1860693003

Cr-Commit-Position: refs/heads/master@{#385093}