Rename bindings test IDL members to use "origin trial" terminology
This is the second CL in a series to cleanup the bindings tests/results
for the [OriginTrialEnabled] attribute. The cleanups are in separate CLs
to simplify the final CL to make the [OriginTrialEnabled] and
[RuntimeEnabled] attributes mutually exclusive.

This CL renames any members in the test IDLs that still use the old
"experiment" naming. Some code was missed in previous CLs to rename the
attribute to [OriginTrialEnabled] from [APIExperimentEnabled].

CLs in the series:
1) https://codereview.chromium.org/1828973003/, Fix indentation in template
2) This CL, rename test IDL members
3) https://codereview.chromium.org/1838813002/, Separate bindings tests for [OriginTrialEnabled]

BUG=586594

Review URL: https://codereview.chromium.org/1830113003

Cr-Commit-Position: refs/heads/master@{#385177}