Fixing the Chromoting Test Driver host online retry logic
In order to increase the reliability of the CTD, logic was added to
refetch the host list every few seconds for ~1 minute until we saw an
online host with a matching name and jid.  This was to combat a backend
propagation delay we had occasionally seen.

We are still seeing failures for this code on the waterfall though as
the logic is not actually executed if the first check fails.  This logic
needs to be fixed so the tool is more tolerant to potential propagation
delays.

The code that checks for the host being available did not have any unit
tests so I added those as well.

BUG=600763

Review URL: https://codereview.chromium.org/1864433005

Cr-Commit-Position: refs/heads/master@{#385319}