Omnibox - Keyword Provider Shouldn't Bold What-You-Typed Queries
Currently we don't bold what-you-typed queries from the default search
provider, not do we bold what-you-typed queries when the keyword is
typed exactly.  We do bold what-you-typed queries for inexact keywords.

The change makes it so we don't.

It also retires a string that is no longer needed after this change.

BUG=593245

Review URL: https://codereview.chromium.org/1784233003

Cr-Commit-Position: refs/heads/master@{#385323}