[WebCL] Let WebCLEvent.prototype.getProfilingInfo() return 64-bit profiling info.
The profiling info of an event is a 64-bit value that describes the
current device time counter in nanoseconds
(https://www.khronos.org/registry/cl/sdk/1.1/docs/man/xhtml/clGetEventProfilingInfo.html),
but in previous implementation it was implicitly casted as a 32-bit
integer so that the return value can fit the binding code generated from
the IDL interface, and thus made the return value almost useless. I
believe it is a typo in the WebCL specification that caused the original
defective IDL interface, and this commit fixes the problem.

BUG=XWALK-6705