cc: Use scroll layer bounds as scrollable area size
LTI::ScrollableSize suspiciously took the bounds of the first child
of the scrollable layer as the scrollable area size because of some
racing condition in blink.

Now we are not sure if the racing condition still exists. This CL
tries to use the bounds of the scrollable layer itself and see if
it breaks anything.

BUG=233853, 598381
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1862623003

Cr-Commit-Position: refs/heads/master@{#385464}