Remove RefCountedLeakCounter.
This machinery was actually broken by the refactoring work to remove
DEFINE_DEBUG_ONLY_GLOBAL in https://codereview.chromium.org/1429713002/

The way the RefCountedLeakCounter worked was by keeping track of
instance counts and then logging in the destructor. The above CL
incorrectly switched to using DEFINE_STATIC_LOCAL to hold the counters
which meant the destructor would never run and the leak detection was
disabled.

These days leak detection uses InstanceCounters and WebLeakDetector
anyway so lets just remove this class. This doesn't add back the
instrumentation for classes which are not tracked by InstanceCounters
but were tracked by the broken RefCountedLeakDetector because many of
them seemed really arbitrary and the detection has been broken for
5 months now without anyone noticing. If those teams want to add back
detection they can use InstanceCounter in the future

Review URL: https://codereview.chromium.org/1862283002

Cr-Commit-Position: refs/heads/master@{#385606}