ServiceWorker: Remove iframes explicitly after a test is finished
On Firefox, reloading a test page which has an iframe controlling
unregistered SW doesn't release the worker though the iframe is already
released, and this results failure of the subsequent test.
Removing iframes after the test is finished would be a solution for this problem.

BUG=594032
TEST=./third_party/WebKit/Tools/Scripts/run-webkit-tests -t Default http/tests/serviceworker

Review URL: https://codereview.chromium.org/1859343002

Cr-Commit-Position: refs/heads/master@{#385699}