Settings People Revamp: Add unit test for ManageProfileHandler.
This is not a complete unit test, but tests the parts of
ManageProfileHandler that are "done" and in-use. This test will be
completed once the desktop shortcuts feature is implemented in Settings.

This test also makes a few simplifications / minor fixes to the
actual handler.

BUG=563721

Review URL: https://codereview.chromium.org/1862273003

Cr-Commit-Position: refs/heads/master@{#385796}