Enhance unit tests for normals.
We asserted that changing vertex winding shouldn't change a
polygon's normal when it doesn't have to.
We asserted that inversion on the X axis (or Y-Axis) should change the normal, but it shouldn't.
We also tested the normal calculation of degenerate polygons (colinear vertices) when these don't have to work, and weren't what the test was claiming to check.

This introduces DISABLED tests to fix, separates the set of normal
calculations into separate tests so that failures are reported better, and uses better formed polygons for the touching and splitting test cases.

BUG=595820
TEST=self
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1860823003

Cr-Commit-Position: refs/heads/master@{#386470}