Fix /manifestuac problems in remoting_desktop.exe
The gn build of remoting_desktop.exe was hitting these errors:

   d:\src\temp\lnk{99E31AC4-E039-4456-A69A-A8DCF4647836}.tmp :
   general error c1010070: Failed to load and parse the manifest.
   Windows was unable to parse the requested XML data.

These are caused by misuse of the /manifestuac linker option, but
that is not obvious from the error message so I filed a VS bug for
that issue:
https://connect.microsoft.com/VisualStudio/feedback/details/2555946

The first /manifestuac problem is that the level is supposed to be
specified as a string, not as a number. That is, requireAdministrator
is supposed to be specified, not 2.

The second problem is that all but the last MANIFESTUAC option is
ignored, and we were specifying three MANIFESTUAC options.

The third problem is that the linker requires single quotes around the
'requiresAdministrator' and 'true' options. It is impressive that the
documentation for /manifestuac never mentions this requirement.

The fourth problem is that when you have multiple options for
MANIFESTUAC the linker requires double quotes around the entire set of
arguments. That used to be impossible but was made automatic by
crrev.com/1879723002

I verified that the manifest file embedded in remoting_desktop.exe
contained:
   <requestedExecutionLevel uiAccess="true" level="requiresAdministrator"/>

BUG=354261

Review URL: https://codereview.chromium.org/1875333002

Cr-Commit-Position: refs/heads/master@{#386560}