Removing the check for SM_TABLETPC for determining whether a device is operating as a tablet.
That appears to always return false on Windows 10 regardless of what mode the device is in
(tablet mode) or desktop mode without a keyboard.

This should fix the issues folks are reporting with the keyboard not popping up when
we detach a keyboard with tablet mode off.

BUG=497381

Review URL: https://codereview.chromium.org/1879953002

Cr-Commit-Position: refs/heads/master@{#386700}