Make OneClickSigninBubbleView a real dialog instead of a faux dialog
It supported both dialog and non-dialog version, but the bubble version
was never used. When in dialog mode, it was reimplementing a lot of basic
dialog functionality, so the code is much simpler now.

The UMA is largely removed because it's hard to have the same type of
UMA and I don't want complicate the code for the sake of some UMA that I
doubt anyone ever looks at. The bubble manager work will theoretically
one day add UMA back.

For Cocoa, the dialog and bubble code paths were already separate, so just delete the bubble path.

BUG=585312

Review URL: https://codereview.chromium.org/1859883002

Cr-Commit-Position: refs/heads/master@{#386886}