Improvements in restart scheme for audio input on Mac.
The current version works but it does not work OK for kMaxNumberOfRestartAttempts > 1 (now set to 1). Setting it to 2 would result in an infinite sequence of restart attempts and that is resolved with this CL.
Hence, this change can be seen as a cleanup CL and one that ensures that the scheme does what it says also for
other values of the design constants.

BUG=549021

Review URL: https://codereview.chromium.org/1881903003

Cr-Commit-Position: refs/heads/master@{#387014}