UTF-8 detector for pages missing encoding info
Experiment crbug.com/518968 shows that about 30% of the pages
missing character encoding information are encoded in UTF-8.
This CL runs a quick one-pass scan against them to check if
they are encoded in UTF-8. It helps get the encoding right
for 30% of the pages without having to turn on full auto-encoding
detection logic which is disabled due to slow performance.

BUG=583549
TEST=Layout test: fast/encoding/unlabelled-non-ascii-utf8.html

Review URL: https://codereview.chromium.org/1721373002

Cr-Commit-Position: refs/heads/master@{#387209}