Implement StyleValue.parse for CSS Typed OM
Spec: https://drafts.css-houdini.org/css-typed-om/#stylevalue-objects

For now it only supports parsing values that result in
SimpleLength objects - this is dependent on implementing other
types in the StyleValueFactory.

Note: parse returns an object? instead of a
(StyleValue or sequence<StyleValue>)?, because the code generator
generates "#error not implemented yet" in the latter case.

BUG=545318

Review URL: https://codereview.chromium.org/1867763002

Cr-Commit-Position: refs/heads/master@{#387255}