Decouple GLHelper from media
We would like to share GLHelper with Mus. In order to do so, we would like
to place this file in gpu/. Thus, it cannot depend on media/.

This CL removes the dependency on media::VideoFrame by passing in the
stride and data of the Y, U, and V planes into ReadbackYUV instead of
passing in a media::VideoFrame. The callers are responsible for keeping
media::VideoFrame alive now instead of ReadbackYUV.

This CL also moves media::LetterboxYUV to the callers.

BUG=601866

Review URL: https://codereview.chromium.org/1893473002

Cr-Commit-Position: refs/heads/master@{#387701}