bluetooth: Separate notification tests
We separate the tests so that we don't have to manually stop notifications
after each test is done. Because notifications state is kept in the frame
and the frame cleans up the state after navigation we know the tests
won't interfere with each other.

BUG=554240

Review URL: https://codereview.chromium.org/1858103004

Cr-Commit-Position: refs/heads/master@{#387734}