Attempt to fix a hang in the GPU process while creating a D3D9 device.
Mimicing some of the checks in the firefox sources to avoid creating a device
if we find that the driver does not support H/W format conversion and to
use D3DCREATE_MIXED_VERTEXPROCESSING instead of D3DCREATE_HARDWARE_VERTEXPROCESSING. As per docs it means
that the driver falls back to software vertex processing
which is still better than doing the whole decoding on the CPU.

Additionally I removed the D3DCREATE_DISABLE_PSGP_THREADING flag which restricts
computation to the main thread in case the driver fallsback to software vertex
processing which could affect performance.

The expectation with this patch is we may be able to avoid device creation for
a bunch of buggy drivers.

BUG=602389

Review URL: https://codereview.chromium.org/1884133004

Cr-Commit-Position: refs/heads/master@{#387738}