Mac: Update the tooltip view tooltip when a tab's title changes.
Currently, when muting a tab, the tooltip only updates when hovering off
of a tab, then hovering back on.

This is because only the tab view's tooltip updates, but there's an
optimization on mac that involves adding a separate transparent view to
handle tab tooltips. If that tooltip view is for the changed tab, ensure
its tooltip is updated as well.

BUG=599401
TEST=On Mac, reveal a tab's tooltip, then mute/unmute a tab and
immediately hover over it again. Ensure the tooltip is updated
appropriately.

Review URL: https://codereview.chromium.org/1855503002

Cr-Commit-Position: refs/heads/master@{#387840}