. so just require a paint invalidation and ensure visual overflow gets recomputed.
Another go at https://codereview.chromium.org/561303002 which was reverted by
https://codereview.chromium.org/747493002.

This includes wangxianzhu's fix for tables in https://codereview.chromium.org/744493002.

The intent of this patch is to avoid layout for all types of LayoutBlock when
the outline, border-image-outset or box-shadow are updated.

BUG=396825,434700,437012

Review URL: https://codereview.chromium.org/1809643008

Cr-Commit-Position: refs/heads/master@{#387862}