[Temp] Make |enable_webvr==0| work.
**IMPORTANT**
We actually intend to fix this properly. We are cherry-picking this
commit from the crosswalk-18 branch first in order to have a stable
baseline with no WebVR support propagating from Crosswalk 21 to 20 to
19. After that, we can start working on fixing this in 21 and 20 and, if
there is enough time, 19 too.

*From the original commit message:*

Move `CardboardVRDevice.java` to a different directory so that it is not
built by the `content_java` target even if `enable_webvr==0`.

This is a prerequisite for getting the code to build with
`enable_webvr=0`.

BUG=XWALK-6746

(cherry picked from commit e5934186311cdf68e737a88b6cf1dbf23eb0464a)

M53 rebasing notes: added GN support:

> [Temp] Make |enable_webvr==false| work in GN.
>
> This is a change that should be squashed into "[Temp] Make
> |enable_webvr==0| work", and just adds the required changes to the GN
> build so that it is aware of the fact that CardboardVRDevice.java
> resides in another directory.
>
> BUG=XWALK-6746