Don't clear metrics_client_id on shutdown
Because we use the same path for disabling as shutdown, the
metrics_client_id crash key gets cleared during browser shutdown. This
is undesireable because then it won't be there for crashes that happen
late in shutdown. Avoid this by simply not clearing it when shutting
down.

R=asvitkine@chromium.org
BUG=604487

Review URL: https://codereview.chromium.org/1894973005

Cr-Commit-Position: refs/heads/master@{#388215}