Fixing a Dr. Memory hang in the remoting_unittests
There were two problems that needed to be addressed.  One was a failure that
occurred if a stream writer wrote the message type and payload in separate
write actions (timing problem) the second was a test bug in the
MultipleMessages test which did not check for success/failure for each loop
iteration.

The first problem was a timing issue where the reader might fail if the message
payload was written separately from the type and was not written quickly.  This
was not repro'd on normal builds but we see it on the Dr. Memory bot since
operations are slowed down as a result of the build flags used for the tool. I
fixed this problem by reading in the type and payload separately.

The second problem was in the unit test itself.  If an error occurred in the
Multiple messages test, the loop would write to the end of the pipe and wait
(forever) for an answer.  We now verify the message was written after every
iteration.

BUG=603357

Review URL: https://codereview.chromium.org/1900733002

Cr-Commit-Position: refs/heads/master@{#388307}