Fix an issue that referrer field is no longer exposed to pdfviewer
New pdfviewer starts to handle http scheme.
As a result, clicking a pdf link will no longer trigger a download.
Instead, the navigation will be intercepted and passed to pdfviewer.
We previously added the referrer extra for analytics purpose for pdfviewer.
But that only happens in the download code.
This CL fixes the referrer field issue for intercepted navigations.

BUG=603730

Review URL: https://codereview.chromium.org/1887173003

Cr-Commit-Position: refs/heads/master@{#388535}