Invalidate LinkHash state for links in shadow roots
When link history state changed, :visited style can change, but
it used to work only for links in document node tree.
This CL fixes to make it work also for links in shadow node trees.

Writing a layout test for this is harder than expected; tried using
<iframe> to change history state, but it did not working as expected.

Here's a manual test for this case.
http://jsbin.com/yipozul/6/edit?html,output

BUG=405707
TEST=manual

Review URL: https://codereview.chromium.org/1910433002

Cr-Commit-Position: refs/heads/master@{#388710}