Fix tests under the password-separated sign in flow before launch.
The password separated sign in flow changes some assumptions that breaks/makes some tests obsolete but it's currently behind a Finch Feature Flag.

This CL either adapts (if relevant) the failing tests or disables the ones that are obsoleted by the new flow in preparation for the Finch rollout.

BUG=533004

Review URL: https://codereview.chromium.org/1888863002

Cr-Commit-Position: refs/heads/master@{#389113}