Remove obsolete label from DRP histogram suffixes
https://codereview.chromium.org/1889873005/ added
<obsolete> for a group of histogram suffixes. The goal was
to mark DataReductionProxy.AutoLoFiAccuracy.* as obsolete.
However, that caused some other unrelated histograms (e.g.,
DataReductionProxy.AutoLoFiRequestHeaderState.*) to be
marked as obsolete.

This CL removes the obsolete label from the suffixes. It
should not be required because the base histogram
DataReductionProxy.AutoLoFiAccuracy is already marked as
obsolete.

BUG=

Review URL: https://codereview.chromium.org/1912903002

Cr-Commit-Position: refs/heads/master@{#389153}