Delete //gpu/blink/ and WebGraphicsContext3DImpl.
There is now only 1 subclass of WebGraphicsContext3DImpl (which is the
infamous WebGraphicsContext3DCommandBufferImpl), so let's delete this
in-between thing.

The use of GLES2TraceImplementation was broken, since
ContextProviderCommandBuffer was returning the underlying
GLES2Implementation directly now (woops). Fix that and move it to
the ContextProviderCommandBuffer.

Move the lost context callback to WebGraphicsContext3DCommandBufferImpl,
and remove the definition of the proxy class out of WebGraphicsContext3D
leaving the blink base class empty now.

Lots of DEPS and include removals all over the place that were already
obsolete.

R=piman, sievers
TBR=chrishtr
BUG=584497

Review URL: https://codereview.chromium.org/1907783002

Cr-Commit-Position: refs/heads/master@{#389257}