GN: VS project calls ninja for single file compiles as well
Previously, compiling a single file in the generated project
invoked the normal compiler in Visual Studio.
Now we call ninja instead, this way it will be compiled with the
same options as it would compile when building the whole project.

It's now also possible to properly compile a file when using a
cross-compiler.

R=brettw@chromium.org, tmoniuszko@opera.com
BUG=

Review URL: https://codereview.chromium.org/1835943002

Cr-Commit-Position: refs/heads/master@{#389443}