Paint media element's play button using padding.
Rather than specify only a box size in CSS, we now specify the box
size and padding to draw the play and pause buttons.  This also
requires the image to be reduced to 32 image pixels square.

In order to handle desktop zoom, the images are now specified with
FloatRect rather than IntRect.

BUG=547955

Review URL: https://codereview.chromium.org/1424733003

Cr-Commit-Position: refs/heads/master@{#389780}