Fix sizing of the verification overlay in unmask prompt.
This patch prevents the verification overlay for card unmasking prompt
from being cut off at the bottom. That happened due to the overlay being
taller than the controls container, while the size of the overlay was
constrained to the size of the controls container.

The fix is accomplished by moving the instructions ("Type the 3 digit
CVC from the back of your card") into the controls container.

Before:

|--------------|     |--------------|
|Type card cvc.|     |Type card cvc.|
|--------------|  <  |--------------|
|cvc:___       |     |      ✓       |
|--------------|     |   Verified   |
                     |--------------|

After:

|--------------|     |--------------|
|Type card cvc.|     |      ✓       |
|cvc:___       |  =  |   Verified   |
|--------------|     |--------------|

BUG=604274

Review-Url: https://codereview.chromium.org/1916923003
Cr-Commit-Position: refs/heads/master@{#390232}