Enabled tests to control material design ink drop animations.
The InkDropAnimationControllerImpl choreographs the complex interaction of the
InkDropAnimations and InkDropHover animations. This interaction relies on
internal observations of when the different animations start and finish. This
change enables tests to control when these animations start and finish and allow
tests to validate states that was previously impossible to inspect when using
the ui::ScopedAnimationDurationScaleMode::ZERO_DURATION value.

BUG=604415
TEST=InkDropAnimationControllerImplTest.*

Review-Url: https://codereview.chromium.org/1913243002
Cr-Commit-Position: refs/heads/master@{#390416}