Make WebContents accept user interactions from OOPIF RenderWidgetHosts
WebContentsImpl::OnUserInteraction currently returns early if it is
called for an event on a RenderWidgetHost other than its main frame's.

This means WebContentsObservers do not receive notifications for
interactions targeted to out-of-process iframes, causing issues for
components that rely on it.

This patch changes OnUserInteraction to recognize when a RWH belongs
to an OOPIF, and accept the interaction.

BUG=589894

Review-Url: https://codereview.chromium.org/1924333002
Cr-Commit-Position: refs/heads/master@{#390680}