Simplify GN target names for //blimp
As part of making the GN configuration for //blimp up to date with
current best practices, this CL changes top-level targets under
//blimp to have the same name as their folder.

As an example, //blimp/common:blimp_common can now be referred to
using only //blimp/common.

The reason why for example blimp_common was previously chosen was
to ensure that the output name of the component did not collide with
other targets in the chromium repository.

This CL ensures that old output names are kept in place by
specifying the output_name GN variable. This is only necessary for
build artifacts, and not for gn templates such as group or source_set.

BUG=597830

Review-Url: https://codereview.chromium.org/1930303002
Cr-Commit-Position: refs/heads/master@{#390682}