Allow dumping raw json fetched from the server.
On top of dumping the whole list of snippets, this adds the feature to
dump the json of the newly fetched snippets.

Furthermore, the CL fixes a minor bug discovered while testing the CL: the list of snippets was truncated to max_length before deleting outdated snippets - sometimes resulting in less snippets.

BUG=607518
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/1927993002
Cr-Commit-Position: refs/heads/master@{#390905}