Trackpad scrolling could lag on MacOSX
Mouse wheel gesture scrolling with synthetic begin and ends
could saturate the gesture event queue. Ensure that
a synthetic gesture begin coalesces with a synthetic
gesture end so that the updates end up coalescing as well.
This prevents laggy scrolling on MacOSX when it takes
a long time per frame to scroll.

BUG=607340

Review-Url: https://codereview.chromium.org/1932363002
Cr-Commit-Position: refs/heads/master@{#390942}