[New experiment] Enroll 1% dev into StrictMode watch.
This CL allows 1% of all dev sessions to upload 1% of the StrictMode
violations they encounter. These stack traces are uploaded to our
crash server and will help to identify jank that we have not found
before. All the known violations have been whitelisted so that only
unknown ones will come through this experiment.

It is intended only for dev channel, this will not affect beta or
stable.

BUG=531546

Review-Url: https://codereview.chromium.org/1925703002
Cr-Commit-Position: refs/heads/master@{#391017}