content_shell fixes for gn
content_shell is big enough, in non-component builds, that incremental
linking is not reliable. Therefore incremental linking needs to be
conditionally disabled. This is the equivalent of GYP's
msvs_large_module_debug_link_mode.

The GYP build disabled incremental linking for content_shell_lib as well
but this is meaningless and is sanely prohibited by gn.

Also, chrome.dll had a redundant check of is_component_build which this
change removes.

Finally, this adds shell.rc to the build, necessary for dialogs and
menus, thus finishing all of the Windows TODO(GYP) items in this file.

BUG=603015

Review-Url: https://codereview.chromium.org/1930073002
Cr-Commit-Position: refs/heads/master@{#391018}