Run tests after main frame loads.
Tests modified by this CL were calling window.addEventListener('message', ...)
to listen for a message posted by a subframe.  The subframe (i.e.
cross-frame-iframe-for-enumeration-test.html) would post the message to the
parent frame in the frame's onload event handler.

The problem with the above is that there is no guarantee that message
listener will be registered *before* the message is posted.  In fact,
before this CL + with --site-per-process the message would be posted
*before* the listener is registered (resulting in runTest never getting
called and a test time out).

This CL fixes the problem above by running the tests after the main
frame loads.  At this point the subframe must have also loaded - there
is no need to synchronize by postmessage.

BUG=477150

Review-Url: https://codereview.chromium.org/1946673002
Cr-Commit-Position: refs/heads/master@{#391401}