[tracing] Add a memory dump provider for DOM storage
This CL adds a memory dump provider for DOM storage in browser. It only
accounts for the memory usage of the buffer map and session storage
database, not the local backing databases. This will be done in next CL.
The areas can be accessed only on DOMStorageTaskRunner's
PRIMARY_SEQUENCE. Since memory infra supports collection of dumps from
SequencedTaskRunner, a SequencedTaskRunner is extracted from
DOMStorageTaskRunner's worker pool to pass to memory-infra.

BUG=605785

Review-Url: https://codereview.chromium.org/1906243002
Cr-Commit-Position: refs/heads/master@{#391455}