Collect UMA data for Origin Trials
This CL records UMA metrics for the results of feature enabled checks,
one per context. Also records metrics for when error messages are
generated.

Token validation is controlled by the embedder, but the enabled check is
performed in Blink. To collect the desired metrics, it was necessary to
plumb a detailed validation return code through the chain.

BUG=587575

Review-Url: https://codereview.chromium.org/1909633003
Cr-Commit-Position: refs/heads/master@{#391505}