Add strings for four new browsing data volume counters.
These counters will be shown next to their respective checkboxes
in the Clear Browsing Data dialog.

For "Cookies and site data" and "Content licenses", we will just show static text.

For "Download history", we will show the number of items.

For "Hosted apps data", we will show the number of apps and up to two examples.

See the design doc:
https://docs.google.com/document/d/1xhDI8DCZi2viM_dUIl7-3yRCNCMGEM3wQR7AJrf24Wo/

BUG=510028

Review-Url: https://codereview.chromium.org/1946243002
Cr-Commit-Position: refs/heads/master@{#391570}