Reland: Use actual audio channel count in Spitzer audio decoder
The actual channel count used by MediaCodec for audio may be
less than requested in case of multichannel AAC.
MediaCodecAudioDecoder obtains the actual channel count
in response to OUTPUT_FORMAT_CHANGED and uses it for the
calculation of the sample length.

In this CL we keep creating output buffers according to the
requested configuration. If the actual number of channels is
less than the requsted this CL fills the missing channels
with zero data.

The reland contains the fix for improper use of the |size| and
|offset| obtained from DequeueOutputBuffer. The size of the
useful decoded data chunk is |size| and does not depend on |offset|.

BUG=607024

> Committed: https://crrev.com/f6d2c9a3aa0a72ff766a132ef7f245ea26f66dfc
> Cr-Commit-Position: refs/heads/master@{#391416}

Review-Url: https://codereview.chromium.org/1932093002
Cr-Commit-Position: refs/heads/master@{#391618}