Use forceClickOnCurrentItem to mean doDefault action rather than performDefaul\ tAction
When navigating to a new site with the below steps, we accidentally force focu\
s back onto the omnibox.
- ctrl+l
- type an address
- hit enter

result:
we stay in compat mode and we see focus is on the omnibox.

This is because:
- when we place focus via ctrl+l, we switch to compat mode
- that results in a switch to the classic keymap
-UU-:----F1  cl_descriptionSCqLE6   Top L1     (Fundamental) ------------------# Enter a description of the change.
Merge to m51: Use forceClickOnCurrentItem to mean doDefault action rather than performDefaultAction

When navigating to a new site with the below steps, we accidentally force focus back onto the omnibox.
- ctrl+l
- type an address
- hit enter

result:
we stay in compat mode and we see focus is on the omnibox.

This is because:
- when we place focus via ctrl+l, we switch to compat mode
- that results in a switch to the classic keymap
- the classic keymap binds the enter key to performDefaultAction
- in cvox2/background/backgorund.js, performDefaultAction does a |doDefault| action which sends a click
- after entering an address, we "click" the address bar
- this is racey with the focus generated by the webview

Solve this by
using the forceClickOnCurrentItem command (which is mapped to Cvox+space) to mean doDefault. This matches classic behavior. performDefaultAction in classic simply meant placing focus correctly when encountering internal links.
TEST=navigate between sites via the above reproduction; ensure classic is on and working.

BUG=605980

Review-Url: https://codereview.chromium.org/1915603002
Cr-Commit-Position: refs/heads/master@{#390248}
(cherry picked from commit 5f2b437436f41145c774d485602bd61edefc443c)

Review URL: https://codereview.chromium.org/1952803002 .

Cr-Commit-Position: refs/branch-heads/2704@{#377}
Cr-Branched-From: 6e53600def8f60d8c632fadc70d7c1939ccea347-refs/heads/master@{#386251}