ChildProcessSecurityPolicy::CanAccessDataForOrigin workaround to suppress bad kills
crbug.com/600441 apparently involves a race where we check
ChildProcessSecurityPolicy for a process_id that it doesn't know -- probably
a renderer shutdown race. Returning true instead of false here is a temporary
workaround to suppress these bad kills, which are affecting the non-oopif
user population. This is a tolerable short-term behavior as far as security
is concerned, since CanAccessDataForOrigin is only meant to offer meaningful
protection in --site-per-process mode, which is not yet launched.

BUG=600441

Review-Url: https://codereview.chromium.org/1945173003
Cr-Commit-Position: refs/heads/master@{#391678}