Remove googlezip.net proxy removal hack.
As a quick fix for http://crbug.com/476610, a hack
(https://codereview.chromium.org/1126413006) was added to remove
any proxy of the form *.googlezip.net from the proxy configuration
via a proxy pref or via the command line.

This hack is no longer necessary (see http://crbug.com/500111), so
this CL removes the hack.

BUG=500111

Review-Url: https://codereview.chromium.org/1929793002
Cr-Commit-Position: refs/heads/master@{#391914}