Add policy for disabling the Chrome OS lock screen
A new user policy "AllowScreenLock" is added. When the policy is set to
"false", the user should not be able to lock the screen (the lock screen
button will be hidden, the lock screen shortcut will be disabled, etc.).

If some other policies or settings request to lock the screen under
specific scenarios, then the user session will be stopped instead. The
examples of such scenarios:
* Device suspends with the "ChromeOsLockOnIdleSuspend" policy (or the
  "Require password to wake from sleep" setting) enabled.
* The screen is locked automatically after the idle timeout passes
  (specified through the "ScreenLockDelays" policy).

BUG=607129
TEST=unit tests, manual test on a dev chromebook with having different
     combinations of the new policy with the
     "ChromeOsLockOnIdleSuspend", "LidCloseAction", "ScreenLockDelays",
     "IdleAction", "IdleActionAC", "IdleActionBattery" policies and/or
     the "Require password to wake from sleep" setting

Review-Url: https://codereview.chromium.org/1932233002
Cr-Commit-Position: refs/heads/master@{#392035}