Keep UMA and crash reporting prefs in sync even when Chrome prefs are cleared.
Previously it was assumed that UMA state will change only from Android code and chromium prefs will be updated accordingly when change happens.
However it appeared that it is possible to change native prefs (for example delete the prefs) which will result in inconsistent behavior.
Therefore, this CL updates native prefs according to Android pref value every time a new session starts to keep all the prefs in sync.

BUG=603216

Review-Url: https://codereview.chromium.org/1916903002
Cr-Commit-Position: refs/heads/master@{#392106}