Use the Mojo StubBindings interface correctly in WebUSB LayoutTests.
Instead of subclassing the stubClass type of the Mojo service being
implemented we should call connection.bindHandleToStub to create a stub
and then set our implementation as its delegate. The existing code
instead unintentionally overrode the generated stub method bodies which
was mostly harmless but broke passing interface pointers.

While making this change it was discovered that if the DeviceManager
connection is closed a call to requestDevices() would still attempt to
access m_deviceManager.

BUG=None

Review-Url: https://codereview.chromium.org/1951723002
Cr-Commit-Position: refs/heads/master@{#392190}