Material WebUI: Switch to poly-icons to generate iconsets
This is a follow-up to https://codereview.chromium.org/1947533002/.

Instead of copying iron-icon definitions from third_party/polymer,
we generate them using https://poly-icon.appspot.com. This has the
advantage of being easier to reproduce and update, as a single URL
can generate the list of SVG <g> definitions neaded.

The actual definitions are the same, but the comments were changed
to reflect the new generation style.

Also, the shared icon file is renamed to "icons.html" to reflect
similar conventions in MD Extensions and (soon) MD settings, and
the iconset prefix is simply "cr".

BUG=605821
R=dbeam@chromium.org
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/1957313002
Cr-Commit-Position: refs/heads/master@{#392360}