[Media Router WebUI] Run 'select cast mode' test in afterNextRender();
There's a race condition where afterNextRender() is not always called before setTimeout(). 'select cast mode' appears to be the only failing test right now, so this changes runs the test within afterNextRender().

However, all of the tests (especially if it relies on the listeners) should be run in afterNextRender() to ensure setup is complete. Further work will be done specifically for Media Router WebUI tests to reflect this.

BUG=609411

Review-Url: https://codereview.chromium.org/1961223002
Cr-Commit-Position: refs/heads/master@{#392516}