Make PassOwnPtr<T> an alias to OwnPtr<T>.
This patch makes PassOwnPtr<T> a mere alias to OwnPtr<T>. That means,
after this is landed, OwnPtr<T> and PassOwnPtr<T> are the same type!

Surprisingly, I didn't have to update most of the user code.

Because MSVC was not very happy with alias templates (it generated
some internal compiler errors), PassOwnPtr is renamed with a macro
on MSVC.

BUG=582350

Review-Url: https://codereview.chromium.org/1942673002
Cr-Commit-Position: refs/heads/master@{#392564}