Set SkipServiceWorker flag in DocumentThreadableLoader::loadActualRequest().
Before a SW controls the page, when the DTL sends a CORS preflight request, the SkipServiceWorker flag is set by RenderFrameImpl::willSendRequest().
But a new SW may be controlling the page when the DTL sends the actual request by calling clients.claim().
In such case, the request goes to the SW and causes several problems.

To avoid this problem, this cl set the flag in DTL::loadActualRequest().

BUG=604583,610400

Review-Url: https://codereview.chromium.org/1964823002
Cr-Commit-Position: refs/heads/master@{#392609}