Use didStartProvisionalLoadForFrame to trigger the progress bar
The progress bar should traverse the page a single time per
page load. To do this, the primary signal for starting the progress
bar should be the loading of the main frame, not the sub-frames.

This changes the triggering of the progress bar to be primarily
based on the didStartProvisionalLoadForFrame event. This event
is the earliest indication that a page has started loading that
also tells us whether or not it is the main frame.

Events including onLoadStarted, onDidNavigateMainFrame, and
onPageLoadStarted either occur for every frame without indicating
if it was the main frame or are called too late to show early
progress.

BUG=573177

Review-Url: https://codereview.chromium.org/1940583002
Cr-Commit-Position: refs/heads/master@{#392421}
(cherry picked from commit aae2cc896cf2e7c48ea637efdc392505502ff74a)

Review URL: https://codereview.chromium.org/1963913004 .

Cr-Commit-Position: refs/branch-heads/2704@{#471}
Cr-Branched-From: 6e53600def8f60d8c632fadc70d7c1939ccea347-refs/heads/master@{#386251}