Remove ApplicationStatus.getApplicationContext.
ContextUtils.getApplicationContext has superceded the method in
ApplicationStatus. The application context should not be cast back
to a ChromeApplication, but the code that currently relies on it must
be refactored before they can avoid the cast. This is safe right now
as all that code is in chrome/ and ChromeApplication initializes its
own reference in ContextUtils on startup.

BUG=552419

Review-Url: https://codereview.chromium.org/1944843002
Cr-Commit-Position: refs/heads/master@{#392706}