[GN] New template convert_plist to convert a plist file to another format.
Add a new target convert_plist to invoke "plutil -convert" only on a
given plist file. This is used as iOS needs to convert all plist to
binary in bundle, including plist that are not Info.plist files.

Fix ios_info_plist and info_plist templates to also forward "visibility"
from the invoker.

BUG=459705

Review-Url: https://codereview.chromium.org/1964393002
Cr-Commit-Position: refs/heads/master@{#392924}