Run queued tasks before shutting down after running a layout test.
Currently, when a layout test run finishes, the BlinkTestController
falls out of scope, which triggers shutdown of the RenderProcessHostImpl
used by the test. However, RenderProcessHostImpl posts a task to delete
itself, which is currently never run; this results in a
RenderProcessHostImpl outliving the UI MessageLoop and its
BrowserContext, inviting use-after-frees. This CL fixes this issue by
running queued tasks before shutting down the main runner.

BUG=610989

Review-Url: https://codereview.chromium.org/1975593003
Cr-Commit-Position: refs/heads/master@{#393204}