Add histograms for notices about other forms of browsing history
Notices about the existence of other forms of browsing history on
history.google.com are shown to users in the header of the History page,
footer of the Clear Browsing Data dialog, and an additional dialog on top
of the CBD dialog after history deletion; but only to those users for
whom they are relevant.

This CL adds histograms that are recorded whenever one of the described
UI elements is shown and records whether we also showed the notice or not
(thus computing the amount and ratio of users for whom we decided to show
the notice).

Implemented:
- History page (shared between Desktop and Android)
- History page (iOS)
- CBD dialog (Desktop WebUI)
- CBD dialog (Desktop MD, only the footer)
- CBD dialog (Android)

Missing (will be added in followup CLs):
- CBD dialog (iOS) - not part of Chromium repo
- CBD dialog (Desktop MD, history deletion dialog) - does not exist yet

BUG=608842
TBR=dbeam@chromium.org,jyquinn@chromium.org,dfalcantara@chromium.org,rkaplow@chromium.org
NOTRY=true
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/1948633002
Cr-Commit-Position: refs/heads/master@{#391682}
(cherry picked from commit 1f22966d3e45226f3048405f9f5d7d6898285d76)

NOTE! A conflict appeared in
chrome/browser/ui/webui/settings/settings_clear_browsing_data_handler.cc,
as in the 2704 branch this file does not contain the functionality which
the histogram was supposed to measure. There, we simply keep the state
of the file on HEAD.

Review-Url: https://codereview.chromium.org/1968983002
Cr-Commit-Position: refs/branch-heads/2704@{#519}
Cr-Branched-From: 6e53600def8f60d8c632fadc70d7c1939ccea347-refs/heads/master@{#386251}