Use gold and the bundled binutils on x86 by default
The only reason we were not using gold by default on x86
is that using icf was buggy. However, we think we've updated
to a working version of gold, and so we should now use it.

R=mcgrathr@chromium.org, thakis@chromium.org
BUG=590004
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_chromium_compile_dbg_32_ng,linux_chromium_dbg_32_ng

Review-Url: https://codereview.chromium.org/1951133002
Cr-Commit-Position: refs/heads/master@{#393645}