media: supportsOverlayFullscreenVideo() now stays constant while fullscreen
This is a follow up to https://codereview.chromium.org/1966053002 to
handle considers the edge case of changing decoders, and therefore
possibly changing the return value of supportsOverlayFullscreenVideo(),
while fullscreen. Without this change we can get into an inconsistent
states because some of the OverlayFullscreenVideo set up happens only
once when entering fullscreen. Specifically, it was possible to see
a white background if a H/W underlay video was fullscreened before it
was loaded.

BUG=610812

Review-Url: https://codereview.chromium.org/1980653003
Cr-Commit-Position: refs/heads/master@{#393684}