[WebCL] Include what you use.
The patch fixes most (if not all) style issues to follow the principle
of including what you use. A self-contained header should include the
necessary headers so that we can simply use it in other files without
worrying about its dependencies. On the other hand, a header file should
only include what is really needed and use forward declarations for
external symbols if possible. This will not only help in faster
compilation and fewer recompilation, but also avoid circular header
inclusion which can actually cause compilation errors that are hard to
diagnose in some cases. Also, if a header file is needed, then it should
be included *explicitly*, instead of depending on the fact that it is
*implicitly* included by some other random files which are subject to
change in the future.