Move in-GPU mojo:media app to new thread
This app was moved to the main thread of the GPU
process when I killed shell::Loader.
See https://codereview.chromium.org/1889253002.

We want this running on its own background thread
because it does slow things.

This CL adds the option for embedded mojo apps to
be run on their own thread and uses the option
for mojo:media in the GPU process.

BUG=611070
TEST=watched cat videos on youtube.com with android build, which exercises in-GPU mojo:media

Review-Url: https://codereview.chromium.org/1978843003
Cr-Commit-Position: refs/heads/master@{#393852}