Revert of Client side display item cache flag (patchset #20 id:380001 of https://codereview.chromium.org/1508223005/ )
Reason for revert:
This caused crbug.com/609218. It seems that there is some short-lived
DisplayItemClient but needs more investigations to find which
DisplayItemClient it is.

Revert this to unblock m-52 branch and beta release.

Will reland on m-53 and add some code in release builds to find the
reason.

BUG=609218

Original issue's description:
> Client side display item cache flag
>
> This is to reduce cost of hash lookups in PaintController.
> See platform/graphics/paint/README.md and DisplayItemClient.h
> for how it works.
>
> BUG=550044
>
> Committed: https://crrev.com/93b971703db1f49a431ddeec2ebc884a232ee6de
> Cr-Commit-Position: refs/heads/master@{#390763}

TBR=chrishtr@chromium.org,pdr@chromium.org
# Not skipping CQ checks because original CL landed more than 1 days ago.
BUG=550044

Review-Url: https://codereview.chromium.org/1979133003
Cr-Commit-Position: refs/heads/master@{#393930}