Remove extra code in attribute service providers.
All the properties for our attribute service providers are read-only, so all
the code in the ::Set methods is useless. We also don't ever return the value
property from Get/GetAll (it is optional and we have the ReadValue method to
read it) so we need to get rid of all the code that calls the delegate to
get the value of an attribute also.

As a side-benefit, this simplifies our code for GetManagedObjects.

R=xiyuan@chromium.org
BUG=None.

Review-Url: https://codereview.chromium.org/1981043002
Cr-Commit-Position: refs/heads/master@{#393975}