media: Set RUNPATH in cdm adapter in GN component build
The CDM should always be able to find the CDM, regardless of how the
loader of the CDM adapter is configured.

In tests, currently, browser_tests sets a RPATH of $ORIGIN, which
makes the test work. But if we put the CDM adapter and CDM in a
different folder, then the test will break in GN component build.

See BUG for more detailed discussion.

BUG=612007
TEST=browser_tests pass in GN component build

Review-Url: https://codereview.chromium.org/1981723002
Cr-Commit-Position: refs/heads/master@{#394272}