Fix GPU IDC locking of CPU backed images
When the GPU Image Decode Controller uploads images to the GPU, it no
longer needs the discardable data once the upload is complete. However,
for images that are too large to upload directly, the GPU IDC creates a
thin wrapper image around the discardable to hand off to Skia. We need
to ensure that the discardable stays pinned for the lifetime of this
wrapper. This change adjusts the unpinning logic to wait until all
references to the image are released in the case of a CPU backed image.

BUG=609612,610544
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review-Url: https://codereview.chromium.org/1982913003
Cr-Commit-Position: refs/heads/master@{#394291}