Forbid Oilpan GC in blink::testing::runPendingTasks
As webkit unittests are run outside of the message loop, tasks run in
blink::testing::runPendingTasks are executed as non-nested tasks. That means
the registered GCTaskObserver will run GC with NoHeapPointerOnStack. This is
problematic because we often have heap pointers on stack.

This change forbid Oilpan GC in runPendingTasks to avoid such a problem.

BUG=613115

Review-Url: https://codereview.chromium.org/1995733003
Cr-Commit-Position: refs/heads/master@{#394743}