Move bindToCurrentThread out to the creators of WebGL contexts.
The ContextProvider::BindToCurrentThread method is going away, with
the ContextProvider being bound to whatever thread it is created on
and deferred creation via helper Factory classes.

The WebGL contexts currently optionally BindToCurrentThread when
the context is created. Instead, move it to the caller that creates
the provider.

In the future the WebGraphicsContext3DProvider will hold the
ContextProvider::Factory, and when bindToCurrentThread is called
it will use the Factory to instantiate the real ContextProvider.

In order to keep this optional, just have the creator of the context
always call WebGraphicsContext3DProvider::bindToCurrentThread.

R=kbr@chromium.org
TBR=pfeldman
BUG=487471
CQ_INCLUDE_TRYBOTS=tryserver.chromium.win:win_optional_gpu_tests_rel;tryserver.chromium.mac:mac_optional_gpu_tests_rel

Review-Url: https://codereview.chromium.org/1990063004
Cr-Commit-Position: refs/heads/master@{#394877}