Correctly set -stdlib=libc++ when building with a custom libc++.
The libcxx_proxy target used to set the link-time configuration to
correctly link libc++ sets it as a public_config. However, the sanitizer
deps targets have only private dependencies on libcxx_proxy so the
config does not apply to their dependents. This changes those
dependencies to be public.

This disables sanitizers for all non-default gcc_toolchain toolchains;
the android host toolchain build does not support building with libc++
and the GYP build does not use sanitizers for building host targets.

BUG=573288

Review-Url: https://codereview.chromium.org/1979973002
Cr-Commit-Position: refs/heads/master@{#394910}