Record histograms for DRP secure proxy check failures caused by CPs.
When the user is on a network with a captive portal, oftentimes the Data
Reduction Proxy's secure proxy check will fail when it receives a
redirect to the captive portal's login page. This CL causes these
failures to be recorded in the existing
DataReductionProxy.ProbeURLNetError histogram, together with the other
possible causes of failure.

BUG=613023

Review-Url: https://codereview.chromium.org/1996513002
Cr-Commit-Position: refs/heads/master@{#394984}