[Media Router WebUI] Update element positioning after delayed imports.
This change adds a call to update the positioning of dialog elements
after the delayed html imports have finished. Delaying the imports seems
to cause the first run flow element (which is conditionally templated)
to have an offsetHeight of 0 after it is stamped but before the imports
complete. Adding another update call lets the update correctly see this
element's offsetHeight when it is supposed to be visible.

R=apacible@chromium.org
BUG=612726
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/1996233002
Cr-Commit-Position: refs/heads/master@{#395116}