Functionality to allow blacklist and whitelist of custom schemes.
These changes will allow admins to define custom schemes as part of URLBlacklist and URLWhitelist policies.
Only the formats scheme://* and scheme:* are accepted.
The URLs of other formats containing custom schemes will be considered invalid and ignored.
In case a custom scheme is defined in whitelist, all the URLs for that scheme will proceed to be launched without asking the user.
In case a custom scheme is defined in blacklist, all the URLs with that scheme will have the same behavior as standard blacklisted URLs.

Most important changes done:

url_blacklist_manager.h - New enum URLBlacklistState added, to identify if the URL is in blacklist, whitelist or none. It is used to separate behavior of custom scheme URLs, so that:
 - if URL is whitelisted it will be launched without any warning/notification.
 - if URL is blacklisted it will redirect user to blocked page information.
 - if URL is not matched with any schemes from above, the old behavior is preserved meaning the user is asked if he wants to launch the URL.
resource_dispatcher_host_impl.cc - Pass the ResourceContext as additional parameter to ResourceDispatcherHostDelegate when handling external protocol.
resource_dispatcher_host_delegate.cc - uses PorfileIOData based on ResourceContext to detect if URL is blacklisted or whitelisted.
url_blacklist_manager.cc - Implement the new method to retrieve the URL state, and add code to allow the custom schemes in filter definitions.

BUG=158436

Review-Url: https://codereview.chromium.org/1692503002
Cr-Commit-Position: refs/heads/master@{#395305}