Remove user32 delay imports for XP, more GN imports match GYP.
This delay loads additional libraries so the set of delay-loaded and non-delay-loaded libraries is the same between GYP and GN for 32-bit Windows official builds of chrome.exe and chrome.dll. I checked the imported libraries but not the individual functions.

Remove user32 delay imports. This was a hack for Windows XP. I checked all the functions in the list of imports and they are all available on Windows 7. There was a lot of machinery required for this that was removed.

BUG=512861
Reland of https://crrev.com/080e7e885be49295f8373eb0fa54e6772206bc64 with fix.

Review-Url: https://codereview.chromium.org/2008643002
Cr-Commit-Position: refs/heads/master@{#395432}