Mac fullscreen low power: Add layout interactions and flag
This instantiates the NSWindow used for fullscreen low power mode, by
creating a FullscreenLowPowerCoordinator when entering fullscreen mode,
and adding it to the fullscreen transition group.

This also adds logic for determining if the layout of the browser chrome
is on-screen, and prevents moving the fullscreen low power window
in front if it is present.

Add a flag to enable this feature during development.

BUG=594460

Review-Url: https://codereview.chromium.org/1990163003
Cr-Commit-Position: refs/heads/master@{#395457}