[Media Router] Simplify representation of off the record routes in UI.
- Removes offTheRecord field from Route objects in UI. They are unused.
- Removes offTheRecord field from MediaRouterContainer. Instead of
sending the full controller URL for each route to the UI and not show
the controller when we see the container is OTR, we scrub out the the
controller URL before sending the route over to the UI. This also fixes
the issue with broken controller UI in incognito (which was previously
fixed in M51 but regressed in M52).
- We don't need to scrub out the custom controller path in the route
response. It's better to scrub it out in MediaRouterWebUIMessageHandler
so that we can cover the updated route list case as well.

This needs to be in M52.

BUG=590376
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2007653002
Cr-Commit-Position: refs/heads/master@{#395755}